### 2024-04-16 #6627 @editorialbot
[PRE REVIEW]: NGMT: NeuroGeriatric Motion Toolbox - An Open-Source Python Toolbox for Analyzing Neurological Motion Data from Various Recording Modalities
**Submitting author:** @JuliusWelzel (<a href="http://orcid.org/0000-0001-8958-0934">Julius Welzel</a>)
**Repository:** https://github.com/neurogeriatricskiel/NGMT
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.4
