### 2024-04-16 #6627 @Kevin-Mattheus-Moerman
On the license this is an OSI approved license: https://opensource.org/license/nposl-3-0
