### 2024-04-16 1f31c1d
Added REQUIRED_COLUMNS to ngmt_dataclass.py
