### 2024-04-16 51994d8
Imported REQUIRED_COLUMNS from ngmt_dataclass.py in datasets/keepcontrol.py
