### 2024-04-16 14020ce
Updated REQUIRED_DATACLASS in validate_channels method of NGMTRecording
