+++ docs/contributing.md
@@ -1,12 +1,12 @@
 ## Contributing guide
 Thanks for considering contributing to our toolbox! NGMT is an open-source project and we welcome contributions from anyone to further enhance this project
 
-There are lots of ways to contribute, such as:
-- Use the software, and when you find bugs, tell us about them! We can only fix the bugs we know about.
-- Tell us about parts of the documentation that you find confusing or unclear.
-- Tell us about things you wish NGMT could do, or things it can do but you wish they were easier.
-- Fix bugs.
-- Implement new features.
+There are lots of ways to contribute, such as:  
+- Use the software, and when you find bugs, tell us about them! We can only fix the bugs we know about.  
+- Tell us about parts of the documentation that you find confusing or unclear.  
+- Tell us about things you wish NGMT could do, or things it can do but you wish they were easier.  
+- Fix bugs.  
+- Implement new features.  
 - Improve existing tutorials or write new ones.
 
 To report bugs, request new features, or ask about confusing documentation, it’s usually best to open a [new issue](https://github.com/neurogeriatricskiel/NGMT/issues/new/choose) on GitHub. For better reproducibility, be sure to include information about your operating system and NGMT version, and (if applicable) include a reproducible code sample that is as short as possible and ideally uses one of [our example datasets](https://neurogeriatricskiel.github.io/NGMT/datasets/).
