### 2024-04-16 #6627 @JuliusWelzel
@Kevin-Mattheus-Moerman, thanks for taking over the editorial process of our manuscript.

For potential reviewer we would suggest
1. [sappelhoff](https://github.com/sappelhoff) - for his expertise in [BIDS](https://bids-specification.readthedocs.io/en/stable/)
2. [davidpagnon](https://github.com/davidpagnon) - for his expertise in biomechanics
3. [AKuederle](https://github.com/AKuederle) - for his experience with python package handling
