### 2024-04-17 4cfcde8
file_name handling to ensure consistency, using pathlib.Path.
