+++ docs/datasets/keepcontrol.md
@@ -1,6 +1,6 @@
-## KeepControl
+## Keep Control
 
-The keepcontrol dataset derived from the [keepcontrol project](https://www.keep-control.eu/) and is a Industrial Academic Initial Training Network working towards specific diagnosis and treatment of age-related gait and balance deficits.
+The Keep Control dataset derived from the [Keep Control](https://www.keep-control.eu/) project and is a Industrial Academic Initial Training Network working towards specific diagnosis and treatment of age-related gait and balance deficits. Part of the dataset was made publicly available on [figshare](https://figshare.com/articles/dataset/Full-body_mobility_data_to_validate_inertial_measurement_unit_algorithms_in_healthy_and_neurological_cohorts/20238006), and it was published as Warmerdam *et al*., Data, 2022, Full-Body Mobility Data to Validate Inertial Measurement Unit Algorithms in Healthy and Neurological Cohorts, doi: [10.3390/data7100136](https://doi.org/10.3390/data7100136).
 
 For this dataset a simple load function is provided to load the data into the NGMT dataclasses.
 
