+++ docs/datasets/mobilised.md
@@ -1,6 +1,6 @@
 ## Mobilise-D
 
-The mobilse-D dataset derived from the [mobilise-D consortuium](https://www.imi.europa.eu/projects-results/project-factsheets/mobilise-d) and is a European project that aims to improve the assessment of Parkinson’s disease using wearable sensors.
+The Mobilise-D dataset derived from the [Mobilise-D](https://mobilise-d.eu/) consortium and is a European project that aims to develop a comprehensive system to monitor and evaluate people's gait based on digital technologies, including sensors worn on the body, such as a low back-worn inertial measurement unit (IMU). Example data were made publicly available as Micó-Amigo *et al*., Zenodo, 2023, Assessing real-world gait with digital technology? Validation, insights and recommendations from the Mobilise-D consortium [[Data set]](https://zenodo.org/records/7547125), doi: [10.5281/zenodo.7547125](https://doi.org/10.5281/zenodo.7547125) and results for the entire dataset were published as Micó-Amigo *et al*., Journal of NeuroEngineering and Rehabilitation, 2023, Assessing real-world gait with digital technology? Validation, insights and recommendations from the Mobilise-D consortium, doi: [10.1186/s12984-023-01198-5](https://doi.org/10.1186/s12984-023-01198-5).
 
 For this dataset a simple load function is provided to load the data into the NGMT dataclasses.
 
