+++ docs/datasets/fairpark.md
@@ -1,7 +1,0 @@
-## Fair-Park II
-
-The Fair-Park II dataset derived from the [FairPark II project](https://www.fairpark2.eu/) and is a European project that primarily investigates the effects of an iron chelation therapy on the progression of handicap in Parkinson's disease.
-
-For this dataset a simple load function is provided to load the data into the NGMT dataclasses.
-
-::: datasets.fairpark
