+++ docs/datasets/index.md
@@ -6,6 +6,3 @@ the avaliable **datasets**.
 ## [KeepControl](keepcontrol.md)
 
 ## [Mobilise-D](mobilised.md)
-
-## [Fair-Park II](fairpark.md)
-
