+++ ngmt/test/test_modules.py
@@ -25,6 +25,7 @@ import pytest
 import numpy as np
 import pandas as pd
 from unittest.mock import patch
+from ngmt.datasets import keepcontrol
 from ngmt.modules.gsd import ParaschivIonescuGaitSequenceDetection
 from ngmt.modules.icd import ParaschivIonescuInitialContactDetection
 from ngmt.modules.pam import PhysicalActivityMonitoring
@@ -571,7 +572,7 @@ def detector():
 
 @pytest.fixture
 def gyro_data_numpy():
-    # Create a sample accelerometer and gyroscope data
+    # Sample accelerometer and gyroscope data
     accel_data = pd.DataFrame(np.random.rand(1000, 3), columns=['Accel_X', 'Accel_Y', 'Accel_Z'])
     gyro_data = pd.DataFrame(np.random.rand(1000, 3), columns=['Gyro_X', 'Gyro_Y', 'Gyro_Z'])
     sample_data = pd.concat([accel_data, gyro_data], axis=1)
@@ -582,7 +583,7 @@ def gyro_data_numpy():
     return gyro
 
 def test_bias_calculation_valid(detector, gyro_data_numpy):
-    # Create a sample accelerometer and gyroscope data
+    # Sample accelerometer and gyroscope data
     accel_data = pd.DataFrame(np.random.rand(1000, 3), columns=['Accel_X', 'Accel_Y', 'Accel_Z'])
     gyro_data = pd.DataFrame(np.random.rand(1000, 3), columns=['Gyro_X', 'Gyro_Y', 'Gyro_Z'])
     sample_data = pd.concat([accel_data, gyro_data], axis=1)
@@ -596,7 +597,7 @@ def test_bias_calculation_valid(detector, gyro_data_numpy):
     # Retrieve the calculated bias from the detector
     gyro_bias_detected = detector.gyro_bias
 
-# Soem additional test functions for Turn detection algorithm
+# Some additional test functions for Turn detection algorithm
 class TestPhamTurnDetection:
     @staticmethod
     def create_mock_data(num_samples=1000, sampling_freq_Hz=200):
@@ -658,7 +659,7 @@ class TestPhamTurnDetection:
 
     @pytest.fixture
     def sample_data(self):
-        # Sample data for testing
+        # Sample data for test
         diff_yaw = np.array([10, 20, 30, 40, 50, 60, 70, 80, 90, 100])
         flags_start_90 = [2, 5]
         flags_end_90 = [7, 10]
@@ -733,7 +734,7 @@ class TestPhamTurnDetection:
         assert duration_90 == [0.3, 0.3, 0.3]
 
 def test_invalid_plot_results_pham_td():
-    # Create a sample accelerometer and gyroscope data
+    # Sample accelerometer and gyroscope data
     accel_data = pd.DataFrame(np.random.rand(1000, 3), columns=['Accel_X', 'Accel_Y', 'Accel_Z'])
     gyro_data = pd.DataFrame(np.random.rand(1000, 3), columns=['Gyro_X', 'Gyro_Y', 'Gyro_Z'])
     sample_data = pd.concat([accel_data, gyro_data], axis=1)
@@ -748,7 +749,7 @@ def test_invalid_plot_results_pham_td():
 
 # Test function for PhamTurnDetection class
 @pytest.fixture
-def detector():
+def pham_detection_instance():
     return PhamTurnDetection()
 
 @pytest.fixture
@@ -757,30 +758,66 @@ def invalid_data():
     data = pd.DataFrame(np.random.rand(100, 5), columns=['Accel_X', 'Accel_Y', 'Accel_Z', 'Gyro_X', 'Gyro_Y'])
     return data
 
-def test_data_shape_invalid(detector, invalid_data):
[…]
