+++ ngmt/test/test_calc.py
@@ -29,6 +29,10 @@ import numpy.testing as npt
 import pytest
 import unittest
 import scipy
+from unittest.mock import patch
+from pathlib import Path
+from ngmt.datasets import keepcontrol
+from ngmt.utils.ngmt_dataclass import NGMTRecording
 from ngmt.utils.preprocessing import (
     resample_interpolate,
     lowpass_filter,
@@ -2117,6 +2121,172 @@ def test_pham_turn_plot_results_no_plot(monkeypatch):
     # Call the function
     pham_turn_plot_results(accel, gyro, detected_turns, sampling_freq_Hz)
 
+# Test function NGMT dataclass: case 1
+# Sample dataset for test
+@pytest.fixture
+def sample_data():
+    # Sample data for testing
+    data = pd.DataFrame({
+        'Time': [0.1, 0.2, 0.3],
+        'Accel_x': [1.0, 2.0, 3.0],
+        'Accel_y': [0.5, 1.5, 2.5],
+        'Accel_z': [0.2, 0.3, 0.4]
+    })
+    return data
+
+@pytest.fixture
+def sample_channels():
+    # Sample channels data for testing
+    channels = pd.DataFrame({
+        'name': ['Accel_x', 'Accel_y', 'Accel_z'],
+        'component': ['x', 'y', 'z'],
+        'ch_type': ['ACCEL', 'ACCEL', 'ACCEL']
+    })
+    return channels
+
+@pytest.fixture
+def sample_events():
+    # Sample events data for testing
+    events = pd.DataFrame({
+        'onset': [0.1, 0.2],
+        'duration': [0.1, 0.2],
+        'event_type': ['stimulus', 'response'],
+        'name': ['Stimulus A', 'Response B']
+    })
+    return events
+
+def test_add_events(sample_data, sample_channels, sample_events):
+    # Test add_events method
+    recording = NGMTRecording(data={'tracking_system_1': sample_data},
+                              channels={'tracking_system_1': sample_channels})
+    recording.add_events(tracking_system='tracking_system_1', new_events=sample_events)
+    assert 'tracking_system_1' in recording.events
+    assert len(recording.events['tracking_system_1']) == len(sample_events)
+
+def test_add_info():
+    # Test add_info method
+    recording = NGMTRecording(data={}, channels={})
+    recording.add_info(key='Subject', value='001')
+    assert recording.info['Subject'] == '001'
+
+def test_invalid_info_key():
+    # Test add_info with invalid info key
+    recording = NGMTRecording(data={}, channels={})
+    with pytest.raises(ValueError):
+        recording.add_info(key='InvalidKey', value='001')
+
+def test_invalid_file_path(tmp_path):
+    # Test export_events with invalid file path
+    recording = NGMTRecording(data={}, channels={}, events={}, info={})
+    file_path = tmp_path / 'invalid_path'  
+    with pytest.raises(ValueError):
+        recording.export_events(file_path=str(file_path))
+
+# Test function NGMT dataclass: case 2
+# Sample dataset for test
+@pytest.fixture
+def sample_data():
+    return pd.DataFrame({
+        'Time': [0.1, 0.2, 0.3],
+        'Accel_x': [1.0, 2.0, 3.0],
+        'Accel_y': [0.5, 1.5, 2.5],
+        'Accel_z': [0.2, 0.3, 0.4]
+    })
+
+@pytest.fixture
+def sample_channels():
+    return pd.DataFrame({
+        'name': ['Accel_x', 'Accel_y', 'Accel_z'],
+        'component': ['x', 'y', 'z'],
+        'ch_type': ['ACCEL', 'ACCEL', 'ACCEL']
+    })
+
[…]
