### 2024-04-30 #6627 @Kevin-Mattheus-Moerman
@JuliusWelzel apologies for the delay so far. I am currently looking for reviewers for this work. I'll check your suggestions also. One point that worries me at the moment is the warning in your README: `The toolbox is currently under development and is not yet ready for use.`. Please can you clarify this? I am afraid this would make the work unacceptable for JOSS.
