+++ ngmt/utils/importers.py
@@ -1,8 +1,9 @@
 import actipy
+import h5py
 import numpy as np
-import pandas as pd
 from ngmt.utils.ngmt_dataclass import NGMTRecording
 from ngmt.utils.file_io import get_unit_from_type
+import pandas as pd
 from pathlib import Path
 
 
@@ -66,3 +67,80 @@ def import_axivity(file_path: str, tracked_point: str):
     )
 
     return recording
+
+
+def import_mobilityLab(fullFileName: str) -> NGMTRecording:
+    """
+    Imports data from a mobility lab system from the specified file path and constructs an NGMTRecording object.
+
+    Args:
+        file_path (str or Path): The path to the mobility lab data file.
+        tracked_point (str): The name of the tracked point.
+
+    Returns:
+        NGMTRecording: The NGMTRecording object containing the imported data.
+
+    Examples:
+        >>> file_path = "/path/to/sensor_data.h5"
+        >>> tracked_point = "lowerBack"
+        >>> recording = import_mobilityLab(file_path, tracked_point)
+    """
+
+    # Convert file_path to a Path object if it is a string
+    if isinstance(file_path, str):
+        file_path = Path(file_path)
+
+    with h5py.File(fullFileName, 'r') as hfile:
+        monitor_labels = hfile.attrs['MonitorLabelList']
+        case_ids = hfile.attrs['CaseIdList']
+        channels_dict = {
+            "name": [],
+            "component": [],
+            "type": [],
+            "tracked_point": [],
+            "units": [],
+            "sampling_frequency": []
+        }
+        data_dict = {}
+        for ixSensor in range(len(monitor_labels)):
+            monitor_label = monitor_labels[ixSensor].decode('utf-8').strip()
+            case_id = case_ids[ixSensor].decode('utf-8')[:9]
+            sample_rate = hfile[case_id].attrs['SampleRate']
+            # Raw data
+            rawAcc = hfile[case_id]['Calibrated']['Accelerometers'][:]
+            rawAcc = hfile[case_id]['Calibrated']['Gyroscopes'][:]
+            rawMagn = hfile[case_id]['Calibrated']['Magnetometers'][:]
+
+            data_dict[f'{monitor_label}'] = pd.DataFrame({
+                f'{monitor_label}_ACCEL_x': rawAcc[:,0],
+                f'{monitor_label}_ACCEL_y': rawAcc[:,1],
+                f'{monitor_label}_ACCEL_z': rawAcc[:,2],
+                f'{monitor_label}_GYRO_x': rawAcc[:,0],
+                f'{monitor_label}_GYRO_y': rawAcc[:,1],
+                f'{monitor_label}_GYRO_z': rawAcc[:,2],
+                f'{monitor_label}_MAGN_x': rawMagn[:,0],
+                f'{monitor_label}_MAGN_y': rawMagn[:,1],
+                f'{monitor_label}_MAGN_z': rawMagn[:,2],
+            })
+            channels_dict["name"].extend([
+                f"{monitor_label}_ACCEL_x",
+                f"{monitor_label}_ACCEL_y",
+                f"{monitor_label}_ACCEL_z",
+                f"{monitor_label}_GYRO_x",
+                f"{monitor_label}_GYRO_y",
+                f"{monitor_label}_GYRO_z",
+                f"{monitor_label}_MAGN_x",
+                f"{monitor_label}_MAGN_y",
+                f"{monitor_label}_MAGN_z",
+            ])
+
+            channels_dict["component"].extend(['x', 'y', 'z'] * 3)
[…]
