### 2024-05-13 6c5c5e3
algorithms and examples were upadated.
