+++ docs/examples/05_tutorial_sit_to_stand_stand_to_sit_detection.md
@@ -2,7 +2,7 @@
 
 **Author:** Masoud Abedinifar
 
-**Last update:** Fri 03 May 2024
+**Last update:** Mon 13 May 2024
 
 ## Learning objectives  
 By the end of this tutorial:
@@ -50,7 +50,7 @@ To implement Pham Sit to Stand and Stand to Sit Detection algorithm, we load exa
 ```python
 # The 'file_path' variable holds the absolute path to the data file
 file_path = (
-    r"C:\Users\Project\Desktop\bigprojects\neurogeriatrics_data\Keep Control\Data\lab dataset\raw data\sub-pp002\motion\sub-pp002_task-tug_tracksys-imu_motion.tsv"
+    r"Keep Control\Data\lab dataset\raw data\sub-pp002\motion\sub-pp002_task-tug_tracksys-imu_motion.tsv"
 )
 
 # In this example, we use "imu" as tracking_system and "pelvis" as tracked points.
@@ -67,10 +67,20 @@ acceleration_data = recording.data[tracking_sys][
     ["pelvis_ACC_x", "pelvis_ACC_y", "pelvis_ACC_z"]
 ]
 
+# Get the acceleration data unit from the recording
+acceleration_data_unit = recording.channels[tracking_sys][
+    recording.channels[tracking_sys]["name"].str.contains("ACC", case=False)
+]["units"].iloc[0]
+
 # Load lower back gyro data
 gyro_data = recording.data[tracking_sys][
     ["pelvis_ANGVEL_x", "pelvis_ANGVEL_y", "pelvis_ANGVEL_z"]
 ]
+
+# Get the gyro data unit from the recording
+gyro_data_unit = recording.channels[tracking_sys][
+    recording.channels[tracking_sys]["name"].str.contains("ANGVEL", case=False)
+]["units"].iloc[0]
 
 # Print acceleration and gyro data
 print(f"acceleration_data (g): {acceleration_data}")
@@ -169,13 +179,15 @@ plt.show()
     
 
 ## Applying Pham sit to stand and stand to sit detection algorithm
-Now, we are running Pham sit to stand and stand to sit detection algorithm from pham module [`NGMT.ngmt.modules.ssd._pham.SittoStandStandtoSitDetection`](https://github.com/neurogeriatricskiel/NGMT/tree/main/ngmt/modules/ssd/_pham.py) to detect sit to stand and stand to sit movements.
+Now, we are running Pham sit to stand and stand to sit detection algorithm from pham module [`NGMT.ngmt.modules.ssd._pham.SittoStandStandtoSitDetection`](https://github.com/neurogeriatricskiel/NGMT/tree/main/ngmt/modules/ssd/_pham.py) to detect sit to stand and stand to sit.
 
-The following python code first prepares the input data by combining acceleration and gyro data into a single DataFrame called `input_data`.
+The following code first prepares the input data by combining acceleration and gyro data into a single DataFrame called `input_data`.
 
-Then, in order to apply sit to stand and stand to sit algorithm, an instance of the PhamSittoStandStandtoSitDetection class is created using the constructor, `PhamSittoStandStandtoSitDetection()`. The `pham` variable holds the instance, allowing us to access its methods. The inputs of the algorithm are as follows:
+Then, in order to apply sit to stand and stand to sit algorithm, an instance of the PhamSittoStandStandtoSitDetection class is created using the constructor, `PhamSittoStandStandtoSitDetection()`. The `pham` variable holds this instance, allowing us to access its methods. The inputs of the algorithm are as follows:
 
-- **Input Data:** `data` includes accelerometer and gyro data (N, 6) for x, y, and z axes in pandas Dataframe format.
[…]

[…]
