### 2024-05-13 0ad8a30
Improve error handling in import_axivity function and remove hardcoded tracked point
