+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "ngmt"
 version = "0.0.4"
-description = "The NeuroGeriatricMotionToolbox (NGMT) is a Python-based toolbox for processing motion data."
+description = "The NeuroGeriatricsMotionToolbox (NGMT) is a Python-based toolbox for processing motion data."
 license = "MIT"
 readme = "README.md"
 homepage = "https://neurogeriatricskiel.github.io/NGMT/"
