+++ ngmt/test/test_modules.py
@@ -549,7 +549,7 @@ def sample_data():
 def test_detection_output_shape(sample_data):
     # Test if the output DataFrame shape is correct
     detection = PhamSittoStandStandtoSitDetection()
-    result = detection.detect(sample_data, sampling_freq_Hz=100)
+    result = detection.detect(sample_data, acceleration_data_unit='g', gyro_data_unit='deg/s', sampling_freq_Hz=100, plot_results=False)
 
 
 def test_plot_results_pham():
@@ -561,6 +561,8 @@ def test_plot_results_pham():
     with pytest.raises(ValueError):
         gsd.detect(
             data=sample_data,
+            acceleration_data_unit='g',
+            gyro_data_unit='deg/s',
             sampling_freq_Hz=200,
             plot_results=invalid_plot_results,
         )
@@ -589,7 +591,7 @@ def test_bias_calculation_valid(detector, gyro_data_numpy):
     sample_data = pd.concat([accel_data, gyro_data], axis=1)
 
     # Perform detection
-    detector.detect(sample_data, 100)
+    detector.detect(sample_data, acceleration_data_unit='g', gyro_data_unit='deg/s', sampling_freq_Hz=100, plot_results=False)
     
     # Manually calculate the gyro bias for comparison
     gyro_bias_manual = np.mean(gyro_data_numpy[:100], axis=0)
@@ -612,50 +614,50 @@ class TestPhamTurnDetection:
     def test_detect_returns_instance(self):
         pham_detector = PhamTurnDetection()
         data, sampling_freq_Hz = self.create_mock_data()
-        result = pham_detector.detect(data=data, sampling_freq_Hz=sampling_freq_Hz)
+        result = pham_detector.detect(data=data, acceleration_data_unit='g', gyro_data_unit='deg/s', sampling_freq_Hz=sampling_freq_Hz, plot_results=False)
 
     def test_detect_raises_value_error_non_dataframe(self):
         pham_detector = PhamTurnDetection()
         data, sampling_freq_Hz = np.array([1, 2, 3]), 100
         with pytest.raises(ValueError):
-            pham_detector.detect(data=data, sampling_freq_Hz=sampling_freq_Hz)
+            pham_detector.detect(data=data,acceleration_data_unit='g', gyro_data_unit='deg/s', sampling_freq_Hz=sampling_freq_Hz, plot_results=False)
 
     def test_detect_raises_value_error_wrong_shape(self):
         pham_detector = PhamTurnDetection()
         data, sampling_freq_Hz = self.create_mock_data()
         data.drop(columns=['gyro_z'], inplace=True)
         with pytest.raises(ValueError):
-            pham_detector.detect(data=data, sampling_freq_Hz=sampling_freq_Hz)
+            pham_detector.detect(data=data, acceleration_data_unit='g', gyro_data_unit='deg/s', sampling_freq_Hz=sampling_freq_Hz, plot_results=False)
 
     def test_detect_raises_value_error_negative_sampling_freq(self):
         pham_detector = PhamTurnDetection()
         data, sampling_freq_Hz = self.create_mock_data()
         with pytest.raises(ValueError):
-            pham_detector.detect(data=data, sampling_freq_Hz=-100)
+            pham_detector.detect(data=data, acceleration_data_unit='g', gyro_data_unit='deg/s', sampling_freq_Hz=-100, plot_results=False)
 
     def test_detect_raises_value_error_non_boolean_plot_results(self):
         pham_detector = PhamTurnDetection()
         data, sampling_freq_Hz = self.create_mock_data()
         with pytest.raises(ValueError):
[…]
