### 2024-05-16 af87321
error handling in gsd algorithm.
