### 2024-05-24 fdbdfcf
final contact detection was added.
