+++ ngmt/modules/icd/_paraschiv.py
@@ -1,5 +1,5 @@
 # Import libraries
-from typing import Optional
+from typing import Optional, Literal, Union, TypeVar
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
@@ -8,6 +8,7 @@ import scipy.signal
 from ngmt.utils import preprocessing
 from ngmt.config import cfg_colors
 
+Self = TypeVar("Self", bound="ParaschivIonescuInitialContactDetection")
 
 class ParaschivIonescuInitialContactDetection:
     """
@@ -37,6 +38,14 @@ class ParaschivIonescuInitialContactDetection:
             0   5       initial contact  0          SU
             1   5.6     initial contact  0          SU
 
+        Calculate temporal parameters
+
+        >>> icd.calculate_temporal_parameters()
+        >>> print(icd.parameters_)
+                stride time [s]  swing time [s]  stance time [s]
+            0   1.2              0.4             0.8
+            1   1.1              0.3             0.8
+
     References:
         [1] Paraschiv-Ionescu et al. (2019). Locomotion and cadence detection using a single trunk-fixed accelerometer...
 
@@ -59,7 +68,7 @@ class ParaschivIonescuInitialContactDetection:
         gait_sequences: Optional[pd.DataFrame] = None,
         dt_data: Optional[pd.Series] = None,
         tracking_system: Optional[str] = None,
-    ) -> pd.DataFrame:
+    ) -> Self:
         """
         Detects initial contacts based on the input accelerometer data.
 
@@ -107,6 +116,9 @@ class ParaschivIonescuInitialContactDetection:
 
         # Initialize an empty list to store all onsets
         all_onsets = []
+        
+        # Initialize an empty list to store event types
+        event_types = []
 
         # Process each gait sequence
         if gait_sequences is None:
@@ -116,22 +128,36 @@ class ParaschivIonescuInitialContactDetection:
         for _, gait_seq in gait_sequences.iterrows():
             # Calculate start and stop indices for the current gait sequence
             start_index = int(sampling_freq_Hz * gait_seq["onset"])
-            stop_index = int(
-                sampling_freq_Hz * (gait_seq["onset"] + gait_seq["duration"])
-            )
-            accv_gait_seq = acc_vertical[start_index:stop_index].to_numpy()
+            stop_index = int(sampling_freq_Hz * (gait_seq["onset"] + gait_seq["duration"]))
+            accv_gait_seq = acc_vertical[start_index:stop_index + 2].to_numpy()
 
             try:
-                # Perform Signal Decomposition Algorithm for Initial Contacts (ICs)
-                initial_contacts_rel, _ = preprocessing.signal_decomposition_algorithm(
+                # Perform Signal Decomposition Algorithm for Initial Contacts (ICs) and Final Contacts (FCs)
+                initial_contacts_rel, final_contacts_rel = preprocessing.signal_decomposition_algorithm(
                     accv_gait_seq, sampling_freq_Hz
                 )
                 initial_contacts = gait_seq["onset"] + initial_contacts_rel
+                final_contacts = gait_seq["onset"] + final_contacts_rel
 
-                gait_seq["IC"] = initial_contacts.tolist()
+                # Combine and sort initial and final contacts
+                all_contacts = np.sort(np.concatenate((initial_contacts, final_contacts)))
+                all_event_types = ["initial contact"] * len(initial_contacts) + ["final contact"] * len(final_contacts)
+                all_event_types = [x for _, x in sorted(zip(np.concatenate((initial_contacts, final_contacts)), all_event_types))]
 
-                # Append onsets to the all_onsets list
-                all_onsets.extend(initial_contacts)
+                # Ensure first and last elements are initial contacts
+                if all_event_types[0] != "initial contact":
+                    all_contacts = all_contacts[1:]
[…]
