### 2024-05-24 8c3f6cc
Return an Error when "invalid" tracked points is given.
