### 2024-05-24 892ea14
[FIX] axivity importer returns data and channels dict
