### 2024-05-25 #6627 @Kevin-Mattheus-Moerman
@davidpagnon @sappelhoff @AKuederle @marcoghislieri @aaa34169

would you be interested in helping to review this submission for JOSS? The work is titled "_NGMT: NeuroGeriatric Motion Toolbox - An Open-Source Python Toolbox for Analyzing Neurological Motion Data from Various Recording Modalities_".

Software: https://github.com/neurogeriatricskiel/NGMT
Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06627/joss.06627/10.21105.joss.06627.pdf

JOSS reviews focus on evaluation of the software as well as the review of the very short paper. JOSS reviews happen in the open here on GitHub. See here for more information on our review process: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

If you are interested you can let me know here. Thanks!
