### 2024-05-25 #6627 @davidpagnon
Sure, why not!
