### 2024-05-26 #6627 @Kevin-Mattheus-Moerman
@JuliusWelzel the minimum is two. However I'll give it a bit more time to hopefully attract another. If I haven't found in the next couple of days, I'll start the review.
