### 2024-05-27 bde17a2
documentaion was updated for gait parameters.
