+++ docs/examples/modules_02_icd.md
@@ -2,15 +2,16 @@
 
 **Author:** Masoud Abedinifar
 
-**Last update:** Wed 27 Mar 2024
+**Last update:** Mon 27 May 2024
 
 ## Learning objectives
 By the end of this tutorial, you will be able to: 
 
-- You can load data from a recording that belongs to one of the available datasets,
+- Load data from a recording that belongs to one of the available datasets.
 - Apply the Paraschiv-Ionescu initial contact detection algorithm to accelerometer data.  
-- Visualize the results of initial contact detection.  
-- Interpret the detected initial contacts for further analysis.
+- Visualize the results of initial contact detection algorithm.
+- Extract temporal parameters for the detected strides.
+- Interpret the parameters for further analysis.
 
 # Paraschiv Initial Contact Detection
 
@@ -26,7 +27,7 @@ The algorithm takes accelerometer data as input, specifically the vertical accel
 [`2`] Paraschiv-Ionescu et al. (2020). Real-world speed estimation using a single trunk IMU: methodological challenges for impaired gait patterns. Annual International Conference of the IEEE Engineering in Medicine and Biology Society. IEEE Engineering in Medicine and Biology Society. https://doi.org/10.1109/EMBC44109.2020.9176281
 
 ## Import Libraries
-The necessary libraries such as numpy, matplotlib.pyplot, dataset (mobilised), Paraschiv-Ionescu gait sequence detection, and Paraschiv-Ionescu initial contact detection algorithms are imported from their corresponding modules. Make sure that you have all the required libraries and modules installed before running this code. You also may need to install the `ngmt` library and its dependencies if you haven't already.
+The necessary libraries such as numpy, matplotlib.pyplot, dataset (keepcontrol), Paraschiv-Ionescu gait sequence detection, and Paraschiv-Ionescu initial contact detection algorithms are imported from their corresponding modules. Make sure that you have all the required libraries and modules installed before running this code. You also may need to install the `ngmt` library and its dependencies if you haven't already.
 
 
 ```python
@@ -34,8 +35,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 import os
 from pathlib import Path
-
-from ngmt.datasets import mobilised
+from ngmt.datasets import keepcontrol
 from ngmt.modules.gsd import ParaschivIonescuGaitSequenceDetection
 from ngmt.modules.icd import ParaschivIonescuInitialContactDetection
 from ngmt.config import cfg_colors
@@ -43,143 +43,69 @@ from ngmt.config import cfg_colors
 
 ## Data Preparation
 
-To implement the Paraschiv-Ionescu initial contact algorithm, we load example data from a congestive heart failure (CHF) cohort, which is publicly available on the Zenodo repository [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7547125.svg)](https://doi.org/10.5281/zenodo.7547125). 
+To implement the Paraschiv-Ionescu initial contact detection algorithm, we load example data from keepcontrol dataset.
 
-The participant was assessed for 2.5 hours in the real-world while doing different daily life activities and also was asked to perform specific tasks such as outdoor walking, walking up and down a slope and stairs and moving from one room to another [`1`].
[…]
