+++ ngmt/utils/importers.py
@@ -1,12 +1,13 @@
 import actipy
 import h5py
 import numpy as np
+import pandas as pd
 from ngmt.utils.ngmt_dataclass import NGMTRecording
 from ngmt.utils.file_io import get_unit_from_type
-import pandas as pd
 from pathlib import Path
+from typing import Union
 
-
+# Importher for Axivity
 def import_axivity(file_path: str, tracked_point: str):
     """
     Imports Axivity data from the specified file path and 
@@ -64,53 +65,50 @@ def import_axivity(file_path: str, tracked_point: str):
     return data, channels
 
 
-# Importher for APDM Mobility Lab system
-def import_mobilityLab(
-    file_name: str | Path,
-    tracked_points: str | list[str],
+# Importher for APDM Mobility Lab for different versions
+def import_apdm_mobilityLab(
+    file_name: str | Path, 
+    tracked_points: str | list[str]
 ) -> tuple[pd.DataFrame, pd.DataFrame]:
     """
-    Imports data from an APDM Mobility Lab system from the specified file path.
+    Imports data from an APDM Mobility Lab system's different versions from the specified file path.
 
     Args:
         file_name (str or Path): The absolute or relative path to the data file.
-        tracked_point (str or list of str]):
-            Defines for which tracked points data are to be returned.
+        tracked_points (str or list of str]): Defines for which tracked points data are to be returned.
 
     Returns:
-        dict, dict: The loaded data and channels as dictionaries.
+        tuple[pd.DataFrame, pd.DataFrame]: The loaded data and channels as dataframes.
 
     Examples:
         >>> file_path = "/path/to/sensor_data.h5"
         >>> tracked_point = "Lumbar"
-        >>> recording = import_mobilityLab(file_path, tracked_point)
+        >>> data, channels = import_mobilityLab_all(file_path, tracked_point)
     """
     # Convert file_name to a Path object if it is a string
     if isinstance(file_name, str):
         file_name = Path(file_name)
 
-    # Convert tracked_points into a list if the it is provided as a string
+    # Convert tracked_points into a list if it is provided as a string
     if isinstance(tracked_points, str):
         tracked_points = [tracked_points]
 
-    with h5py.File(file_name, 'r') as hfile:
-        # Get monitor labels and case IDs
-        monitor_labels = hfile.attrs['MonitorLabelList']
-        monitor_labels = [
-            s.decode("UTF-8").strip()
-            for s in monitor_labels
-        ]
-        case_ids = hfile.attrs['CaseIdList']
-        case_ids = [
-            s.decode("UTF-8")[:9]
-            for s in case_ids
-        ]
-        
-        # Track invalid tracked points
-        invalid_tracked_points = [tp for tp in tracked_points if tp not in monitor_labels]
+    # Mapping of sensor types to make them consistent with NGMT dataclass definition
+    sensor_type_mapping = {
+        'Accelerometer': 'ACCEL',
+        'Gyroscope': 'GYRO',
+        'Magnetometer': 'MAGN',
+        'Accelerometers': 'ACCEL',
+        'Gyroscopes': 'GYRO',
+        'Magnetometers': 'MAGN'
+    }
 
-        if invalid_tracked_points:
-            raise ValueError(f"The following tracked points do not exist in monitor labels: {invalid_tracked_points}")
+    with h5py.File(file_name, 'r') as hfile:
+        # Check if there is an attribute or dataset that indicates the version
+        if 'FileFormatVersion' in hfile.attrs:
+            version = hfile.attrs['FileFormatVersion']
+        else:
+            raise ValueError("Version attribute not found in the h5 file.")
 
         # Initialize dictionaries to store channels and data frames
         channels_dict = {
@@ -121,55 +119,99 @@ def import_mobilityLab(
[…]
