### 2024-06-03 b854d98
Added an initial merge method for the recording object
