+++ ngmt/modules/td/_pham.py
@@ -3,67 +3,83 @@ import numpy as np
 import pandas as pd
 import scipy
 from ngmt.utils import preprocessing
-from ngmt.config import cfg_colors 
+from typing import Optional, TypeVar
+
+Self = TypeVar("Self", bound="PhamTurnDetection")
 
 
 class PhamTurnDetection:
     """
-    This algorithm aims to detect turns using accelerometer and gyroscope data collected from a lower back 
+    This algorithm aims to detect turns using accelerometer and gyroscope data collected from a lower back
     inertial measurement unit (IMU) sensor.
 
-    The core of the algorithm lies in the detect method, where turns are identified using accelerometer and 
-    gyroscope data. The method first processes the gyro data, converting it to rad/s and computing 
-    the variance to identify periods of low variance, which may indicate bias. It then calculates the gyro bias 
-    and subtracts it from the original gyro signal to remove any biases. Next, the yaw angle is computed by 
-    integrating the gyro data, and zero-crossings indices are found to detect turns. Then, turns are identified 
-    based on significant changes in the yaw angle. 
-    
-    The algorithm also accounts for hesitations, which are brief pauses or fluctuations in the signal that may 
-    occur within a turn. Hesitations are marked based on specific conditions related to the magnitude and 
+    The core of the algorithm lies in the detect method, where turns are identified using accelerometer and
+    gyroscope data. The method first processes the gyro data, converting it to rad/s and computing
+    the variance to identify periods of low variance, which may indicate bias. It then calculates the gyro bias
+    and subtracts it from the original gyro signal to remove any biases. Next, the yaw angle is computed by
+    integrating the gyro data, and zero-crossings indices are found to detect turns. Then, turns are identified
+    based on significant changes in the yaw angle.
+
+    The algorithm also accounts for hesitations, which are brief pauses or fluctuations in the signal that may
+    occur within a turn. Hesitations are marked based on specific conditions related to the magnitude and
     continuity of the yaw angle changes.
 
-    Then, the detected turns are characterized by their start and end times, duration, angle of turn, peak 
-    angular velocity, and direction (left or right). Turns with angles equal to or greater than 90 degrees 
-    and durations between 0.5 and 10 seconds are selected for further analysis. Finally, the detected turns 
-    along with their characteristics (onset, duration, direction, etc.) are stored in a pandas DataFrame 
+    Then, the detected turns are characterized by their onset and duration. Turns with angles equal to or greater
+    than 90 degrees and durations between 0.5 and 10 seconds are selected for further analysis. Finally, the detected
+    turns along with their characteristics (onset, duration, etc.) are stored in a pandas DataFrame
     (detected_turns attribute).
 
-    Optionally, if plot_results is set to True, the algorithm generates a plot visualizing the accelerometer 
-    and gyroscope data alongside the detected turns. This visualization aids in the qualitative assessment of 
+    In addition, spatial-temporal parameters are calculated using detected turns and their characteristics by
+    the spatio_temporal_parameters method. As a return, the turn id along with its spatial-temporal parameters including direction (left or right), angle of turn
+    and peak angular velocity are stored in a pandas DataFrame (detected_turns attribute).
+
+    Optionally, if plot_results is set to True, the algorithm generates a plot visualizing the accelerometer
[…]
