### 2024-06-05 #6842 @editorialbot
[REVIEW]: KielMAT: Kiel Motion Analysis Toolbox - An Open-Source Python Toolbox for Analyzing Neurological Motion Data from Various Recording Modalities
**Submitting author:** @JuliusWelzel (<a href="http://orcid.org/0000-0001-8958-0934">Julius Welzel</a>)
**Repository:** https://github.com/neurogeriatricskiel/NGMT
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.5
**Archive:** 10.17605/OSF.IO/8M2SP
