+++ ngmt/modules/td/_pham.py
@@ -3,6 +3,7 @@ import numpy as np
 import pandas as pd
 import scipy
 from ngmt.utils import preprocessing
+from ngmt.utils import viz_utils
 from typing import Optional, TypeVar
 
 Self = TypeVar("Self", bound="PhamTurnDetection")
@@ -14,11 +15,11 @@ class PhamTurnDetection:
     inertial measurement unit (IMU) sensor.
 
     The core of the algorithm lies in the detect method, where turns are identified using accelerometer and
-    gyroscope data. The method first processes the gyro data, converting it to rad/s and computing
+    gyroscope data. The method first processes the gyro data, converting it to rad/s if needed and computing
     the variance to identify periods of low variance, which may indicate bias. It then calculates the gyro bias
     and subtracts it from the original gyro signal to remove any biases. Next, the yaw angle is computed by
-    integrating the gyro data, and zero-crossings indices are found to detect turns. Then, turns are identified
-    based on significant changes in the yaw angle.
+    integrating the vertical component of the gyro data, and zero-crossings indices are found to detect turns.
+    Then, turns are identified based on significant changes in the yaw angle.
 
     The algorithm also accounts for hesitations, which are brief pauses or fluctuations in the signal that may
     occur within a turn. Hesitations are marked based on specific conditions related to the magnitude and
@@ -27,59 +28,67 @@ class PhamTurnDetection:
     Then, the detected turns are characterized by their onset and duration. Turns with angles equal to or greater
     than 90 degrees and durations between 0.5 and 10 seconds are selected for further analysis. Finally, the detected
     turns along with their characteristics (onset, duration, etc.) are stored in a pandas DataFrame
-    (detected_turns attribute).
+    (turns_ attribute).
 
     In addition, spatial-temporal parameters are calculated using detected turns and their characteristics by
-    the spatio_temporal_parameters method. As a return, the turn id along with its spatial-temporal parameters including direction (left or right), angle of turn
-    and peak angular velocity are stored in a pandas DataFrame (detected_turns attribute).
+    the spatio_temporal_parameters method. As a return, the turn id along with its spatial-temporal parameters
+    including direction (left or right), angle of turn and peak angular velocity are stored in a pandas DataFrame
+    (parameters_ attribute).
 
     Optionally, if plot_results is set to True, the algorithm generates a plot visualizing the accelerometer
     and gyroscope data alongside the detected turns. This visualization aids in the qualitative assessment of
     the algorithm's performance and provides insights into the dynamics of the detected turns.
 
     Methods:
-        detect(data, accel_unit, gyro_unit, sampling_freq_Hz, plot_results, tracking_system, tracked_point):
+        detect(data, gyro_vertical, accel_unit, gyro_unit, sampling_freq_Hz, dt_data, tracking_system, tracked_point, plot_results):
             Detects turns using accelerometer and gyro signals.
 
             Args:
                 data (pd.DataFrame): Input accelerometer and gyro data (N, 6) for x, y, and z axes.
+                gyro_vertical (str): The column name that corresponds to the vertical component gyro.
                 accel_unit (str): Unit of acceleration data.
                 gyro_unit (str): Unit of gyro data.
                 sampling_freq_Hz (float, int): Sampling frequency of the signals in Hz.
-                plot_results (bool, optional): If True, generates a plot. Default is False.
+                dt_data (pd.Series, optional): Original datetime in the input data. If original datetime is provided, the output onset will be based on that.
                 tracking_system (str, optional): Tracking systems.
[…]
