+++ ngmt/modules/ptd/_pham.py
@@ -1,0 +1,484 @@
+# Import libraries
+import numpy as np
+import pandas as pd
+import scipy.signal
+from ngmt.utils import preprocessing
+from ngmt.utils import viz_utils
+from typing import Optional, TypeVar
+
+Self = TypeVar("Self", bound="PhamPosturalTransitionDetection")
+
+class PhamPosturalTransitionDetection:
+    """
+    This algorithm aims to detect postural transitions (e.g., sit to stand or stand to sit movements)
+    using accelerometer and gyroscope data collected from a lower back inertial measurement unit (IMU)
+    sensor.
+
+    The algorithm is designed to be robust in detecting postural transitions using inertial sensor data 
+    and provides detailed information about these transitions. It starts by loading the accelerometer and 
+    gyro data, which includes three columns corresponding to the acceleration and gyro signals across 
+    the x, y, and z axes, along with the sampling frequency of the data. It first checks the validity of 
+    the input data. Then, it calculates the sampling period, selects accelerometer and gyro data. Tilt angle 
+    estimation is performed using gyro data in lateral or anteroposterior direction which represent movements 
+    or rotations in the mediolateral direction. The tilt angle is decomposed using wavelet transformation to 
+    identify stationary periods. Stationary periods are detected using accelerometer variance and gyro variance. 
+    Then, peaks in the wavelet-transformed tilt signal are detected as potential postural transition events.
+
+    If there's enough stationary data, further processing is done to estimate the orientation using
+    quaternions and to identify the beginning and end of postural transitions using gyro data. Otherwise,
+    if there's insufficient stationary data, direction changes in gyro data are used to infer postural
+    transitions. Finally, the detected postural transitions along with their characteristics (onset, duration, etc.) 
+    are stored in a pandas DataFrame (postural_transitions_ attribute).
+    
+    In addition, spatial-temporal parameters are calculated using detected postural transitions and their 
+    characteristics by the spatio_temporal_parameters method. As a return, the postural transition id along 
+    with its spatial-temporal parameters including type of postural transition (sit to stand or stand to sit), 
+    angle of postural transition, maximum flexion velocity, and maximum extension velocity are stored in a pandas 
+    DataFrame (parameters_ attribute).
+    
+    If requested (plot_results set to True), it generates plots of the accelerometer and gyroscope data
+    along with the detected postural transitions.
+
+    Methods:
+        detect(data, gyro_vertical, accel_unit, gyro_unit, sampling_freq_Hz, dt_data, tracking_system, tracked_point, plot_results):
+            Detects  sit to stand and stand to sit using accelerometer and gyro signals.
+
+            Args:
+                data (pd.DataFrame): Input accelerometer and gyro data (N, 6) for x, y, and z axes.
+                gyro_mediolateral (str): It corresponds to the gyroscope component representing movements or rotations in the mediolateral direction. This could correspond to rotation around the y-axis in your coordinate system.
+                accel_unit (str): Unit of acceleration data.
+                gyro_unit (str): Unit of gyro data.
+                sampling_freq_Hz (float, int): Sampling frequency of the signals.
+                dt_data (pd.Series, optional): Original datetime in the input data. If original datetime is provided, the output onset will be based on that.
+                tracking_system (str, optional): Tracking systems.
+                tracked_point (str, optional): Tracked points on the body.
+                plot_results (bool, optional): If True, generates a plot. Default is False.
+
+            Returns:
[…]
