+++ ngmt/modules/ssd/__init__.py
@@ -1,1 +1,0 @@
-from ._pham import PhamSittoStandStandtoSitDetection
