+++ ngmt/test/test_modules.py
@@ -1,15 +1,14 @@
 # Introduction and explanation regarding the test suite
 """
 This code is a test suite for various modules which exist in the NGMT toolbox. 
-It employs pytest, a Python testing framework, to verify the correctness of these functions. 
+It employs pytest, a Python testing framework, to verify the correctness of these modules. 
 Here's a brief explanation of the code structure:
 
-1. Import necessary libraries, pytest and the functions to be tested.
+1. Import necessary libraries, pytest and the modules to be tested.
 2. Generate a random input signal for testing purposes.
-3. Define a series of test functions, each targeting a specific function from the specific module.
-4. Inside each test module, we validate the correctness of the corresponding function and its inputs.
-5. We make use of 'assert' statements to check that the functions return expected results.
-6. The code is organized for clarity and maintainability.
+3. Inside each test module, we validate the correctness of the modules.
+4. We make use of 'assert' statements to check that the functions return expected results.
+5. The code is organized for clarity and maintainability.
 
 To run the tests, follow these steps:
 
@@ -24,8 +23,9 @@ By running these tests, the reliability and correctness of the modules will be e
 import pytest
 import numpy as np
 import pandas as pd
+import os
+from datetime import datetime
 from unittest.mock import patch
-from ngmt.datasets import keepcontrol
 from ngmt.modules.gsd import ParaschivIonescuGaitSequenceDetection
 from ngmt.modules.icd import ParaschivIonescuInitialContactDetection
 from ngmt.modules.pam import PhysicalActivityMonitoring
@@ -34,7 +34,7 @@ from ngmt.modules.td import PhamTurnDetection
 
 
 ## Module test
-# Test for gait sequence detection algorithm
+# Test funtions for gait sequence detection algorithm
 num_samples = 50000  # Number of samples
 acceleration_data = {
     "LowerBack_ACCEL_x": np.random.uniform(-2, 2, num_samples),
@@ -44,7 +44,6 @@ acceleration_data = {
 acceleration_data = pd.DataFrame(acceleration_data)
 sampling_frequency = 100  # Sampling frequency
 
-
 def test_gsd_detect():
     # Initialize the class
     gsd = ParaschivIonescuGaitSequenceDetection()
@@ -52,7 +51,6 @@ def test_gsd_detect():
     # Call the detect method
     gsd.detect(data=acceleration_data, sampling_freq_Hz=sampling_frequency)
     gait_sequences_ = gsd.gait_sequences_
-
 
 def test_invalid_sampling_freq():
     # Initialize the class
@@ -63,15 +61,6 @@ def test_invalid_sampling_freq():
     with pytest.raises(ValueError):
         gsd.detect(data=acceleration_data, sampling_freq_Hz=invalid_sampling_freq)
 
-
-def test_gait_sequence_detection():
-    # Initialize the class
-    gsd = ParaschivIonescuGaitSequenceDetection()
-
-    # Call the detect method
-    gsd.detect(data=acceleration_data, sampling_freq_Hz=sampling_frequency)
-
-
 def test_invalid_input_data_type():
     # Initialize the class
     gsd = ParaschivIonescuGaitSequenceDetection()
@@ -81,7 +70,6 @@ def test_invalid_input_data_type():
     with pytest.raises(ValueError):
         gsd.detect(data=invalid_data, sampling_freq_Hz=sampling_frequency)
 
-
 def test_invalid_sampling_freq_type():
     # Initialize the class
[…]
