### 2024-06-11 #6842 @davidpagnon
Hi @JuliusWelzel,
It seems like you did not allow for issues to be posted on NGMT, while it is needed for the reviewing process. Could you fix it? I'll post my potential remarks there once it is done.

> Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the [target repository](https://github.com/neurogeriatricskiel/NGMT) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)
