### 2024-06-12 #6842 @JuliusWelzel
Hi,

issues seem to be allowed:
[image: image.png]

Let me know if this is working for you.

Best,
Julius

Am Di., 11. Juni 2024 um 19:35 Uhr schrieb David PAGNON <
***@***.***>:

> Hi @JuliusWelzel <https://github.com/JuliusWelzel>,
> It seems like you did not allow for issues to be posted on NGMT, while it
> is needed for the reviewing process. Could you fix it? I'll post my
> potential remarks there once it is done.
>
> Please avoid lengthy details of difficulties in the review thread.
> Instead, please create a new issue in the target repository
> <https://github.com/neurogeriatricskiel/NGMT> and link to those issues
> (especially acceptance-blockers) by leaving comments in the review thread
> below. (For completists: if the target issue tracker is also on GitHub,
> linking the review thread in the issue or vice versa will create
> corresponding breadcrumb trails in the link target.)
>
