### 2024-06-12 #6842 @JuliusWelzel
![image](https://github.com/openjournals/joss-reviews/assets/52565341/d20b296c-188b-4363-999b-52c3d7fd3cfb)
