+++ docs/examples/modules_05_td.md
@@ -2,12 +2,12 @@
 
 **Author:** Masoud Abedinifar
 
-**Last update:** Thu 06 Jun 2024
+**Last update:** Wed 12 Jun 2024
 
 ## Learning Objectives
 By the end of this tutorial:
 
-- You can load data from a recording that belongs to one of the available datasets,
+- You can load data from keepcontrol which is one of available datasets.
 - Apply the Pham turn detection algorithm.
 - Visualize the results of the algorithm.
 - Extract spatio-temporal parameters of the detected turns.
