+++ docs/examples/modules_04_ptd.md
@@ -2,12 +2,12 @@
 
 **Author:** Masoud Abedinifar
 
-**Last update:** Thu 06 Jun 2024
+**Last update:** Wed 12 Jun 2024
 
 ## Learning objectives
 By the end of this tutorial:
 
-- You can load data from a recording that belongs to one of the available datasets,
+- You can load data from keepcontrol which is one of available datasets.
 - Apply the Postural Transition Detection algorithm.
 - Visualize the results of the algorithm.
 - Extract spatio-temporal parameters of the detected postural transitions.
