+++ ngmt/test/test_calc.py
@@ -1682,7 +1682,7 @@ thresholds_mg = {
 }
 
 
-# Test function for gsd_plot_results without plotting
+# Test function for gsd_plot_results without plotting uisng monkeypatch which allows testing function without plotting
 def test_gsd_plot_results_without_plot(monkeypatch):
     # A mock function for plt.show() that does nothing
     def mock_show():
