### 2024-06-12 #6842 @davidpagnon
Here is a first pass for the review:

## General checks
- **License:** Non-Profit Open Software License version 3.0" is OSI approved: https://opensource.org/license/nposl-3-0
- **Substantial scholarly effort:** The tools offered by NGMT are useful in the field of gait analysis, and are straightforward to put into application
- **Data sharing:** Two large datasets are natively supported (Keep Control and Mobilize-D), and tools are provided to load them.
  *Example files for tutorials are missing for now.*
  *It would be good to add a line stating that it has been gathered in accordance with the Declaration of Helsinki.*

## Functionality
- **Installation:** Installation ran properly on python 3.11 (on python 3.12, I had to add the --user flag: `pip install ngmt --user`)
- **Functionality:** *I could not test all tutorials yet, due to missing example files*.
  The theory behind each module is well explained on the website and in the docstrings. *However, I would appreciate it if you provided the gist of the theory behind each module in the paper.*

## Documentation
- **Statement of need:** The statement of need is clear and the target audience (motion researchers) is also specified.
  *It may be worth specifying that the expected motion data are 3D accelerations (from the wrist or the lower-back depending on the module).*
  *Note that the list is not rendered on the website (all items are on the same line).*
- **Example usage:** *Example data files are currently missing*
- **Code quality:** The code is clearly documented (both on the website and in the docstrings), it is very well written, classes and functions are easy to read. Tests are extensive.
- **Community guidelines:** Issues have been activated on the repository so the user can report issues or seek support; there are contribution guidelines on the website and in the paper, but it would be good to refer to them in the Readme.md file.

## Software paper
- **Summary:**  *It is not immediately clear from the paper what kind of data the toolbox uses:* marker-based data? Video input? RGB-D? IMUs? I suppose any piece of hardware could be suitable as long as it can provide acceleration data (from the pelvis? the wrist?), but it is not stated. I would be more specific very early on, for example by specifying which type of motion data is expected L12: "NGMT offers a range of 12 algorithms for the processing of motion data in neuroscience and biomechanics".
- **Statement of need:**   I would like more detail to be given about each of the 4 modules. I feel like someone skimming the paper could easily miss that there are actually 4 main modules, and if they do, they may not understand what they consist in. *Why are these 4 modules (+ the ones under development: "postural transition analysis" and "turns") specifically useful?*

  *Why are accelerometers better than vision-based approaches in your context?* Supposedly because unlike being filmed, one can wear accelerometers all day long and at any place.

  I do not understand the end of the sentence L33: *"DHT [...] offers a new dimension of measuring daily function, that is, performance"*

  One subsidiary remark: NGMT is just a sequence of letters that will not bear any meaning to anyone who comes across it; moreover, I think this toolbox could be useful to many more cases than the very specific field of neurogeriatric motion. *For the sake of dissemination, would it be worth finding another name?*
- **State of the field:** It may be interesting to cite other approaches not using acceleration-based data, such  as the Mobile gait-lab ([paper](https://www.nature.com/articles/s41467-020-17807-z) and [code](https://github.com/stanfordnmbl/mobile-gaitlab)): although it works from video input and focuses on different quantities (walking speed, cadence, maximum knee flexion, and gait deviation), the approach is globally similar in that it is open-source and meant to be easily accessible to non computer-science specialists. [Sit2Stand](https://sit2stand.ai/) and [Sports2D](https://github.com/davidpagnon/Sports2D/) are other repositories that could make it to this section.
- **Quality of writing:** I am not a native English speaker but no obvious error jumps to mind.
