+++ ngmt/modules/ptd/_pham.py
@@ -200,7 +200,7 @@ class PhamPosturalTransitionDetection:
         # Select gyro data and convert it to numpy array format
         gyro = data[gyro_columns].copy().to_numpy()
         self.gyro = gyro
-        
+
         # Extract mediolateral gyro data using the specified index
         self.gyro_mediolateral = data[gyro_mediolateral].to_numpy()
 
@@ -216,7 +216,9 @@ class PhamPosturalTransitionDetection:
         # Check unit of gyro data if it is in deg/s or rad/s
         if gyro_unit in ["rad/s", "radians per second"]:
             # Convert gyro data from rad/s to deg/s (if not already is in deg/s)
-            gyro = np.rad2deg(gyro)
+            self.gyro = np.rad2deg(self.gyro)
+            self.gyro_mediolateral = np.rad2deg(self.gyro_mediolateral)
+
         elif gyro_unit in ["deg/s", "degrees per second"]:
             pass  # Gyro data is already in rad/s
         else:
@@ -279,7 +281,7 @@ class PhamPosturalTransitionDetection:
         stationary_2 = (accel_var <= self.thr_gyro_var).astype(int)
 
         # Convert unit of the gyro data from deg/s to rad/s
-        gyro = np.deg2rad(self.gyro)
+        self.gyro = np.deg2rad(self.gyro)
 
         # Calculate stationary of gyro variance
         gyro_norm = np.sqrt(self.gyro[:, 0] ** 2 + self.gyro[:, 1] ** 2 + self.gyro[:, 2] ** 2)
@@ -346,7 +348,7 @@ class PhamPosturalTransitionDetection:
             self.left_side = np.ones_like(self.local_peaks)
 
             # Initialize right side indices with length of gyro data
-            self.lef_side = len(self.gyro_mediolateral) * np.ones_like(self.local_peaks)
+            self.right_side = len(self.gyro_mediolateral) * np.ones_like(self.local_peaks)
 
             # Loop through each local peak
             for i in range(len(self.local_peaks)):
@@ -380,25 +382,34 @@ class PhamPosturalTransitionDetection:
                                 break
                 for j in range(len(dist2peak_right_side)):
                     if gyro_y_diff[postural_transitions + dist2peak_right_side[j]] < 0 and dist2peak_right_side[j] > (0.2*sampling_freq_Hz):
-                        self.lef_side[i] = postural_transitions + dist2peak_right_side[j]
+                        self.right_side[i] = postural_transitions + dist2peak_right_side[j]
                         break
 
-            # Calculate postural transition angle
-            self.postural_transition_angle = np.abs(tilt_angle_deg[self.local_peaks] - tilt_angle_deg[self.left_side])
-            if self.left_side[0] == 1:
-                self.postural_transition_angle[0] = abs(
-                    tilt_angle_deg[self.local_peaks[0]] - tilt_angle_deg[self.lef_side[0]]
-                )
+                # Calculate postural transition angle
+                self.postural_transition_angle = np.abs(tilt_angle_deg[self.local_peaks] - tilt_angle_deg[self.left_side])
+                if self.left_side[0] == 1:
+                    self.postural_transition_angle[0] = abs(
+                        tilt_angle_deg[self.local_peaks[0]] - tilt_angle_deg[self.right_side[0]]
+                    )
 
-            # Calculate duration of each postural transition
-            duration = (self.lef_side - self.left_side) / sampling_freq_Hz
+                # Calculate duration of each postural transition
+                duration = (self.right_side - self.left_side) / sampling_freq_Hz
 
-            # Convert peak times to integers
-            postural_transition_onset = time[self.local_peaks]
+                # Convert peak times to integers
+                postural_transition_onset = time[self.local_peaks] 
 
         # Remove too small postural transitions
[…]
