### 2024-06-17 960ad53
data name revision based on dataclass standard.
