### 2024-06-17 ca2c147
data coloumn names were revised to match dataclass standards.
