+++ ngmt/test/test_modules.py
@@ -36,12 +36,12 @@ from ngmt.modules.td import PhamTurnDetection
 ## Module test
 # Test funtions for gait sequence detection algorithm
 num_samples = 50000  # Number of samples
-acceleration_data = {
+accel_data = {
     "LowerBack_ACCEL_x": np.random.uniform(-2, 2, num_samples),
     "LowerBack_ACCEL_y": np.random.uniform(-2, 2, num_samples),
     "LowerBack_ACCEL_z": np.random.uniform(-2, 2, num_samples),
 }
-acceleration_data = pd.DataFrame(acceleration_data)
+accel_data = pd.DataFrame(accel_data)
 sampling_frequency = 100  # Sampling frequency
 
 def test_gsd_detect():
@@ -49,7 +49,7 @@ def test_gsd_detect():
     gsd = ParaschivIonescuGaitSequenceDetection()
 
     # Call the detect method
-    gsd.detect(data=acceleration_data, sampling_freq_Hz=sampling_frequency)
+    gsd.detect(data=accel_data, sampling_freq_Hz=sampling_frequency)
     gait_sequences_ = gsd.gait_sequences_
 
 def test_invalid_sampling_freq():
@@ -59,7 +59,7 @@ def test_invalid_sampling_freq():
     # Test with invalid sampling frequency
     invalid_sampling_freq = "invalid"
     with pytest.raises(ValueError):
-        gsd.detect(data=acceleration_data, sampling_freq_Hz=invalid_sampling_freq)
+        gsd.detect(data=accel_data, sampling_freq_Hz=invalid_sampling_freq)
 
 def test_invalid_input_data_type():
     # Initialize the class
@@ -77,7 +77,7 @@ def test_invalid_sampling_freq_type():
     # Test with invalid sampling frequency type
     invalid_sampling_freq = "invalid"
     with pytest.raises(ValueError):
-        gsd.detect(data=acceleration_data, sampling_freq_Hz=invalid_sampling_freq)
+        gsd.detect(data=accel_data, sampling_freq_Hz=invalid_sampling_freq)
 
 def test_plot_results_type():
     # Initialize the class
@@ -87,7 +87,7 @@ def test_plot_results_type():
     invalid_plot_results = "invalid"
     with pytest.raises(ValueError):
         gsd.detect(
-            data=acceleration_data,
+            data=accel_data,
             sampling_freq_Hz=sampling_frequency,
             plot_results=invalid_plot_results,
         )
@@ -100,7 +100,7 @@ def test_invalid_dt_data_type():
     invalid_dt_data = "invalid"
     with pytest.raises(ValueError):
         gsd.detect(
-            data=acceleration_data,
+            data=accel_data,
             sampling_freq_Hz=sampling_frequency,
             plot_results=False,
             dt_data=invalid_dt_data,
@@ -114,7 +114,7 @@ def test_invalid_dt_data_format():
     invalid_dt_data = pd.Series(["2024-03-22 10:00:00", "2024-03-22 10:00:01"])
     with pytest.raises(ValueError):
         gsd.detect(
-            data=acceleration_data,
+            data=accel_data,
             sampling_freq_Hz=sampling_frequency,
             plot_results=False,
             dt_data=invalid_dt_data,
@@ -126,12 +126,12 @@ def test_valid_dt_data():
 
     # Create valid datetime data
     valid_dt_data = pd.Series(
-        pd.date_range(start="2024-03-22", periods=len(acceleration_data), freq="S")
+        pd.date_range(start="2024-03-22", periods=len(accel_data), freq="S")
     )
 
     # Call the detect method with valid datetime data
     gsd.detect(
-        data=acceleration_data,
[…]
