+++ ngmt/modules/ptd/_pham.py
@@ -8,34 +8,35 @@ from typing import Optional, TypeVar
 
 Self = TypeVar("Self", bound="PhamPosturalTransitionDetection")
 
+
 class PhamPosturalTransitionDetection:
     """
     This algorithm aims to detect postural transitions (e.g., sit to stand or stand to sit movements)
     using accelerometer and gyroscope data collected from a lower back inertial measurement unit (IMU)
     sensor.
 
-    The algorithm is designed to be robust in detecting postural transitions using inertial sensor data 
-    and provides detailed information about these transitions. It starts by loading the accelerometer and 
-    gyro data, which includes three columns corresponding to the acceleration and gyro signals across 
-    the x, y, and z axes, along with the sampling frequency of the data. It first checks the validity of 
-    the input data. Then, it calculates the sampling period, selects accelerometer and gyro data. Tilt angle 
-    estimation is performed using gyro data in lateral or anteroposterior direction which represent movements 
-    or rotations in the mediolateral direction. The tilt angle is decomposed using wavelet transformation to 
-    identify stationary periods. Stationary periods are detected using accelerometer variance and gyro variance. 
+    The algorithm is designed to be robust in detecting postural transitions using inertial sensor data
+    and provides detailed information about these transitions. It starts by loading the accelerometer and
+    gyro data, which includes three columns corresponding to the acceleration and gyro signals across
+    the x, y, and z axes, along with the sampling frequency of the data. It first checks the validity of
+    the input data. Then, it calculates the sampling period, selects accelerometer and gyro data. Tilt angle
+    estimation is performed using gyro data in lateral or anteroposterior direction which represent movements
+    or rotations in the mediolateral direction. The tilt angle is decomposed using wavelet transformation to
+    identify stationary periods. Stationary periods are detected using accelerometer variance and gyro variance.
     Then, peaks in the wavelet-transformed tilt signal are detected as potential postural transition events.
 
     If there's enough stationary data, further processing is done to estimate the orientation using
     quaternions and to identify the beginning and end of postural transitions using gyro data. Otherwise,
     if there's insufficient stationary data, direction changes in gyro data are used to infer postural
-    transitions. Finally, the detected postural transitions along with their characteristics (onset, duration, etc.) 
+    transitions. Finally, the detected postural transitions along with their characteristics (onset, duration, etc.)
     are stored in a pandas DataFrame (postural_transitions_ attribute).
-    
-    In addition, spatial-temporal parameters are calculated using detected postural transitions and their 
-    characteristics by the spatio_temporal_parameters method. As a return, the postural transition id along 
-    with its spatial-temporal parameters including type of postural transition (sit to stand or stand to sit), 
-    angle of postural transition, maximum flexion velocity, and maximum extension velocity are stored in a pandas 
+
+    In addition, spatial-temporal parameters are calculated using detected postural transitions and their
+    characteristics by the spatio_temporal_parameters method. As a return, the postural transition id along
+    with its spatial-temporal parameters including type of postural transition (sit to stand or stand to sit),
+    angle of postural transition, maximum flexion velocity, and maximum extension velocity are stored in a pandas
     DataFrame (parameters_ attribute).
-    
+
[…]
