### 2024-06-25 c94143d
Added DOIs for (conference) papers
