+++ ngmt/modules/td/_pham.py
@@ -210,11 +210,11 @@ class PhamTurnDetection:
         elif accel_unit in ["g", "G"]:
             pass  # No conversion needed
         else:
-            raise ValueError("Unsupported unit for acceleration data")
+            raise ValueError("Invalid unit for acceleration data. Must be 'm/s^2' or 'g'")
 
         # Convert variations of gyro unit to "deg/s"
-        if accel_unit in ["degrees per second", "°/s"]:
-            accel_unit = "deg/s"
+        if gyro_unit in ["degrees per second", "°/s"]:
+            gyro_unit = "deg/s"
 
         # Check unit of gyro data if it is in deg/s or rad/s
         if gyro_unit == "deg/s":
