### 2024-06-26 dd55c38
Added pooch and tqdm as dependencies
