+++ pyproject.toml
@@ -43,6 +43,8 @@ json-repair = "^0.8.0"
 pyarrow = "^15.0.0"
 bids-validator = "^1.14.4"
 actipy = "^3.0.5"
+pooch = "^1.8.2"
+tqdm = "^4.66.4"
 
 [tool.poetry.extras]
 docs = ["mkdocs", "mkdocs-material", "pymdownx.superfences"]
