+++ ngmt/datasets/mobilised.py
@@ -1,6 +1,9 @@
 import numpy as np
 import pandas as pd
-import pathlib
+from pathlib import Path
+from pooch import DOIDownloader
+from zipfile import ZipFile
+from typing import Literal, Any
 from ngmt.utils import matlab_loader
 from ngmt.utils.ngmt_dataclass import NGMTRecording
 
@@ -30,8 +33,162 @@ MAP_CHANNEL_UNITS = {
 }
 
 
+def fetch_dataset(
+    progressbar: bool = True,
+    dataset_path: str | Path = Path(__file__).parent / "_mobilised",
+) -> None:
+    dataset_path = Path(dataset_path) if isinstance(dataset_path, str) else dataset_path
+
+    # Check if zip archive has already been downloaded
+    if not dataset_path.exists():
+        dataset_path.parent.joinpath("_mobilised").mkdir(parents=True, exist_ok=True)
+    _output_file = dataset_path.joinpath("Mobilise-D_dataset.zip")
+
+    if not _output_file.exists():
+        # Set the URL to the dataset
+        _url = "doi:10.5281/zenodo.7547125/Mobilise-D dataset_1-18-2023.zip"
+
+        # Instantiate a downloader object
+        downloader = DOIDownloader(progressbar=progressbar)
+        downloader(url=_url, output_file=_output_file, pooch=None)
+
+    # Extract the dataset
+    with ZipFile(_output_file, "r") as zip_ref:
+        zip_ref.extractall(dataset_path)
+    return
+
+
+def load_file(
+    tracking_systems: None | str | list[str] = None,
+    tracked_points: None | dict[str, str] | dict[str, list[str]] = None,
+    cohort: Literal["PFF", "PD", "MS", "HA", "COPD", "CHF"] = "PFF",
+    file_name: str = "data.mat",
+    dataset_path: str | Path = Path(__file__).parent / "_mobilised",
+    progressbar: None | bool = None,
+) -> dict[str, Any]:
+    # Local reference to the available tracking systems and tracked points
+    _AVAILABLE_TRACKING_SYSTEMS = ["SU", "PressureInsoles_raw", "DistanceModule_raw"]
+    _AVAILABLE_TRACKED_POINTS = {
+        "SU": ["LowerBack"],
+        "PressureInsoles_raw": ["LeftFoot", "RightFoot"],
+        "DistanceModule_raw": ["LeftFoot", "RightFoot"],
+    }
+
+    # Check the tracking systems
+    tracking_systems = (
+        _AVAILABLE_TRACKING_SYSTEMS if tracking_systems is None else tracking_systems
+    )
+    tracking_systems = (
+        [tracking_systems] if isinstance(tracking_systems, str) else tracking_systems
+    )
+    if any(
+        [track_sys not in _AVAILABLE_TRACKING_SYSTEMS for track_sys in tracking_systems]
+    ):
+        raise ValueError(
+            f"Invalid tracking system. Available options are: {_AVAILABLE_TRACKING_SYSTEMS}"
+        )
+
+    # Set the tracked points
+    tracked_points = (
+        _AVAILABLE_TRACKED_POINTS if tracked_points is None else tracked_points
+    )
+    if any([track_sys not in tracking_systems for track_sys in tracked_points.keys()]):
+        raise ValueError(
+            f"You have specified tracked points for a tracking system that you have not requested (requested: {tracking_systems})."
+        )
+    for track_sys in tracked_points.keys():
+        if isinstance(tracked_points[track_sys], str):
+            tracked_points[track_sys] = [tracked_points[track_sys]] # type: ignore
+
+    # Loop over the tracked points and check if they are valid
+    for track_sys, tracked_points_list in tracked_points.items():
+        if any(
+            [
[…]
