+++ ngmt/datasets/mobilised.py
@@ -14,7 +14,7 @@ MAP_CHANNEL_TYPES = {
     "Gyr": "GYRO",
     "Mag": "MAGN",
     "Bar": "BARO",
-    # "Temp": "TEMP"
+    # "Temp": "TEMP",
 }
 
 MAP_CHANNEL_COMPONENTS = {
@@ -37,6 +37,12 @@ def fetch_dataset(
     progressbar: bool = True,
     dataset_path: str | Path = Path(__file__).parent / "_mobilised",
 ) -> None:
+    """Fetch the Mobilise-D dataset from the Zenodo repository.
+
+    Args:
+        progressbar (bool, optional): Whether to display a progressbar. Defaults to True.
+        dataset_path (str | Path, optional): The path where the dataset is stored. Defaults to Path(__file__).parent/"_mobilised".
+    """
     dataset_path = Path(dataset_path) if isinstance(dataset_path, str) else dataset_path
 
     # Check if zip archive has already been downloaded
@@ -58,59 +64,25 @@ def fetch_dataset(
     return
 
 
-def load_file(
-    tracking_systems: None | str | list[str] = None,
-    tracked_points: None | dict[str, str] | dict[str, list[str]] = None,
+def load_recording(
     cohort: Literal["PFF", "PD", "MS", "HA", "COPD", "CHF"] = "PFF",
     file_name: str = "data.mat",
     dataset_path: str | Path = Path(__file__).parent / "_mobilised",
     progressbar: None | bool = None,
-) -> dict[str, Any]:
-    # Local reference to the available tracking systems and tracked points
-    _AVAILABLE_TRACKING_SYSTEMS = ["SU", "PressureInsoles_raw", "DistanceModule_raw"]
-    _AVAILABLE_TRACKED_POINTS = {
-        "SU": ["LowerBack"],
-        "PressureInsoles_raw": ["LeftFoot", "RightFoot"],
-        "DistanceModule_raw": ["LeftFoot", "RightFoot"],
-    }
+) -> NGMTRecording:
+    """Load a recording from the Mobilise-D dataset.
 
-    # Check the tracking systems
-    tracking_systems = (
-        _AVAILABLE_TRACKING_SYSTEMS if tracking_systems is None else tracking_systems
-    )
-    tracking_systems = (
-        [tracking_systems] if isinstance(tracking_systems, str) else tracking_systems
-    )
-    if any(
-        [track_sys not in _AVAILABLE_TRACKING_SYSTEMS for track_sys in tracking_systems]
-    ):
-        raise ValueError(
-            f"Invalid tracking system. Available options are: {_AVAILABLE_TRACKING_SYSTEMS}"
-        )
+    If the dataset has not yet been downloaded, then is fetched from the Zenodo repository using the pooch package.
 
-    # Set the tracked points
-    tracked_points = (
-        _AVAILABLE_TRACKED_POINTS if tracked_points is None else tracked_points
-    )
-    if any([track_sys not in tracking_systems for track_sys in tracked_points.keys()]):
-        raise ValueError(
-            f"You have specified tracked points for a tracking system that you have not requested (requested: {tracking_systems})."
-        )
-    for track_sys in tracked_points.keys():
-        if isinstance(tracked_points[track_sys], str):
-            tracked_points[track_sys] = [tracked_points[track_sys]] # type: ignore
+    Args:
+        cohort (Literal[&quot;PFF&quot;, &quot;PD&quot;, &quot;MS&quot;, &quot;HA&quot;, &quot;COPD&quot;, &quot;CHF&quot;], optional): The cohort from which data should be loaded. Defaults to "PFF".
+        file_name (str, optional): The filename of the data file. Defaults to "data.mat".
[…]
