+++ ngmt/datasets/keepcontrol.py
@@ -1,13 +1,44 @@
 import numpy as np
 import pandas as pd
-import pathlib
+from pathlib import Path
+import openneuro
 import os
 from ngmt.utils.ngmt_dataclass import NGMTRecording
 from ngmt.utils.ngmt_dataclass import REQUIRED_COLUMNS
 
+def fetch_dataset(
+    dataset_path: str | Path = Path(__file__).parent / "_keepcontrol",
+) -> None:
+    """Fetch the Keep Control dataset from the OpenNeuro repository.
 
+    Args:
+        progressbar (bool, optional): Whether to display a progressbar. Defaults to True.
+        dataset_path (str | Path, optional): The path where the dataset is stored. Defaults to Path(__file__).parent/"_keepcontrol".
+    """
+    dataset_path = Path(dataset_path) if isinstance(dataset_path, str) else dataset_path
+
+    # Check if target folder exists, if not create it
+    if not dataset_path.exists():
+        dataset_path.parent.joinpath("_keepcontrol").mkdir(parents=True, exist_ok=True)
+
+    # check if the dataset has already been downloaded (directory is not empty), if not download it
+    if not any(dataset_path.iterdir()):
+        # Download the dataset using openneuro-py
+        openneuro.download(
+            dataset="ds005258", # this is the example Keep Control dataset on OpenNeuro, maintained by Julius Welzel
+            target_dir=dataset_path,
+        )
+        
+        print(f"Dataset has been downloaded to {dataset_path}")
+
+    # print message to user if dataset has already been downloaded
+    else:
+        print(f"Dataset has already been downloaded to {dataset_path}")
+
+    return
+        
 def load_recording(
-    file_name: pathlib.Path,
+    file_name: Path,
     tracking_systems: str | list[str],
     tracked_points: str | list[str] | dict[str, str] | dict[str, list[str]],
 ):
@@ -25,6 +56,13 @@ def load_recording(
     Returns:
         NGMTRecording : An instance of the NGMTRecording dataclass containing the loaded data and channels.
     """
+
+    # Fetch the dataset if it does not exist
+    progressbar = False if not progressbar else progressbar
+    file_path = Path(dataset_path) / cohort / file_name
+    if not file_path.exists():
+        fetch_dataset()
+
     # Put tracking systems in a list
     if isinstance(tracking_systems, str):
         tracking_systems = [tracking_systems]
@@ -75,9 +113,15 @@ def load_recording(
                     )
                 ],
             ]
+            if df_data.empty: 
+                print(f"Specified tracked point not found in the tracking system {tracksys}.")
+            
             df_channels = df_channels[
-                (df_channels["tracked_point"].isin(tracked_points[tracksys]))
+                df_channels["tracked_point"].str.contains("|".join(tracked_points[tracksys]))
             ]
+
+            # Add sampling frequency to the channels DataFrame
+            df_channels["sampling_frequency"] = 200
 
             # Put data and channels in output dictionaries
             col_names = [c for c in REQUIRED_COLUMNS] + [
