+++ pyproject.toml
@@ -44,6 +44,7 @@ pyarrow = "^15.0.0"
 bids-validator = "^1.14.4"
 actipy = "^3.0.5"
 pooch = "^1.8.2"
+openneuro-py = "^2024.2.0"
 
 [tool.poetry.extras]
 docs = ["mkdocs", "mkdocs-material", "pymdownx.superfences"]
