+++ README.md
@@ -11,13 +11,15 @@
 
 Welcome to the NeuroGeriatricsMotionToolbox (NGMT). We are a Python based toolbox for processing motion data.
 
-The toolbox is aimed at motion researchers who want to use python based open source software to process their data.
-We have implemented validated algorithms in modules to process motion data, such as:
-   - Gait sequence detection (GSD)
-   - Inital contact detection (ICD)
-   - Physical activity monitoring (PAM)
-   - Postural transition detection (SSD)
-   - More to follow ...
+The toolbox is aimed at motion researchers who want to use Python-based open-source software to process their data. We have implemented validated algorithms in modules to process motion data, as shown in the table below:
+
+| Module                         | Description                                    | Data                               |
+|--------------------------------|------------------------------------------------|----------------------------------------|
+| Gait sequence detection (GSD)  | Detects gaits                      | 3D accelerations from the lower back   |
+| Initial contact detection (ICD)| Detects initial contact during gait     | 3D accelerations from the lower back   |
+| Postural transition detection (SSD) | Detects sit-to-stand and stand-to-sit movements        | 3D accelerations and gyroscope from the lower back   |
+| Physical activity monitoring (PAM) | Monitors physical activity levels          | 3D accelerations from the wrist        |
+| More to follow...              | Additional modules to be added                |                                        |
 
 The idea is that various motion data can be loaded into our dedicated dataclass which rely on principles from the [Motion-BIDS](https://bids-specification.readthedocs.io/en/latest/modality-specific-files/motion.html) standard.
 
