### 2024-07-17 a62e731
summary revision
