### 2024-07-17 d8f471f
State of the Field revision & remove unsued files
