+++ paper/paper.md
@@ -54,7 +54,7 @@ Physical mobility is an essential aspect of health, since impairment of mobility
 These modules are pivotal because they enable researchers and clinicians to extract meaningful insights from motion data captured in various environments and conditions. These modules are designed to process data from wearable devices, which offer distinct advantages over vision-based approaches. wearable devices such as IMUs provide continuous monitoring capabilities, enabling users to wear them throughout the day in diverse settings without logistical constraints posed by camera-based systems.
 
 # State of the field
-With the growing availability of digital health data, open-source implementations of relevant algorithms are increasingly becoming available. From the Mobilise-D consortium, the recommended algorithms for assessing real-world gait were released, but these algorithms were developed in MATLAB, that is not free to use [@mobilised:2023]. Likewise, an algorithm for the estimation of gait quality was released, but it is also only available in MATLAB [@gaitqualitycomposite:2016].  Alternatively, open-source, Python packages are available, for example to detect gait and extract gait features from a low back-worn inertial measurement unit (IMU) [@czech:2019], or from two feet-worn IMUs [@kuederle:2024]. NGMT builds forth on these toolboxes by providing a module software package that goes beyond the analysis of merely gait, and extends these analyses by additionally allowing for the analysis of general physical activity and other daily life-relevant movements, such as sit-to-stand and stand-to-sit transitions [@pham:2017] as well as turns [@pham:2018].
+With the growing availability of digital health data, open-source implementations of relevant algorithms are increasingly becoming available. From the Mobilise-D consortium, the recommended algorithms for assessing real-world gait were released, but these algorithms were developed in MATLAB, that is not free to use [@mobilised:2023]. Likewise, an algorithm for the estimation of gait quality was released, but it is also only available in MATLAB [@gaitqualitycomposite:2016]. Alternatively, open-source, Python packages are available, for example to detect gait and extract gait features from a low back-worn inertial measurement unit (IMU) [@czech:2019], or from two feet-worn IMUs [@kuederle:2024]. These advancements facilitate broader accessibility and usability across research and clinical applications. Additionally, innovative approaches like Mobile GaitLab focus on video input for predicting key gait parameters such as walking speed, cadence, knee flexion angle at maximum extension, and the Gait Deviation Index, leveraging open-source principles and designed to be accessible to non-computer science specialists [@kidzinski:2020; @mobile-gaitlab:2020]. Moreover, tools such as Sit2Stand and Sports2D contribute to this landscape by offering user-friendly platforms for assessing physical function through automated analysis of movements like sit-to-stand transitions and joint angles from smartphone videos (Sports2D) [@Boswell:2023; @Pagnon:2023]. NGMT builds forth on these toolboxes by providing a module software package that goes beyond the analysis of merely gait, and extends these analyses by additionally allowing for the analysis of general physical activity and other daily life-relevant movements, such as sit-to-stand and stand-to-sit transitions [@pham:2017] as well as turns [@pham:2018].
 
 # Provided Functionality
 NGMT offers a comprehensive suite of algorithms for motion data processing in neuroscience and biomechanics. Currently, the toolbox includes implementations for gait sequence detection (GSD) and initial contact detection (ICD), whereas algorithms for postural transition analysis [@pham:2017] and turns [@pham:2018] are under current development. NGMT is built on principles from the Brain Imaging Data Structure (BIDS) [@gorgolewski:2016] and for the motion analysis data are organized similar to the Motion-BIDS specifications [@jeung:2023].
