+++ paper/references.bib
@@ -1,3 +1,15 @@
+@article{Boswell:2023,
+  author = {Boswell, Melissa A and Kidzi{\'n}ski, {\L}ukasz and Hicks, Jennifer L and Uhlrich, Scott D and Falisse, Antoine and Delp, Scott L},
+  title = {Smartphone videos of the sit-to-stand test predict osteoarthritis and health outcomes in a nationwide study},
+  journal = {npj Digital Medicine},
+  volume = {6},
+  number = {1},
+  pages = {32},
+  year = {2023},
+  publisher={Nature Publishing Group UK London},
+  doi= {10.1038/s41746-023-00775-1}
+}
+
 @article{buckley:2019,
 	author = {Buckley, Christopher and Alcock, Lisa and McArdle, Ríona and Rehman, Rana Zia Ur and Del Din, Silvia and Mazzà, Claudia and Yarnall, Alison J. and Rochester, Lynn},
   title = {The {Role} of {Movement} {Analysis} in {Diagnosing} and {Monitoring} {Neurodegenerative} {Conditions}: {Insights} from {Gait} and {Postural} {Control}},
@@ -72,7 +84,19 @@
   doi={10.31234/osf.io/w6z79}
 }
 
-@ARTICLE{kuederle:2024,
+@article{kidzinski:2020,
+  author = {Kidzi{\'n}ski, {\L}ukasz and Yang, Bryan and Hicks, Jennifer L and Rajagopal, Apoorva and Delp, Scott L and Schwartz, Michael H},
+  title = {Deep neural networks enable quantitative movement analysis using single-camera videos},
+  journal = {Nature communications},
+  volume = {11},
+  number = {1},
+  pages = {4054},
+  year = {2020},
+  publisher={Nature Publishing Group UK London},
+  doi = {10.1038/s41467-020-17807-z}
+}
+
+@article{kuederle:2024,
   author = {Küderle, Arne and Ullrich, Martin and Roth, Nils and Ollenschläger, Malte and Ibrahim, Alzhraa A. and Moradi, Hamid and Richer, Robert and Seifer, Ann-Kristin and Zürl, Matthias and Sîmpetru, Raul C. and Herzer, Liv and Prossel, Dominik and Kluge, Felix and Eskofier, Bjoern M.},
   title={Gaitmap—An Open Ecosystem for IMU-Based Human Gait Analysis and Algorithm Benchmarking}, 
   journal = {IEEE Open Journal of Engineering in Medicine and Biology}, 
@@ -193,6 +217,15 @@
   url = {https://github.com/KimvanS/EstimateGaitQualityComposite}
 }
 
+@misc{mobile-gaitlab:2020,
+  author = {Kidzi{\'n}ski, {\L}ukasz and Yang, Bryan and Hicks, Jennifer L and Rajagopal, Apoorva and Delp, Scott L and Schwartz, Michael H},
+  title = {mobile-gaitlab},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/stanfordnmbl/mobile-gaitlab}
+}
+
 @misc{mobilised:2023,
[…]
