### 2024-07-17 2a11cda
name changed to Kiel Motion Analysis Toolbox (KMAT)
