+++ kmat/modules/ssd/_pham.py
@@ -2,8 +2,8 @@
 import numpy as np
 import pandas as pd
 import scipy.signal
-from ngmt.utils import preprocessing
-from ngmt.config import cfg_colors
+from kmat.utils import preprocessing
+from kmat.config import cfg_colors
 
 
 class PhamSittoStandStandtoSitDetection:
