+++ ngmt/__init__.py
@@ -1,7 +1,0 @@
-# ngmt/__init__.py
-
-"""
-NeuroGeriatrics Motion Toolbox
-"""
-
-__version__ = "0.0.1"
