+++ kmat/test/scripts/test.py
@@ -1,6 +1,6 @@
-from ngmt.datasets import mobilised
-from ngmt.modules.gsd import ParaschivIonescuGaitSequenceDetection
-from ngmt.modules.icd import ParaschivIonescuInitialContactDetection
+from kmat.datasets import mobilised
+from kmat.modules.gsd import ParaschivIonescuGaitSequenceDetection
+from kmat.modules.icd import ParaschivIonescuInitialContactDetection
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
