+++ kmat/__init__.py
@@ -1,0 +1,7 @@
+# kmat/__init__.py
+
+"""
+Kiel Motion Analysis Toolbox
+"""
+
+__version__ = "0.0.1"
