+++ docs/examples/basic_00_intro_ngmt.md
@@ -1,0 +1,455 @@
+
+**Author:** Robbin Romijnders  
+**Last update:** Tue 16 Jan 2024
+
+## Learning objectives
+By the end of this tutorial:
+- you can load data from a recording that belongs to one of the available datasets,
+- you know which attributes are available for an instance of the `KMATRecording`
+- you can do some basic selecting and slicing of data
+
+## Imports
+
+We start by importing some Python libraries. You should be familiar with most of them, and we will not discuss them here.
+
+
+```python
+import numpy as np
+import matplotlib.pyplot as plt
+import os
+from pathlib import Path
+from kmat.datasets import mobilised
+```
+
+## Import data
+
+Let us consider a single recording, namely of the randomly selected subject `sub-3011` from the `Mobilise-D` dataset, and load the data. For that we use the `load_recording()` function that is available in the `kmat.datasets.mobilised` module.
+
+
+```python
+# Set the filepath
+file_path = Path(os.getcwd()).parent.joinpath("examples","data","exMobiliseFreeLiving.mat")
+
+# Load the recording
+recording = mobilised.load_recording(
+    file_name=file_path, tracking_systems=["SU", "SU_INDIP"], 
+    tracked_points=["LowerBack", "LeftFoot", "RightFoot"]
+)
+```
+
+We have loaded the data for two tracking systems, `SU` and `SU_INDIP`, and we have specified three tracked points. The data is assigned to the variable `recording`, so let us take a look at what we have got.
+
+
+```python
+recording.__dict__
+```
+
+
+
+
+    {'data': {'SU':          LowerBack_ACCEL_x  LowerBack_ACCEL_y  LowerBack_ACCEL_z  \
+      0                 0.942178           0.007510          -0.288112   
+      1                 0.942810           0.010240          -0.289560   
+      2                 0.943393           0.006346          -0.295584   
+      3                 0.940556           0.002275          -0.291497   
+      4                 0.941915           0.005847          -0.292111   
+      ...                    ...                ...                ...   
+      1074280           0.971606           0.054683          -0.130928   
+      1074281           0.970852           0.056564          -0.135593   
+      1074282           0.969576           0.056643          -0.132909   
+      1074283           0.973411           0.056775          -0.136337   
+      1074284           0.972761           0.054689          -0.136837   
+      
+               LowerBack_GYRO_x  LowerBack_GYRO_y  LowerBack_GYRO_z  \
+      0                0.137440          0.916976          0.332333   
+      1                0.137510          0.676277          0.561320   
+      2                0.252010          0.595940          0.389748   
+      3               -0.091394          0.550077          0.469760   
+      4               -0.183270          0.630187          0.435476   
+      ...                   ...               ...               ...   
+      1074280          0.607320          0.401071          0.217741   
+      1074281          0.389616          0.412534          0.458358   
+      1074282          0.469829          0.492739          0.297937   
+      1074283          0.561494          0.355235          0.275020   
+      1074284          0.275020          0.423989          0.286479   
+      
+               LowerBack_MAGN_x  LowerBack_MAGN_y  LowerBack_MAGN_z  \
+      0               -3.220996         -5.058998         -7.994002   
+      1               -3.224997         -5.059000         -7.991002   
+      2               -3.229996         -5.059000         -7.986004   
+      3               -3.234996         -5.056002         -7.980005   
+      4               -3.238997         -5.052003         -7.973006   
+      ...                   ...               ...               ...   
+      1074280         -2.690000         -3.682000         -7.408000   
+      1074281         -2.690000         -3.683000         -7.406000   
[…]
