### 2024-07-17 b6ab3ae
Revert "name changed to Kiel Motion Analysis Toolbox (KMAT)"
This reverts commit 2a11cda3ac0e2ec15585a45c4931ae6f623cac42.
