+++ ngmt/__init__.py
@@ -1,0 +1,7 @@
+# ngmt/__init__.py
+
+"""
+NeuroGeriatrics Motion Toolbox
+"""
+
+__version__ = "0.0.1"
