### 2024-07-17 0fd4bc5
renaming to Kiel Motion Analysis Toolbox
