+++ kielmat/test/test_calc.py
@@ -1,12 +1,12 @@
 # Introduction and explanation regarding the test suite
 """
-This code is a test suite for various signal processing and analysis functions which exist in the NGMT toolbox. 
+This code is a test suite for various signal processing and analysis functions which exist in the KielMAT toolbox. 
 It employs pytest, a Python testing framework, to verify the correctness of these functions. 
 Here's a brief explanation of the code structure:
 
 1. Import necessary libraries, pytest and the functions to be tested.
 2. Generate a random input signal for testing purposes.
-3. Define a series of test functions, each targeting a specific function from the 'ngmt.utils.preprocessing' module.
+3. Define a series of test functions, each targeting a specific function from the 'kielmat.utils.preprocessing' module.
 4. Inside each test function, we validate the correctness of the corresponding function and its inputs.
 5. We make use of 'assert' statements to check that the functions return expected results.
 6. The code is organized for clarity and maintainability.
@@ -17,7 +17,7 @@ To run the tests, follow these steps:
 2. Run this script, and pytest will execute all the test functions.
 3. Any failures in tests will be reported as failed with red color, and also the number of passed tests will be represented with green color.
 
-By running these tests, the reliability and correctness of the signal processing functions in the 'ngmt.utils.preprocessing' module will be ensured.
+By running these tests, the reliability and correctness of the signal processing functions in the 'kielmat.utils.preprocessing' module will be ensured.
 """
 
 # Import necessary libraries and functions to be tested.
@@ -26,7 +26,7 @@ import numpy as np
 import warnings
 import numpy.testing as npt
 import pytest
-from ngmt.utils.preprocessing import (
+from kielmat.utils.preprocessing import (
     resample_interpolate,
     lowpass_filter,
     highpass_filter,
@@ -51,7 +51,7 @@ from ngmt.utils.preprocessing import (
     pham_plot_results,
     process_postural_transitions_stationary_periods,
 )
-from ngmt.utils.quaternion import (
+from kielmat.utils.quaternion import (
     quatinv,
     quatnormalize,
     quatnorm,
@@ -140,7 +140,7 @@ def test_resample_interpolate():
 # Test function for the 'lowpass_filter_savgol' function
 def test_lowpass_filter_savgol():
     """
-    Test for lowpass_filter_savgol function in the 'ngmt.utils.preprocessing' module.
+    Test for lowpass_filter_savgol function in the 'kielmat.utils.preprocessing' module.
     """
     # Test with inputs
     test_signal = random_input_signal
@@ -238,7 +238,7 @@ def test_lowpass_filter():
 # Test function for the 'lowpass_filter_fir' function
 def test_lowpass_filter_fir():
     """
-    Test for lowpass_filter_fir function in the 'ngmt.utils.preprocessing' module.
+    Test for lowpass_filter_fir function in the 'kielmat.utils.preprocessing' module.
     """
     # Test with inputs
     test_signal = random_input_signal
@@ -279,7 +279,7 @@ def test_lowpass_filter_fir():
 # Additional test cases for Savitzky-Golay filter
 def test_lowpass_filter_savgol_specific():
     """
-    Test specific parameters for Savitzky-Golay filter in the 'ngmt.utils.preprocessing' module.
+    Test specific parameters for Savitzky-Golay filter in the 'kielmat.utils.preprocessing' module.
     """
     # Test with specific parameters
     test_signal = np.ones(100)
@@ -303,7 +303,7 @@ def test_lowpass_filter_savgol_specific():
 # Additional test cases for Butterworth filter
[…]
