+++ mkdocs.yml
@@ -1,4 +1,4 @@
-site_name: NeuroGeriatricsMotionToolbox
+site_name: KielMotionAnalysisToolbox
 
 extra_css:
   - 'style.css'
