+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "kielmat"
 version = "0.0.4"
-description = "The NeuroGeriatricsMotionToolbox (KielMAT) is a Python-based toolbox for processing motion data."
+description = "The KielMotionAnalysisToolbox (KielMAT) is a Python-based toolbox for processing motion data."
 license = "MIT"
 readme = "README.md"
 homepage = "https://neurogeriatricskiel.github.io/KielMAT/"
