+++ ngmt/utils/importers.py
@@ -7,6 +7,7 @@ from ngmt.utils.file_io import get_unit_from_type
 from pathlib import Path
 from typing import Union
 
+
 # Importer for Axivity
 def import_axivity(file_path: str, tracked_point: str):
     """
@@ -67,8 +68,7 @@ def import_axivity(file_path: str, tracked_point: str):
 
 # Importer for APDM Mobility Lab for different versions
 def import_apdm_mobilitylab(
-    file_name: str | Path, 
-    tracked_points: str | list[str]
+    file_name: str | Path, tracked_points: str | list[str]
 ) -> tuple[pd.DataFrame, pd.DataFrame]:
     """
     Imports data from an APDM Mobility Lab system's different versions from the specified file path.
@@ -95,18 +95,18 @@ def import_apdm_mobilitylab(
 
     # Mapping of sensor types to make them consistent with NGMT dataclass definition
     sensor_type_mapping = {
-        'Accelerometer': 'ACCEL',
-        'Gyroscope': 'GYRO',
-        'Magnetometer': 'MAGN',
-        'Accelerometers': 'ACCEL',
-        'Gyroscopes': 'GYRO',
-        'Magnetometers': 'MAGN'
+        "Accelerometer": "ACCEL",
+        "Gyroscope": "GYRO",
+        "Magnetometer": "MAGN",
+        "Accelerometers": "ACCEL",
+        "Gyroscopes": "GYRO",
+        "Magnetometers": "MAGN",
     }
 
-    with h5py.File(file_name, 'r') as hfile:
+    with h5py.File(file_name, "r") as hfile:
         # Check if there is an attribute or dataset that indicates the version
-        if 'FileFormatVersion' in hfile.attrs:
-            version = hfile.attrs['FileFormatVersion']
+        if "FileFormatVersion" in hfile.attrs:
+            version = hfile.attrs["FileFormatVersion"]
         else:
             raise ValueError("Version attribute not found in the h5 file.")
 
@@ -123,23 +123,33 @@ def import_apdm_mobilitylab(
 
         # Check the version
         if version == 5:
-            sensors_group = hfile['Sensors']
+            sensors_group = hfile["Sensors"]
 
             # Structure for version 5
             monitor_labels = list(sensors_group.keys())
             sensor_to_label = {
-                sensor_id: sensors_group[sensor_id]['Configuration'].attrs['Label 0'].decode('utf-8')
+                sensor_id: sensors_group[sensor_id]["Configuration"]
+                .attrs["Label 0"]
+                .decode("utf-8")
                 for sensor_id in monitor_labels
             }
 
             # Convert tracked_points to sensor IDs using sensor_to_label mapping
-            tracked_points = [sensor for sensor, label in sensor_to_label.items() if label in tracked_points]
+            tracked_points = [
+                sensor
+                for sensor, label in sensor_to_label.items()
+                if label in tracked_points
+            ]
 
             # Track invalid tracked points
-            invalid_tracked_points = [tp for tp in tracked_points if tp not in monitor_labels]
+            invalid_tracked_points = [
+                tp for tp in tracked_points if tp not in monitor_labels
+            ]
 
             if invalid_tracked_points:
-                raise ValueError(f"The following tracked points do not exist in sensor names: {invalid_tracked_points}")
+                raise ValueError(
+                    f"The following tracked points do not exist in sensor names: {invalid_tracked_points}"
+                )
 
             # Iterate over each sensor
             for sensor_name in monitor_labels:
@@ -147,70 +157,87 @@ def import_apdm_mobilitylab(
                     continue  # to next sensor name
 
                 sensor_data = sensors_group[sensor_name]
[…]
