### 2024-07-18 0586e6b
[FIX] example with load KeepControl data from OpenNeuro
