### 2024-07-19 092e19d
[ADD] new logo to docs
