### 2024-07-19 06b20b8
Apply Black Formatting
