+++ kielmat/datasets/keepcontrol.py
@@ -9,19 +9,77 @@ from kielmat.utils.kielmat_dataclass import REQUIRED_COLUMNS
 # Dict of valid tracked points for the Keep Control dataset for each tracking system
 VALID_TRACKED_POINTS = {
     "omc": [
-        'l_toe', 'l_heel', 'l_ank', 'l_sk1', 'l_sk2', 'l_sk3', 'l_sk4', 'l_th1', 'l_th2',
-        'l_th3', 'l_th4', 'r_toe', 'r_heel', 'r_ank', 'r_sk1', 'r_sk2', 'r_sk3', 'r_sk4',
-        'r_th1', 'r_th2', 'r_th3', 'r_th4', 'l_asis', 'r_asis', 'l_psis', 'r_psis',
-        'm_ster1', 'm_ster2', 'm_ster3', 'l_sho', 'l_ua', 'l_elbl', 'l_frm', 'l_wrr',
-        'l_wru', 'l_hand', 'r_sho', 'r_ua', 'r_elbl', 'r_frm', 'r_wrr', 'r_wru', 'r_hand',
-        'lf_hd', 'rf_hd', 'lb_hd', 'rb_hd', 'start_1', 'start_2', 'end_1', 'end_2'
+        "l_toe",
+        "l_heel",
+        "l_ank",
+        "l_sk1",
+        "l_sk2",
+        "l_sk3",
+        "l_sk4",
+        "l_th1",
+        "l_th2",
+        "l_th3",
+        "l_th4",
+        "r_toe",
+        "r_heel",
+        "r_ank",
+        "r_sk1",
+        "r_sk2",
+        "r_sk3",
+        "r_sk4",
+        "r_th1",
+        "r_th2",
+        "r_th3",
+        "r_th4",
+        "l_asis",
+        "r_asis",
+        "l_psis",
+        "r_psis",
+        "m_ster1",
+        "m_ster2",
+        "m_ster3",
+        "l_sho",
+        "l_ua",
+        "l_elbl",
+        "l_frm",
+        "l_wrr",
+        "l_wru",
+        "l_hand",
+        "r_sho",
+        "r_ua",
+        "r_elbl",
+        "r_frm",
+        "r_wrr",
+        "r_wru",
+        "r_hand",
+        "lf_hd",
+        "rf_hd",
+        "lb_hd",
+        "rb_hd",
+        "start_1",
+        "start_2",
+        "end_1",
+        "end_2",
     ],
     "imu": [
-        'head', 'sternum', 'left_upper_arm', 'left_fore_arm', 'right_upper_arm',
-        'right_fore_arm', 'pelvis', 'left_thigh', 'left_shank', 'left_foot',
-        'right_thigh', 'right_shank', 'right_foot', 'left_ankle', 'right_ankle'
-    ]
+        "head",
+        "sternum",
+        "left_upper_arm",
+        "left_fore_arm",
+        "right_upper_arm",
+        "right_fore_arm",
+        "pelvis",
+        "left_thigh",
+        "left_shank",
+        "left_foot",
+        "right_thigh",
+        "right_shank",
+        "right_foot",
+        "left_ankle",
+        "right_ankle",
+    ],
 }
+
 
 def fetch_dataset(
     dataset_path: str | Path = Path(__file__).parent / "_keepcontrol",
@@ -41,7 +99,7 @@ def fetch_dataset(
     if not any(dataset_path.iterdir()):
         # Download the dataset using openneuro-py
         openneuro.download(
-            dataset="ds005258", # this is the example Keep Control dataset on OpenNeuro, maintained by Julius Welzel
+            dataset="ds005258",  # this is the example Keep Control dataset on OpenNeuro, maintained by Julius Welzel
             target_dir=dataset_path,
         )
 
[…]
