+++ kielmat/datasets/mobilised.py
@@ -1,6 +1,9 @@
 import numpy as np
 import pandas as pd
-import pathlib
+from pathlib import Path
+from pooch import DOIDownloader
+from zipfile import ZipFile
+from typing import Literal, Any
 from kielmat.utils import matlab_loader
 from kielmat.utils.kielmat_dataclass import KielMATRecording
 
@@ -11,7 +14,7 @@ MAP_CHANNEL_TYPES = {
     "Gyr": "GYRO",
     "Mag": "MAGN",
     "Bar": "BARO",
-    # "Temp": "TEMP"
+    # "Temp": "TEMP",
 }
 
 MAP_CHANNEL_COMPONENTS = {
@@ -30,107 +33,124 @@ MAP_CHANNEL_UNITS = {
 }
 
 
-def load_recording(
-    file_name: str | pathlib.Path,
-    tracking_systems: str | list[str],
-    tracked_points: str | list[str] | dict[str, str] | dict[str, list[str]],
-):
-    """
-    Load a recording from the Mobilise-D dataset.
+def fetch_dataset(
+    progressbar: bool = True,
+    dataset_path: str | Path = Path(__file__).parent / "_mobilised",
+) -> None:
+    """Fetch the Mobilise-D dataset from the Zenodo repository.
 
     Args:
-        file_name (str or pathlib.Path ): The absolute or relative path to the data file.
-        tracking_systems (str or list of str) : A string or list of strings of tracking systems for which data are to be returned.
-        tracked_points (str or list of str or dict[str, str] or dict[str, list of str]) :
-            Defines for which tracked points data are to be returned.
-            If a string or list of strings is provided, then these will be mapped to each requested tracking system.
-            If a dictionary is provided, it should map each tracking system to either a single tracked point or a list of tracked points.
+        progressbar (bool, optional): Whether to display a progressbar. Defaults to True.
+        dataset_path (str | Path, optional): The path where the dataset is stored. Defaults to Path(__file__).parent/"_mobilised".
+    """
+    dataset_path = Path(dataset_path) if isinstance(dataset_path, str) else dataset_path
+
+    # Check if zip archive has already been downloaded
+    if not dataset_path.exists():
+        dataset_path.parent.joinpath("_mobilised").mkdir(parents=True, exist_ok=True)
+    _output_file = dataset_path.joinpath("Mobilise-D_dataset.zip")
+
+    if not _output_file.exists():
+        # Set the URL to the dataset
+        _url = "doi:10.5281/zenodo.7547125/Mobilise-D dataset_1-18-2023.zip"
+
+        # Instantiate a downloader object
+        downloader = DOIDownloader(progressbar=progressbar)
+        downloader(url=_url, output_file=_output_file, pooch=None)
+
+    # Extract the dataset
+    with ZipFile(_output_file, "r") as zip_ref:
+        zip_ref.extractall(dataset_path)
+    return
+
+
+def load_recording(
+    cohort: Literal["PFF", "PD", "MS", "HA", "COPD", "CHF"] = "PFF",
+    file_name: str = "data.mat",
+    dataset_path: str | Path = Path(__file__).parent / "_mobilised",
+    progressbar: None | bool = None,
+) -> KielMATRecording:
+    """Load a recording from the Mobilise-D dataset.
+
+    If the dataset has not yet been downloaded, then is fetched from the Zenodo repository using the pooch package.
+
+    Args:
+        cohort (Literal[&quot;PFF&quot;, &quot;PD&quot;, &quot;MS&quot;, &quot;HA&quot;, &quot;COPD&quot;, &quot;CHF&quot;], optional): The cohort from which data should be loaded. Defaults to "PFF".
[…]
