### 2024-07-19 20cff73
Initial test for loading recordings for updated Keep Control and Mobilise-D datasets
