+++ pyproject.toml
@@ -52,6 +52,7 @@ docs = ["mkdocs", "mkdocs-material", "pymdownx.superfences"]
 [tool.poetry.group.dev.dependencies]
 pytest-cov = "^4.1.0"
 pytest = "^7.4.3"
+black = "^24.4.2"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
