### 2024-07-19 2a6cab8
[ENH] use logging in favour of printing
