### 2024-07-19 adb009e
Apply Black Formatting
