### 2024-07-19 5f50fcc
[FIX] examples with current .md files in docs
