+++ docs/examples/modules_02_icd.md
@@ -35,17 +35,21 @@ import matplotlib.pyplot as plt
 import os
 from pathlib import Path
 
-from kielmat.datasets import mobilised
+from kielmat.datasets import keepcontrol
 from kielmat.modules.gsd import ParaschivIonescuGaitSequenceDetection
 from kielmat.modules.icd import ParaschivIonescuInitialContactDetection
 from kielmat.config import cfg_colors
 ```
 
+    c:\Users\User\Desktop\kiel\NGMT\.venv\lib\site-packages\tqdm\auto.py:21: TqdmWarning: IProgress not found. Please update jupyter and ipywidgets. See https://ipywidgets.readthedocs.io/en/stable/user_install.html
+      from .autonotebook import tqdm as notebook_tqdm
+    
+
 ## Data Preparation
 
-To implement the Paraschiv-Ionescu initial contact algorithm, we load example data from a congestive heart failure (CHF) cohort, which is publicly available on the Zenodo repository [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7547125.svg)](https://doi.org/10.5281/zenodo.7547125). 
+To implement the Paraschiv-Ionescu initial contact algorithm, we load example data from a KeepControl, which is publicly available on [OpenNeuro](https://openneuro.org/datasets/ds005258). 
 
-The participant was assessed for 2.5 hours in the real-world while doing different daily life activities and also was asked to perform specific tasks such as outdoor walking, walking up and down a slope and stairs and moving from one room to another [`1`].
+The participant was walking at a slow speed for a distance of 20m.
 
 #### Refertences
 
@@ -54,27 +58,21 @@ The participant was assessed for 2.5 hours in the real-world while doing differe
 
 
 ```python
-# The 'file_path' variable holds the absolute path to the data file
-file_path = Path(os.getcwd()).parent.joinpath("examples","data","exMobiliseFreeLiving.mat")
+recording = keepcontrol.load_recording(tracking_systems="imu", tracked_points="pelvis")
+```
 
-# In this example, we use "SU" as tracking_system and "LowerBack" as tracked points.
-tracking_sys = "SU"
-tracked_points = {tracking_sys: ["LowerBack"]}
+    Warning: The value of the key 'Subject' should be lower case. Converted to lower case.
+    Warning: The value of the key 'Task' should be lower case. Converted to lower case.
+    
 
-# The 'mobilised.load_recording' function is used to load the data from the specified file_path
-recording = mobilised.load_recording(
-    file_name=file_path, tracking_systems=[tracking_sys], tracked_points=tracked_points
-)
 
-# Load lower back acceleration data
-acceleration_data = recording.data[tracking_sys][
-    ["LowerBack_ACCEL_x", "LowerBack_ACCEL_y", "LowerBack_ACCEL_z"]
-]
+```python
+# find the channel with acceleration data
+chns_oi = [ch for ch in recording.channels["imu"]["name"] if "accel" in ch]
+acceleration_data = recording.data["imu"][chns_oi]
 
-# Get the corresponding sampling frequency directly from the recording
-sampling_frequency = recording.channels[tracking_sys][
-    recording.channels[tracking_sys]["name"] == "LowerBack_ACCEL_x"
-]["sampling_frequency"].values[0]
+# set the sampling frequency
+sampling_frequency = float(recording.channels["imu"]["sampling_frequency"].unique()[0])
 ```
 
 ## Visualisation of the Data
@@ -96,7 +94,7 @@ colors = cfg_colors["raw"]
 for i in range(3):
     plt.plot(
[…]
