+++ docs/examples/index.md
@@ -12,5 +12,3 @@ This is a collection of examples to explain various parts of the KielMAT toolbox
 
 ### [Example 1: Gait Sequence Detection](modules_01_gsd.md)
 ### [Example 2: Initial Contact Detection](modules_02_icd.md)
-### [Example 3: Physical Activity Monitoring](03_tutorial_physical_activity_monitoring.md)
-### [Example 4: Sit to Stand and Stand to Sit Detection](05_tutorial_sit_to_stand_stand_to_sit_detection.md)
