### 2024-07-19 63d1d55
[ADD] test for HDF5Decoder
