+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "kielmat"
-version = "0.0.4"
+version = "0.0.5"
 description = "The KielMotionAnalysisToolbox (KielMAT) is a Python-based toolbox for processing motion data."
 license = "MIT"
 readme = "README.md"
