### 2024-07-22 eb4ec51
some updates based on comments from reviewer (JOSS)
