+++ paper/references.bib
@@ -190,6 +190,18 @@
 	doi = {10.3389/fneur.2018.00652}
 }
 
+@article{van2013separating,
+  title={Separating movement and gravity components in an acceleration signal and implications for the assessment of human daily physical activity},
+  author={Van Hees, Vincent T and Gorzelniak, Lukas and Dean Le{\'o}n, Emmanuel Carlos and Eder, Martin and Pias, Marcelo and Taherian, Salman and Ekelund, Ulf and Renstr{\"o}m, Frida and Franks, Paul W and Horsch, Alexander and others},
+  journal={PloS one},
+  volume={8},
+  number={4},
+  pages={e61691},
+  year={2013},
+  publisher={Public Library of Science San Francisco, USA},
+  doi = {10.1371/journal.pone.0061691}
+}
+
 @article{warmerdam:2020,
 	author = {Warmerdam, Elke and Hausdorff, Jeffrey M and Atrsaei, Arash and Zhou, Yuhan and Mirelman, Anat and Aminian, Kamiar and Espay, Alberto J and Hansen, Clint and Evers, Luc J W and Keller, Andreas and Lamoth, Claudine and Pilotto, Andrea and Rochester, Lynn and Schmidt, Gerhard and Bloem, Bastiaan R and Maetzler, Walter},
   title = {Long-term unsupervised mobility assessment in movement disorders},
