### 2024-07-23 #6842 @JuliusWelzel
Thanks @davidpagnon for your comprehensive review.
We have updated the toolbox and made a new release, so the example datasets can now be fetched from public repos. Let us know if everything works.

> * **Data sharing:** Two large datasets are natively supported (Keep Control and Mobilize-D), and tools are provided to load them.
>   _Example files for tutorials are missing for now._

Examples for ICD and GSD have been added as part of [#102](https://github.com/neurogeriatricskiel/KielMAT/pull/102). They can be simply loaded as shown in [this example](https://neurogeriatricskiel.github.io/KielMAT/examples/modules_02_icd/).

>   _It would be good to add a line stating that it has been gathered in accordance with the Declaration of Helsinki._

This has been added in https://github.com/neurogeriatricskiel/KielMAT/commit/8fa81ad785b6a9120e2a3c9cdf9524c27acfe3ca

>   _It may be worth specifying that the expected motion data are 3D accelerations (from the wrist or the lower-back depending on the module)._

A table with an overview has been added on the landing page in

> * **Community guidelines:** Issues have been activated on the repository so the user can report issues or seek support; there are contribution guidelines on the website and in the paper, but it would be good to refer to them in the Readme.md file.

We have added a [link and statement about contributing](https://github.com/neurogeriatricskiel/KielMAT?tab=readme-ov-file#contributing) in the README.

>
> ## Software paper
> * **Summary:**  _It is not immediately clear from the paper what kind of data the toolbox uses:_ marker-based data? Video input? RGB-D? IMUs? I suppose any piece of hardware could be suitable as long as it can provide acceleration data (from the pelvis? the wrist?), but it is not stated. I would be more specific very early on, for example by specifying which type of motion data is expected L12: "NGMT offers a range of 12 algorithms for the processing of motion data in neuroscience and biomechanics".

We have updated the summary in https://github.com/neurogeriatricskiel/KielMAT/commit/a62e73108f44192dade141ce92d335053f51e768

> * **Statement of need:**   I would like more detail to be given about each of the 4 modules. I feel like someone skimming the paper could easily miss that there are actually 4 main modules, and if they do, they may not understand what they consist in. _Why are these 4 modules (+ the ones under development: "postural transition analysis" and "turns") specifically useful?_ I would also appreciate it if you provided the gist of the theory behind each module.
>   _Why are accelerometers better than vision-based approaches in your context?_ Supposedly because unlike being filmed, one can wear accelerometers all day long and at any place.
>   I do not understand the end of the sentence L33: _"DHT [...] offers a new dimension of measuring daily function, that is, performance"_

We have updated the text in various commits to address these valuable comments

>   One subsidiary remark: NGMT is just a sequence of letters that will not bear any meaning to anyone who comes across it; moreover, I think this toolbox could be useful to many more cases than the very specific field of neurogeriatric motion. _For the sake of dissemination, would it be worth finding another name?_

We thank you for pointing this out. We have change the name of the toolbox to KielMotionAnalysisToolbox (KielMAT).

> * **State of the field:** It may be interesting to cite other approaches not using acceleration-based data, such  as the Mobile gait-lab ([paper](https://www.nature.com/articles/s41467-020-17807-z) and [code](https://github.com/stanfordnmbl/mobile-gaitlab)): although it works from video input and focuses on different quantities (walking speed, cadence, maximum knee flexion, and gait deviation), the approach is globally similar in that it is open-source and meant to be easily accessible to non computer-science specialists. [Sit2Stand](https://sit2stand.ai/) and [Sports2D](https://github.com/davidpagnon/Sports2D/) are other repositories that could make it to this section.

We adapted the text in the following commit: https://github.com/neurogeriatricskiel/KielMAT/commit/d8f471f1e023a043a831aef5cedbb6a16e7568e7
