### 2024-07-23 #6842 @JuliusWelzel
@Kevin-Mattheus-Moerman We made a new release as parts of our responses to @davidpagnon. The available version is now 0.0.5.
