### 2024-07-24 #6842 @davidpagnon
Thanks for addressing my points! I tried all the modules with the new version and they all work seamlessly.

Three last remarks:

1. It would be nice if there was a progress bar and/or message telling us that the mobilize dataset is being downloaded. I did not have any of it and it took a while, so I exited. When I relaunched it it did not work (invalid zip file) so I had to find where the dataset was stored, remove it, and type the command again. I believe you would just have to set this to True and it would show the progress bar:
https://github.com/neurogeriatricskiel/KielMAT/blob/a2b757a2f785aa3f43aba72dcfe357d363737dc2/kielmat/datasets/mobilised.py#L71
   I had a timeout issue when downloading the keepcontrol dataset, but then the demo still worked so I would not worry. It may be on my end, my connection is not especially fast.
   `RuntimeError: Timeout when trying to download D:\softs\KielMAT\kielmat\datasets\_keepcontrol\sub-pp001\motion\sub-pp001_task-calibration3_tracksys-omc_motion.tsv.`

2. The user needs to pip install kielmat, but then also to `git clone https://github.com/neurogeriatricskiel/KielMAT.git`, and to `cd keilmat` or [this tutorial](https://neurogeriatricskiel.github.io/KielMAT/examples/basic_01_load_Data/) will not work. It may be worth specifying it.

3. I feel like the Installation section should be at the very beginning of the home page. Right now the user has to dig to find how to install it and that may add a small obstacle to the adoption of the software.
