### 2024-07-25 52a0386
[FIX] make download progressbar default for mobilised data
