### 2024-07-25 #6842 @JuliusWelzel
Hi,

thanks for getting back so quickly. All three points have been addressed as follows:

> 1. It would be nice if there was a progress bar and/or message telling us that the mobilize dataset is being downloaded. I did not have any of it and it took a while, so I exited. When I relaunched it it did not work (invalid zip file) so I had to find where the dataset was stored, remove it, and type the command again. I believe you would just have to set this to True and it would show the progress bar.

We made set the progressbar `True` as default in: [52a0386b1e4f2df6e8950db50abdf1f94aef5a7a](https://github.com/neurogeriatricskiel/KielMAT/commit/52a0386b1e4f2df6e8950db50abdf1f94aef5a7a)

> 2. The user needs to pip install kielmat, but then also to `git clone https://github.com/neurogeriatricskiel/KielMAT.git`, and to `cd keilmat` or [this tutorial](https://neurogeriatricskiel.github.io/KielMAT/examples/basic_01_load_Data/) will not work. It may be worth specifying it.

Good point, we now allow the data for the example to be downloaded form the online repo directly in: [d79baaa419fa487ece3f0bbc3514817a07375075](https://github.com/neurogeriatricskiel/KielMAT/commit/d79baaa419fa487ece3f0bbc3514817a07375075)

> 3. I feel like the Installation section should be at the very beginning of the home page. Right now the user has to dig to find how to install it and that may add a small obstacle to the adoption of the software.

Agreed, we moved it up on the main page in: [3a001628f85c50405528c78cf5739f574d9c4c92](https://github.com/neurogeriatricskiel/KielMAT/commit/3a001628f85c50405528c78cf5739f574d9c4c92)
