### 2024-07-31 f71a1d4
[ENH] make consited use of `subject_id` instead of `id`
