### 2024-07-31 6d4fd66
Initial pipeline for marker-based analysis
