+++ kielmat/datasets/keepcontrol.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import pathlib
 import os
-from ngmt.utils.ngmt_dataclass import NGMTRecording
+from kielmat.utils.ngmt_dataclass import NGMTRecording
 
 
 # See: https://bids-specification.readthedocs.io/en/stable/modality-specific-files/motion.html#restricted-keyword-list-for-channel-type
