+++ kielmat/modules/pam/_pam.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import numpy as np
-from ngmt.utils import preprocessing
-from ngmt.utils import viz_utils
+from kielmat.utils import preprocessing
+from kielmat.utils import viz_utils
 
 
 class PhysicalActivityMonitoring:
