+++ kielmat/modules/gsd/_paraschiv.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pandas as pd
 import scipy.signal
-from ngmt.utils import preprocessing
-from ngmt.utils import viz_utils
+from kielmat.utils import preprocessing
+from kielmat.utils import viz_utils
 
 
 class ParaschivIonescuGaitSequenceDetection:
