### 2024-08-01 bcab330
[ENH] allows additional information from `participant.tsv` and `motion.json` file in recording information
