### 2024-08-01 b04be0d
Apply Black Formatting
