+++ kielmat/test/test_modules.py
@@ -1,16 +1,5 @@
 # Introduction and explanation regarding the test suite
 """
-<<<<<<< HEAD
-This code is a test suite for various modules which exist in the NGMT toolbox. 
-It employs pytest, a Python testing framework, to verify the correctness of these modules. 
-Here's a brief explanation of the code structure:
-
-1. Import necessary libraries, pytest and the modules to be tested.
-2. Generate a random input signal for testing purposes.
-3. Inside each test module, we validate the correctness of the modules.
-4. We make use of 'assert' statements to check that the functions return expected results.
-5. The code is organized for clarity and maintainability.
-=======
 This code is a test suite for various signal processing and analysis functions which exist in the KielMAT toolbox. 
 It employs pytest, a Python testing framework, to verify the correctness of these functions. 
 Here's a brief explanation of the code structure:
@@ -21,7 +10,6 @@ Here's a brief explanation of the code structure:
 4. Inside each test module, we validate the correctness of the corresponding function and its inputs.
 5. We make use of 'assert' statements to check that the functions return expected results.
 6. The code is organized for clarity and maintainability.
->>>>>>> main
 
 To run the tests, follow these steps:
 
@@ -36,22 +24,6 @@ By running these tests, the reliability and correctness of the modules will be e
 import pytest
 import numpy as np
 import pandas as pd
-<<<<<<< HEAD
-import os
-from datetime import datetime
-from unittest.mock import patch
-from kielmat.modules.gsd import ParaschivIonescuGaitSequenceDetection
-from kielmat.modules.icd import ParaschivIonescuInitialContactDetection
-from kielmat.modules.pam import PhysicalActivityMonitoring
-from kielmat.modules.ptd import PhamPosturalTransitionDetection
-from kielmat.modules.td import PhamTurnDetection
-
-
-## Module test
-# Test funtions for gait sequence detection algorithm
-num_samples = 50000  # Number of samples
-accel_data = {
-=======
 from kielmat.modules.gsd import ParaschivIonescuGaitSequenceDetection
 from kielmat.modules.icd import ParaschivIonescuInitialContactDetection
 from kielmat.modules.pam import PhysicalActivityMonitoring
@@ -61,36 +33,23 @@ from kielmat.modules.ssd import PhamSittoStandStandtoSitDetection
 # Test for gait sequence detection algorithm
 num_samples = 50000  # Number of samples
 acceleration_data = {
->>>>>>> main
     "LowerBack_ACCEL_x": np.random.uniform(-2, 2, num_samples),
     "LowerBack_ACCEL_y": np.random.uniform(-2, 2, num_samples),
     "LowerBack_ACCEL_z": np.random.uniform(-2, 2, num_samples),
 }
-<<<<<<< HEAD
-accel_data = pd.DataFrame(accel_data)
-sampling_frequency = 100  # Sampling frequency
-
-=======
 acceleration_data = pd.DataFrame(acceleration_data)
 sampling_frequency = 100  # Sampling frequency
 
 
->>>>>>> main
 def test_gsd_detect():
     # Initialize the class
     gsd = ParaschivIonescuGaitSequenceDetection()
 
     # Call the detect method
-<<<<<<< HEAD
-    gsd.detect(data=accel_data, sampling_freq_Hz=sampling_frequency)
-    gait_sequences_ = gsd.gait_sequences_
-
-=======
[…]
