+++ kielmat/test/test_calc.py
@@ -1465,7 +1465,9 @@ def test_signal_decomposition_algorithm_invalid_input_type():
     initial_sampling_frequency = 100
 
     # Check that the function raises an error for invalid input type
-    with pytest.raises(ValueError, match="vertical_acceleration_data must be a numpy.ndarray"):
+    with pytest.raises(
+        ValueError, match="vertical_acceleration_data must be a numpy.ndarray"
+    ):
         signal_decomposition_algorithm(
             vertical_acceleration_data, initial_sampling_frequency
         )
@@ -1725,7 +1727,14 @@ def test_pham_plot_results(monkeypatch):
     monkeypatch.setattr("matplotlib.pyplot.show", mock_show)
 
     # Call the function
-    plot_postural_transitions(accel, gyro, accel_unit='g', gyro_unit='deg/s', postural_transitions_=postural_transitions_, sampling_freq_Hz=sampling_freq_Hz)
+    plot_postural_transitions(
+        accel,
+        gyro,
+        accel_unit="g",
+        gyro_unit="deg/s",
+        postural_transitions_=postural_transitions_,
+        sampling_freq_Hz=sampling_freq_Hz,
+    )
 
 
 # Test function for test_organize_and_pack_results
@@ -1774,6 +1783,7 @@ def test_organize_and_pack_results(walking_periods, peak_steps, expected_results
     actual_results_flat = [
         item for sublist in actual_results for item in sublist.items()
     ]
+
 
 # Generate test data
 quaternions = np.array(
@@ -1870,6 +1880,7 @@ def test_quatconj(q, expected):
 )
 def test_quatmultiply(q1, q2):
     result = quatmultiply(q1, q2)
+
 
 # Test function for rotm2quat function
 @pytest.mark.parametrize(
@@ -2081,23 +2092,24 @@ def test_axang2rotm(axang, expected):
     result = axang2rotm(axang)
     assert np.allclose(result, expected)
 
+
 # Test function for without plotting
 def test_pham_turn_plot_results_no_plot(monkeypatch):
     # a mock function for plt.show() that does nothing
     def mock_show():
         pass
+
     # Generate mock data
     accel = np.random.rand(100, 3)  # Acceleration data
-    gyro = np.random.rand(100, 3)   # Gyroscope data
-    detected_turns = pd.DataFrame({
-        "onset": [10.5, 20.0, 35.2],
-        "duration": [2.0, 1.5, 3.0]
-    })  # Detected turns DataFrame
+    gyro = np.random.rand(100, 3)  # Gyroscope data
+    detected_turns = pd.DataFrame(
+        {"onset": [10.5, 20.0, 35.2], "duration": [2.0, 1.5, 3.0]}
+    )  # Detected turns DataFrame
     sampling_freq_Hz = 50  # Sampling frequency
 
     # Monkeypatch plt.show() with the mock function
     monkeypatch.setattr("matplotlib.pyplot.show", mock_show)
-    
+
     # Unit of acceleration data.
     accel_unit = "g"
 
@@ -2107,132 +2119,175 @@ def test_pham_turn_plot_results_no_plot(monkeypatch):
     # Call the function
     plot_turns(accel, gyro, accel_unit, gyro_unit, detected_turns, sampling_freq_Hz)
 
+
 # Test function NGMT dataclass: case 1
 # Sample dataset for test
 @pytest.fixture
 def sample_data():
     # Sample data for testing
-    data = pd.DataFrame({
-        'Time': [0.1, 0.2, 0.3],
-        'ACCEL_x': [1.0, 2.0, 3.0],
[…]
