### 2024-08-01 b8289bf
remove SSD from modules in docs
