+++ mkdocs.yml
@@ -52,7 +52,6 @@ nav:
     - Gait Sequence Detection: modules/gsd.md
     - Initial Contact Detection: modules/icd.md
     - Physical Activity Monitoring: modules/pam.md
-    - Sit to Stand and Stand to Sit Detection: modules/ssd.md
   - Dataclass: dataclass.md
   - Datasets:
     - datasets/index.md
