+++ kielmat/test/test_calc.py
@@ -31,6 +31,10 @@ import matplotlib as plt
 import warnings
 import numpy.testing as npt
 import pytest
+from pathlib import Path
+from kielmat.utils.kielmat_dataclass import KielMATRecording
+from bids_validator import BIDSValidator
+from kielmat.utils import viz_utils
 from kielmat.utils.preprocessing import (
     resample_interpolate,
     lowpass_filter,
@@ -1688,7 +1692,7 @@ def test_gsd_plot_results_without_plot(monkeypatch):
     monkeypatch.setattr("matplotlib.pyplot.show", mock_show)
 
     # Call the function
-    plot_gait(target_sampling_freq_Hz, detected_activity_signal, gait_sequences_)
+    viz_utils.plot_gait(target_sampling_freq_Hz, detected_activity_signal, gait_sequences_)
 
 
 # Test function for pam_plot_results without plotting
@@ -1701,7 +1705,7 @@ def test_pam_plot_results_without_plot(monkeypatch):
     monkeypatch.setattr("matplotlib.pyplot.show", mock_show)
 
     # Call the function
-    plot_pam(hourly_average_data, thresholds_mg)
+    viz_utils.plot_pam(hourly_average_data, thresholds_mg)
 
 
 # Test function for test_pham_plot_results
@@ -1727,7 +1731,7 @@ def test_pham_plot_results(monkeypatch):
     monkeypatch.setattr("matplotlib.pyplot.show", mock_show)
 
     # Call the function
-    plot_postural_transitions(
+    viz_utils.plot_postural_transitions(
         accel,
         gyro,
         accel_unit="g",
@@ -2117,223 +2121,108 @@ def test_pham_turn_plot_results_no_plot(monkeypatch):
     gyro_unit = "rad/s"
 
     # Call the function
-    plot_turns(accel, gyro, accel_unit, gyro_unit, detected_turns, sampling_freq_Hz)
+    viz_utils.plot_turns(accel, gyro, accel_unit, gyro_unit, detected_turns, sampling_freq_Hz)
 
-
-# Test function NGMT dataclass: case 1
-# Sample dataset for test
+# Fixture to provide sample data for testing
 @pytest.fixture
 def sample_data():
-    # Sample data for testing
-    data = pd.DataFrame(
-        {
-            "Time": [0.1, 0.2, 0.3],
-            "ACCEL_x": [1.0, 2.0, 3.0],
-            "ACCEL_y": [0.5, 1.5, 2.5],
-            "ACCEL_z": [0.2, 0.3, 0.4],
-        }
-    )
-    return data
+    """Fixture to provide a sample DataFrame with accelerometer data."""
+    return pd.DataFrame({
+        "Time": [0.1, 0.2, 0.3], 
+        "ACCEL_x": [1.0, 2.0, 3.0], 
+        "ACCEL_y": [0.5, 1.5, 2.5],  
+        "ACCEL_z": [0.2, 0.3, 0.4]   
+    })
 
-
+# Fixture to provide sample channel information for testing
 @pytest.fixture
 def sample_channels():
-    # Sample channels data for testing
-    channels = pd.DataFrame(
-        {
-            "name": ["ACCEL_x", "ACCEL_y", "ACCEL_z"],
-            "component": ["x", "y", "z"],
-            "ch_type": ["ACCEL", "ACCEL", "ACCEL"],
-        }
-    )
-    return channels
+    """Fixture to provide a sample DataFrame with channel information."""
+    return pd.DataFrame({
+        "name": ["ACCEL_x", "ACCEL_y", "ACCEL_z"],  # Channel names
+        "component": ["x", "y", "z"],  # Component of the measurement
[…]
