### 2024-08-01 2f6dc50
Apply Black Formatting
