+++ kielmat/test/test_calc.py
@@ -1692,7 +1692,9 @@ def test_gsd_plot_results_without_plot(monkeypatch):
     monkeypatch.setattr("matplotlib.pyplot.show", mock_show)
 
     # Call the function
-    viz_utils.plot_gait(target_sampling_freq_Hz, detected_activity_signal, gait_sequences_)
+    viz_utils.plot_gait(
+        target_sampling_freq_Hz, detected_activity_signal, gait_sequences_
+    )
 
 
 # Test function for pam_plot_results without plotting
@@ -2121,51 +2123,68 @@ def test_pham_turn_plot_results_no_plot(monkeypatch):
     gyro_unit = "rad/s"
 
     # Call the function
-    viz_utils.plot_turns(accel, gyro, accel_unit, gyro_unit, detected_turns, sampling_freq_Hz)
+    viz_utils.plot_turns(
+        accel, gyro, accel_unit, gyro_unit, detected_turns, sampling_freq_Hz
+    )
+
 
 # Fixture to provide sample data for testing
 @pytest.fixture
 def sample_data():
     """Fixture to provide a sample DataFrame with accelerometer data."""
-    return pd.DataFrame({
-        "Time": [0.1, 0.2, 0.3], 
-        "ACCEL_x": [1.0, 2.0, 3.0], 
-        "ACCEL_y": [0.5, 1.5, 2.5],  
-        "ACCEL_z": [0.2, 0.3, 0.4]   
-    })
+    return pd.DataFrame(
+        {
+            "Time": [0.1, 0.2, 0.3],
+            "ACCEL_x": [1.0, 2.0, 3.0],
+            "ACCEL_y": [0.5, 1.5, 2.5],
+            "ACCEL_z": [0.2, 0.3, 0.4],
+        }
+    )
+
 
 # Fixture to provide sample channel information for testing
 @pytest.fixture
 def sample_channels():
     """Fixture to provide a sample DataFrame with channel information."""
-    return pd.DataFrame({
-        "name": ["ACCEL_x", "ACCEL_y", "ACCEL_z"],  # Channel names
-        "component": ["x", "y", "z"],  # Component of the measurement
-        "type": ["ACCEL", "ACCEL", "ACCEL"],  # Type of measurement
-        "tracked_point": ["n/a", "n/a", "n/a"],  # Point being tracked
-        "units": ["m/s^2", "m/s^2", "m/s^2"],  # Units of measurement
-        "sampling_frequency": [100, 100, 100]  # Sampling frequency
-    })
+    return pd.DataFrame(
+        {
+            "name": ["ACCEL_x", "ACCEL_y", "ACCEL_z"],  # Channel names
+            "component": ["x", "y", "z"],  # Component of the measurement
+            "type": ["ACCEL", "ACCEL", "ACCEL"],  # Type of measurement
+            "tracked_point": ["n/a", "n/a", "n/a"],  # Point being tracked
+            "units": ["m/s^2", "m/s^2", "m/s^2"],  # Units of measurement
+            "sampling_frequency": [100, 100, 100],  # Sampling frequency
+        }
+    )
+
 
 # Fixture to provide sample event data for testing
 @pytest.fixture
 def sample_events():
     """Fixture to provide a sample DataFrame with event information."""
-    return pd.DataFrame({
-        "onset": [0.1, 0.2],  # Onset times of events
-        "duration": [0.1, 0.2],  # Durations of events
-        "event_type": ["stimulus", "response"],  # Type of events
-        "name": ["Stimulus A", "Response B"]  # Names of events
-    })
+    return pd.DataFrame(
+        {
+            "onset": [0.1, 0.2],  # Onset times of events
+            "duration": [0.1, 0.2],  # Durations of events
[…]
