+++ kielmat/test/test_modules.py
@@ -24,9 +24,12 @@ By running these tests, the reliability and correctness of the modules will be e
 import pytest
 import numpy as np
 import pandas as pd
+import os
 from kielmat.modules.gsd import ParaschivIonescuGaitSequenceDetection
 from kielmat.modules.icd import ParaschivIonescuInitialContactDetection
 from kielmat.modules.pam import PhysicalActivityMonitoring
+from kielmat.modules.ptd import PhamPosturalTransitionDetection
+from kielmat.modules.td import PhamTurnDetection
 
 ## Module test
 # Test for gait sequence detection algorithm
@@ -698,23 +701,228 @@ def test_insufficient_columns():
             plot=False,
         )
 
+# Test functions for Turn detection algorithm
+def test_pham_turn_detection_algorithm():
+    # Initialize PhamTurnDetection object
+    pham = PhamTurnDetection()
 
-# Tests for sit to stand and stand to sit detection algorithm
+    # Sample accelerometer and gyroscope data
+    current_dir = os.path.dirname(os.path.abspath(__file__))
+    test_data_file_path = os.path.join(current_dir, "test_data_pham_algorithms.csv")
+    sample_data = pd.read_csv(test_data_file_path, index_col=0)
+    sampling_freq = 200
+    
+    # Check that the target sampling frequency is positive
+    assert sampling_freq > 0, "Assertion Error: Sampling frequency should be greater than 0."
+
+    # Check that the input data is a pandas DataFrame
+    assert isinstance(sample_data, pd.DataFrame), "Input data must be a pandas DataFrame"
+
+    # Perform detection
+    pham.detect(sample_data, accel_unit='g', gyro_unit='deg/s', gyro_vertical="pelvis_GYRO_x", sampling_freq_Hz=sampling_freq, plot_results=False)
+
+    # Perform spatio-temporal detection
+    pham.spatio_temporal_parameters()
+
+def test_invalid_plot_results_pham_td():
+    # Initialize PhamTurnDetection object
+    pham = PhamTurnDetection()
+    
+    # Sample accelerometer and gyroscope data
+    current_dir = os.path.dirname(os.path.abspath(__file__))
+    test_data_file_path = os.path.join(current_dir, "test_data_pham_algorithms.csv")
+    sample_data = pd.read_csv(test_data_file_path, index_col=0)
+    
+    # Test with invalid plot_results
+    invalid_plot_results = "invalid"
+    with pytest.raises(ValueError):
+        pham.detect(data=sample_data,accel_unit='g', gyro_unit='deg/s', gyro_vertical='pelvis_GYRO_x', sampling_freq_Hz=200, plot_results=invalid_plot_results)
+
 @pytest.fixture
-def sample_data():
-    # Generate sample data
-    num_samples = 50000  # Number of samples
-    sample_data = pd.DataFrame(
-        {
-            "Acc_X": np.random.uniform(-2, 2, num_samples),
-            "Acc_Y": np.random.uniform(-2, 2, num_samples),
-            "Acc_Z": np.random.uniform(-2, 2, num_samples),
-            "Gyro_X": np.random.uniform(-150, 150, num_samples),
-            "Gyro_Y": np.random.uniform(-150, 150, num_samples),
-            "Gyro_Z": np.random.uniform(-150, 150, num_samples),
-        }
-    )
-    return sample_data
+def invalid_data():
+    # Create invalid data with less than 6 columns
+    data = pd.DataFrame(np.random.rand(100, 5), columns=['Accel_X', 'Accel_Y', 'Accel_Z', 'Gyro_X', 'Gyro_Y'])
+    return data
+
[…]
