### 2024-08-01 b503046
Apply Black Formatting
