+++ kielmat/test/test_modules.py
@@ -701,6 +701,7 @@ def test_insufficient_columns():
             plot=False,
         )
 
+
 # Test functions for Turn detection algorithm
 def test_pham_turn_detection_algorithm():
     # Initialize PhamTurnDetection object
@@ -711,51 +712,82 @@ def test_pham_turn_detection_algorithm():
     test_data_file_path = os.path.join(current_dir, "test_data_pham_algorithms.csv")
     sample_data = pd.read_csv(test_data_file_path, index_col=0)
     sampling_freq = 200
-    
+
     # Check that the target sampling frequency is positive
-    assert sampling_freq > 0, "Assertion Error: Sampling frequency should be greater than 0."
+    assert (
+        sampling_freq > 0
+    ), "Assertion Error: Sampling frequency should be greater than 0."
 
     # Check that the input data is a pandas DataFrame
-    assert isinstance(sample_data, pd.DataFrame), "Input data must be a pandas DataFrame"
+    assert isinstance(
+        sample_data, pd.DataFrame
+    ), "Input data must be a pandas DataFrame"
 
     # Perform detection
-    pham.detect(sample_data, accel_unit='g', gyro_unit='deg/s', gyro_vertical="pelvis_GYRO_x", sampling_freq_Hz=sampling_freq, plot_results=False)
+    pham.detect(
+        sample_data,
+        accel_unit="g",
+        gyro_unit="deg/s",
+        gyro_vertical="pelvis_GYRO_x",
+        sampling_freq_Hz=sampling_freq,
+        plot_results=False,
+    )
 
     # Perform spatio-temporal detection
     pham.spatio_temporal_parameters()
 
+
 def test_invalid_plot_results_pham_td():
     # Initialize PhamTurnDetection object
     pham = PhamTurnDetection()
-    
+
     # Sample accelerometer and gyroscope data
     current_dir = os.path.dirname(os.path.abspath(__file__))
     test_data_file_path = os.path.join(current_dir, "test_data_pham_algorithms.csv")
     sample_data = pd.read_csv(test_data_file_path, index_col=0)
-    
+
     # Test with invalid plot_results
     invalid_plot_results = "invalid"
     with pytest.raises(ValueError):
-        pham.detect(data=sample_data,accel_unit='g', gyro_unit='deg/s', gyro_vertical='pelvis_GYRO_x', sampling_freq_Hz=200, plot_results=invalid_plot_results)
+        pham.detect(
+            data=sample_data,
+            accel_unit="g",
+            gyro_unit="deg/s",
+            gyro_vertical="pelvis_GYRO_x",
+            sampling_freq_Hz=200,
+            plot_results=invalid_plot_results,
+        )
+
 
 @pytest.fixture
 def invalid_data():
     # Create invalid data with less than 6 columns
-    data = pd.DataFrame(np.random.rand(100, 5), columns=['Accel_X', 'Accel_Y', 'Accel_Z', 'Gyro_X', 'Gyro_Y'])
+    data = pd.DataFrame(
+        np.random.rand(100, 5),
+        columns=["Accel_X", "Accel_Y", "Accel_Z", "Gyro_X", "Gyro_Y"],
+    )
     return data
+
 
 def test_data_shape_invalid(invalid_data):
     # Initialize PhamTurnDetection object
     pham = PhamTurnDetection()
-    
+
     # Test invalid data shape
     with pytest.raises(ValueError, match="Input data must have 6 columns"):
-        pham.detect(data=invalid_data, accel_unit='g', gyro_unit='deg/s', gyro_vertical='Gyro_X', sampling_freq_Hz=100)
+        pham.detect(
+            data=invalid_data,
+            accel_unit="g",
+            gyro_unit="deg/s",
[…]
