+++ kielmat/modules/ptd/_pham.py
@@ -4,7 +4,7 @@ import pandas as pd
 import scipy.signal
 from kielmat.utils import preprocessing
 from kielmat.utils import viz_utils
-from typing import Optional, Self
+from typing import Optional, Union, Tuple, List
 
 
 class PhamPosturalTransitionDetection:
@@ -127,7 +127,7 @@ class PhamPosturalTransitionDetection:
         tracking_system: Optional[str] = None,
         tracked_point: Optional[str] = None,
         plot_results: bool = False,
-    ) -> Self:
+    ) -> 'PhamPosturalTransitionDetection':
         """
         Detects postural transitions based on the input accelerometer and gyro data.
 
@@ -483,7 +483,7 @@ class PhamPosturalTransitionDetection:
         # Return an instance of the class
         return self
 
-    def spatio_temporal_parameters(self: Self) -> None:
+    def spatio_temporal_parameters(self) -> None:
         """
         Extracts spatio-temporal parameters of the detected postural transitions.
 
