### 2024-08-01 757c8a7
Apply Black Formatting
