### 2024-08-01 e07681a
Apply Black Formatting
