### 2024-08-01 eaaf3bb
Apply Black Formatting
