### 2024-08-01 c556ccb
remove unwanted examples from doc
