+++ docs/examples/index.md
@@ -10,10 +10,6 @@ This is a collection of examples to explain various parts of the KielMAT toolbox
 
 ## Modules of KielMAT
 
-### [Example 1: Gait Sequence Detection](01_tutorial_gait_sequence_detection.md)
-### [Example 2: Initial Contact Detection](02_tutorial_initial_contact_detection.md)
-### [Example 3: Physical Activity Monitoring](03_tutorial_physical_activity_monitoring.md)
-### [Example 4: Postural Transition Detection](modules_04_ptd.md)
-### [Example 5: Turn Detection](modules_05_td.md)
 ### [Example 1: Gait Sequence Detection](modules_01_gsd.md)
 ### [Example 2: Initial Contact Detection](modules_02_icd.md)
+
