### 2024-08-01 be1566a
fix smaller docstring error for docs
