+++ kielmat/test/test_datasets.py
@@ -4,9 +4,13 @@ import pytest
 from unittest.mock import patch, MagicMock
 from pathlib import Path
 from kielmat.datasets.mobilised import load_recording as mobilised_load_recording
-from kielmat.datasets.keepcontrol import load_recording as keepcontrol_load_recording, fetch_dataset as keepcontrol_fetch_dataset
+from kielmat.datasets.keepcontrol import (
+    load_recording as keepcontrol_load_recording,
+    fetch_dataset as keepcontrol_fetch_dataset,
+)
 from kielmat.datasets.fairpark import load_recording as fairpark_load_recording
 from kielmat.utils.kielmat_dataclass import KielMATRecording
+
 
 # Test for keepcontrol dataset
 @pytest.fixture
@@ -14,10 +18,12 @@ def mock_download():
     with patch("openneuro.download") as mock:
         yield mock
 
+
 @pytest.fixture
 def mock_zipfile():
     with patch("zipfile.ZipFile") as mock:
         yield mock
+
 
 @pytest.fixture
 def temp_dataset_path(tmp_path):
@@ -26,6 +32,7 @@ def temp_dataset_path(tmp_path):
     path.mkdir(parents=True, exist_ok=True)
     return path
 
+
 def test_fetch_dataset_creates_directory(temp_dataset_path, mock_download):
     """Test that fetch_dataset creates the directory if it does not exist."""
     # Set up the mock to simulate the download action
@@ -33,31 +40,34 @@ def test_fetch_dataset_creates_directory(temp_dataset_path, mock_download):
 
     # Call the function under test
     keepcontrol_fetch_dataset(dataset_path=temp_dataset_path)
-    
+
     # Assert that the directory was created
     assert temp_dataset_path.exists()
+
 
 def test_fetch_dataset_no_download_needed(temp_dataset_path, mock_download):
     """Test that fetch_dataset does not download if the directory already contains files."""
     # Create a file in the directory to simulate that it is not empty
     (temp_dataset_path / "some_file").touch()
-    
+
     # Call the function under test
     keepcontrol_fetch_dataset(dataset_path=temp_dataset_path)
-    
+
     # Assert that the download was not called
     mock_download.assert_not_called()
+
 
 def test_zipfile_validity(temp_dataset_path, mock_zipfile):
     """Test that the ZIP file is valid."""
     # Ensure the ZIP file exists
     zip_file_path = temp_dataset_path / "KeepControl_dataset.zip"
     zip_file_path.touch()
-    
+
     # Set up the mock to simulate a valid zip file
     mock_zipfile.return_value.__enter__.return_value.testzip = MagicMock()
 
-@patch('kielmat.datasets.keepcontrol.fetch_dataset')
+
+@patch("kielmat.datasets.keepcontrol.fetch_dataset")
 def test_keepcontrol_load_recording_missing_dataset(mock_keepcontrol_fetch_dataset):
     dataset_path = Path("path/to/keepcontrol")
     id = "sub001"
@@ -65,17 +75,18 @@ def test_keepcontrol_load_recording_missing_dataset(mock_keepcontrol_fetch_datas
     tracking_systems = ["imu", "omc"]
 
     # Simulate that the file does not exist initially
-    with patch('pathlib.Path.exists', return_value=False):
+    with patch("pathlib.Path.exists", return_value=False):
         recording = keepcontrol_load_recording(
             dataset_path=dataset_path,
             id=id,
             task=task,
-            tracking_systems=tracking_systems
+            tracking_systems=tracking_systems,
         )
 
     # Assert that fetch_dataset was called
[…]
