### 2024-08-02 32a5316
Revert "Merge branch 'main' of https://github.com/neurogeriatricskiel/KielMAT"
This reverts commit e0a7ec4398bb5712dd6a1b0021b1f84f1fee8022, reversing
changes made to 4f5229585112f1581428f9b4daa47255d373766c.
