+++ kielmat/test/test_datasets.py
@@ -3,20 +3,20 @@ import pandas as pd
 import pytest
 from unittest.mock import patch, MagicMock
 from pathlib import Path
-from kielmat.datasets.mobilised import load_recording as mobilised_load_recording
-from kielmat.datasets.keepcontrol import load_recording as keepcontrol_load_recording, fetch_dataset as keepcontrol_fetch_dataset
+from kielmat.datasets.mobilised import (
+    load_recording as mobilised_load_recording, 
+    fetch_dataset as mobilised_fetch_dataset)
+from kielmat.datasets.keepcontrol import (
+    load_recording as keepcontrol_load_recording,
+    fetch_dataset as keepcontrol_fetch_dataset
+)
 from kielmat.datasets.fairpark import load_recording as fairpark_load_recording
-from kielmat.utils.kielmat_dataclass import KielMATRecording
+from kielmat.utils.kielmat_dataclass import KielMATRecording, VALID_COMPONENT_TYPES, VALID_CHANNEL_STATUS_VALUES
 
 # Test for keepcontrol dataset
 @pytest.fixture
 def mock_download():
     with patch("openneuro.download") as mock:
-        yield mock
-
-@pytest.fixture
-def mock_zipfile():
-    with patch("zipfile.ZipFile") as mock:
         yield mock
 
 @pytest.fixture
@@ -75,14 +75,14 @@ def test_keepcontrol_load_recording_missing_dataset(mock_keepcontrol_fetch_datas
 
     # Assert that fetch_dataset was called
     mock_keepcontrol_fetch_dataset.assert_called_once()
-    assert isinstance(recording, KielMATRecording) is False  
+    assert recording is None  # Recording should be None if dataset is missing
 
 def test_tracking_systems_conversion():
     """Test that tracking systems are converted to a list."""
     tracking_systems = "imu"
     expected = ["imu"]
 
-    # Call the function or code block that converts tracking systems to a list
+    # Convert tracking_systems to a list if it is a string
     if isinstance(tracking_systems, str):
         tracking_systems = [tracking_systems]
 
@@ -94,7 +94,7 @@ def test_tracked_points_conversion():
     tracked_points = "head"
     expected = {"imu": ["head"], "omc": ["head"]}
 
-    # Call the function or code block that converts tracked points to a dictionary
+    # Convert tracked_points to a dictionary if it is a list or string
     if isinstance(tracked_points, str):
         tracked_points = [tracked_points]
     if isinstance(tracked_points, list):
@@ -114,11 +114,26 @@ def test_multiple_files_found(mock_warning, temp_dataset_path):
     file_path2 = temp_dataset_path / "sub-pp001" / "sub-pp001_task-walkSlow_tracksys-imu_motion2.tsv"
     file_path2.touch()
 
-    # Call the function or code block that handles file loading
-    file_name = list(temp_dataset_path.glob("sub-pp001/motion/sub-pp001_task-walkSlow_tracksys-imu_*motion.tsv"))
-    if len(file_name) > 1:
-        mock_warning.assert_called()
-        assert "Multiple files found for ID pp001, task walkSlow, and tracking system imu." in [call[0][0] for call in mock_warning.call_args_list]
+
+@patch('pandas.read_csv')
+def test_load_recording_no_files_found(mock_read_csv, temp_dataset_path):
+    """Test the case when no files are found for the given ID, task, and tracking system."""
+    dataset_path = temp_dataset_path / "sub-pp001" / "motion"
+    dataset_path.mkdir(parents=True, exist_ok=True)
+
+    # Mock read_csv to simulate that no files are found
[…]
