+++ kielmat/test/test_utils.py
@@ -7,11 +7,12 @@ from tempfile import NamedTemporaryFile
 from kielmat.utils.matlab_loader import HDF5Decoder, convert_mat_to_dict, load_matlab
 import importlib.resources as resources
 
+
 @pytest.fixture
 def temp_files():
     temp_mat_file = NamedTemporaryFile(delete=False, suffix=".mat")
     temp_h5_file = NamedTemporaryFile(delete=False, suffix=".h5")
-    
+
     # Ensure files are closed before tests start
     temp_mat_file.close()
     temp_h5_file.close()
@@ -22,6 +23,7 @@ def temp_files():
     os.remove(temp_mat_file.name)
     os.remove(temp_h5_file.name)
 
+
 def test_initialization():
     decoder = HDF5Decoder(load_only_su=True)
     assert decoder.load_only_su
@@ -29,18 +31,20 @@ def test_initialization():
     decoder = HDF5Decoder(load_only_su=False)
     assert not decoder.load_only_su
 
+
 def test_mat2dict_mat(temp_files):
     temp_mat_file, _ = temp_files
     with h5py.File(temp_mat_file, "w") as f:
         f.create_dataset("a", data=[1, 2, 3])
         f.create_dataset("b", data=[4, 5, 6])
         f.attrs["MATLAB_class"] = np.string_("struct")
-    
+
     decoder = HDF5Decoder()
     with h5py.File(temp_mat_file, "r") as f:
         result = decoder.mat2dict(f)
         assert "a" in result
         assert "b" in result
+
 
 def test_mat2dict_h5(temp_files):
     _, temp_h5_file = temp_files
@@ -53,17 +57,21 @@ def test_mat2dict_h5(temp_files):
         with pytest.raises(NotImplementedError):
             decoder.mat2dict(f)
 
+
 def test_unpack_mat(temp_files):
     temp_mat_file, _ = temp_files
     with h5py.File(temp_mat_file, "w") as f:
         group = f.create_group("group")
-        dataset = group.create_dataset("dataset", data=np.array([1, 2, 3], dtype="float64"))
+        dataset = group.create_dataset(
+            "dataset", data=np.array([1, 2, 3], dtype="float64")
+        )
         dataset.attrs["MATLAB_class"] = np.string_("double")
 
     decoder = HDF5Decoder()
     with h5py.File(temp_mat_file, "r") as f:
         result = decoder.unpack_mat(f["group/dataset"], depth=0)
     np.testing.assert_array_equal(result, [1, 2, 3])
+
 
 def test_convert_mat(temp_files):
     temp_mat_file, _ = temp_files
@@ -79,36 +87,44 @@ def test_convert_mat(temp_files):
 
     np.testing.assert_array_equal(result, [1, 2, 3])
 
+
 def test_has_refs_without_references():
-    file_name = 'test.h5'
-    with h5py.File(file_name, 'w') as f:
-        dset = f.create_dataset('dataset', data=np.array([[1, 2]]))
+    file_name = "test.h5"
+    with h5py.File(file_name, "w") as f:
+        dset = f.create_dataset("dataset", data=np.array([[1, 2]]))
         decoder = HDF5Decoder()
         assert decoder._has_refs(dset) is False
+
 
 def test_convert_mat_char(temp_files):
     temp_mat_file, _ = temp_files
     with h5py.File(temp_mat_file, "w") as f:
-        dataset = f.create_dataset("char_dataset", data=np.array([72, 101, 108, 108, 111], dtype="uint8"))
+        dataset = f.create_dataset(
+            "char_dataset", data=np.array([72, 101, 108, 108, 111], dtype="uint8")
+        )
[…]
