+++ kielmat/test/test_datasets.py
@@ -9,8 +9,11 @@ from kielmat.datasets.keepcontrol import (
     fetch_dataset as keepcontrol_fetch_dataset,
 )
 from kielmat.datasets.fairpark import load_recording as fairpark_load_recording
-from kielmat.utils.kielmat_dataclass import KielMATRecording, VALID_COMPONENT_TYPES, VALID_CHANNEL_STATUS_VALUES
-
+from kielmat.utils.kielmat_dataclass import (
+    KielMATRecording,
+    VALID_COMPONENT_TYPES,
+    VALID_CHANNEL_STATUS_VALUES,
+)
 
 
 # Test for keepcontrol dataset
