+++ docs/examples/modules_04_ptd.md
@@ -2,13 +2,13 @@
 
 **Author:** Masoud Abedinifar
 
-**Last update:** Mon 17 June 2024
+**Last update:** Mon 5 August 2024
 
 ## Learning objectives
 By the end of this tutorial:
 
-- You can load data from [`keepcontrol`](https://github.com/neurogeriatricskiel/NGMT/tree/main/ngmt/datasets/keepcontrol.py) which is one of available datasets.
-- Apply the [`Postural Transition Detection`](https://github.com/neurogeriatricskiel/NGMT/tree/main/ngmt/modules/ptd/_pham.py) algorithm.
+- You can load data from [`keepcontrol`](https://github.com/neurogeriatricskiel/KielMAT/blob/main/kielmat/datasets/keepcontrol.py) which is one of available datasets.
+- Apply the [`Postural Transition Detection`](https://github.com/neurogeriatricskiel/KielMAT/blob/main/kielmat/modules/ptd/_pham.py) algorithm.
 - Visualize the results of the algorithm.
 - Extract spatio-temporal parameters of the detected postural transitions.
 - Interpret the detected postural transitions for further analysis.
@@ -17,7 +17,7 @@ By the end of this tutorial:
 
 This example can be referenced by citing the package.
 
-The example illustrates how to use Pham Postural Transition Detection algorithm to detect postural transitions (e.g., sit to stand and stand to sit movements) using body acceleration and gyro data recorded with a lower back IMU sensor. The Pham Postural Transition detection algorithm is implemented using [`ngmt.modules.ptd._pham`](https://github.com/neurogeriatricskiel/NGMT/tree/main/ngmt/modules/ptd/_pham.py). This algorithm is based on the research of Pham et al [1].
+The example illustrates how to use Pham Postural Transition Detection algorithm to detect postural transitions (e.g., sit to stand and stand to sit movements) using body acceleration and gyro data recorded with a lower back IMU sensor. The Pham Postural Transition detection algorithm is implemented using [`kielmat.modules.ptd._pham`](https://github.com/neurogeriatricskiel/KielMAT/blob/main/kielmat/modules/ptd/_pham.py). This algorithm is based on the research of Pham et al [1].
 
 This algorithm aims to detect postural transitions (e.g., sit to stand or stand to sit movements) using accelerometer and gyroscope data collected from a lower back inertial measurement unit (IMU) sensor.
 
@@ -35,15 +35,17 @@ If requested (`plot_results` set to True), it generates plots of the acceleromet
 [1] Pham et al. (2018). Validation of a Lower Back "Wearable"-Based Sit-to-Stand and  Stand-to-Sit Algorithm for Patients With Parkinson's Disease and Older Adults in a Home-Like  Environment. Frontiers in Neurology, 9, 652. https://doi.org/10.3389/fneur.2018.00652
 
 ## Import libraries
-The necessary libraries such as numpy, matplotlib.pyplot, dataset and Pham Postural Transition Detection algorithm are imported. Make sure that you have all the required libraries and modules installed before running this code. You also may need to install the `ngmt` library and its dependencies if you haven't already.
+The necessary libraries such as numpy, matplotlib.pyplot, dataset and Pham Postural Transition Detection algorithm are imported. Make sure that you have all the required libraries and modules installed before running this code. You also may need to install the `kielmat` library and its dependencies if you haven't already.
 
 
 ```python
 import numpy as np
 import pandas as pd
+import os
 import matplotlib.pyplot as plt
-from ngmt.datasets import keepcontrol
[…]
