+++ docs/examples/modules_05_td.md
@@ -97,36 +97,36 @@ print(f"gyro_data ({gyro_unit}): {gyro_data}")
 accel_data (g):
 
                         pelvis_ACCEL_x      pelvis_ACCEL_y      pelvis_ACCEL_z
-            0           0.983901           -0.032714            0.164542
-            1           0.992177           -0.027832            0.171870
-            2           0.993660           -0.026852            0.163584
-            3           0.991220           -0.025391            0.171870
-            4           0.984858           -0.030754            0.167977
+            0           0.920901           -0.047850           -0.400888
+            1           0.919441           -0.051282           -0.392583
+            2           0.922828           -0.047359           -0.392093
+            3           0.926741           -0.048830           -0.384279
+            4           0.918973           -0.053218           -0.397947
             ...         ...                 ...                 ...
-            7141        0.984380            0.115229            0.228036
-            7142        0.982897            0.116690            0.218751
-            7143        0.966298            0.100583            0.217772
-            7144        0.957019            0.091299            0.211923
-            7145        0.957019            0.089839            0.211423
+            2903        0.966803           -0.027822           -0.279782
+            2904        0.957517           -0.035152           -0.285636
+            2905        0.960437           -0.034171           -0.291979
+            2906        0.962890           -0.036623           -0.299794
+            2907        0.963883           -0.038584           -0.294921
 
-[7146 rows x 3 columns]
+[2908 rows x 3 columns]
 
 gyro_data (deg/s):        
 
                         pelvis_GYRO_x     pelvis_GYRO_y     pelvis_GYRO_z
-            0           0.351765         -0.437449         -0.087013
-            1          -0.087941         -0.698968          0.174026
-            2           0.698645         -1.573867         -0.174026
-            3           0.087941         -1.661039         -0.786015
-            4           0.874528         -0.611795         -0.350952
-            ...         ...               ...                ...
-            7141        15.560731        -17.223774        -0.699003
-            7142        16.699083        -17.310947        -0.261038
-            7143        18.272255        -17.574051         0.524977
-            7144        18.970900        -18.185846         1.922982
-            7145        21.071722        -16.786325         2.534972
+            0           0.000000         -0.614677          0.436291
+            1           0.000000         -0.700049          0.176093
+            2          -0.172905         -0.261807         -0.262826
+            3           0.262815         -0.261807          0.000000
+            4           0.608625         -0.614677         -0.349559
+            ...         ...               ...               ...
+            2903       -0.089911         -1.309034          0.000000
+            2904        0.525631         -0.438242         -0.436291
+            2905        0.871441         -0.961855          0.086733
+            2906        1.051262         -0.700049          0.176093
+            2907        1.134256         -0.347179         -0.525652
 
-[7146 rows x 3 columns]
+[2908 rows x 3 columns]
 
 #### Load and print sampling frequency of the data
 
@@ -248,11 +248,8 @@ print(f"turn id: {pham.turns_}")
 ```
 
     turn id:    onset   duration    event_type  tracking_systems    tracked_points
-    0           4.04    3.26        turn        imu                 pelvis
-    1           9.44    3.35        turn        imu                 pelvis
-    2           15.99   2.61        turn        imu                 pelvis
-    3           22.26   2.39        turn        imu                 pelvis
-    4           28.49   2.41        turn        imu                 pelvis   
+    0           7.17    2.14        turn        imu                 pelvis
+    1          10.65    1.77        turn        imu                 pelvis 
 
 ## Extraction of spatio-temporal parameters
 
@@ -272,8 +269,5 @@ print(pham.parameters_)
 
                 direction of turn  angle of turn [deg]  peak angular velocity [deg/s]
     turn id                                                                      
-    0           right              197.34               159.38
-    1           right              199.46               144.61
[…]
