### 2024-08-08 #6842 @aaa34169
here is my review of NGMT: NeuroGeriatric Motion Toolbox - An Open-Source Python Toolbox for Analyzing Neurological Motion Data from Various Recording Modalities

**General**

NGMT is implemented in Python (>=3.10) and is available under the Non-Profit Open Software License version 3.0. It can be installed from PyPI or from source on GitHub.

The documentation includes comprehensive installation instructions, a conceptual framework, and tutorial notebooks.

The code utilizes downloadable data from sources such as the *Mobilised* and *OpenNeuro* projects. However, it is important to note that these downloads may encounter SSL errors when accessed from a secure network, such as that of a hospital.

**Functionality**

The library is easily installable with Python 3.10. The local installation using the command

`pip install -e .`

works flawlessly.

The modules serve well-defined purposes.

The library employs data from projects such as Mobilised or OpenNeuro. However, when accessed on a secure network, an "SSLCertVerificationError" exception may be returned. It would be beneficial to mention this potential error and ideally provide information on how to resolve or circumvent it.
The library is intended for users such as rehabilitation center engineers or academic researchers who may want to use it on institutional workstations. It's important to prevent any potential reluctance from these users.

Regarding the data, I have also noted that the file "exAccelDataMobilise.csv" does not allow the example *modules_03_pam* to run successfully.

**Documentation**

The entirety of the library is documented and accessible .

To address common errors, I would recommend that the authors create a "Common Issues" or "Frequently Asked Questions" section.

In general, the documentation is intuitive and provides both high-level and low-level information. However, I find that the organization could be improved, particularly with a focus on the concept of DataClass and its usage.

First, I suggest that the horizontal menu place DataClass before modules. Additionally, on the Home page, it is pertinent to describe the "KielMATRecording" object thoroughly, accompanied specifically by an explanation demonstrating how to construct a KielMATRecording instance from the user’s own data, rather than from the Mobilised project dataset.

Moreover, on the Home page, the link https://neurogeriatricskiel.github.io/KielMAT/examples/00_tutorial_basics/ leads to a page with display errors.

In the Examples section, I find the title "Tutorial basics" to be somewhat misleading, as the content of this page explains how to load a dataset and does not explore the modules. I would propose the following organization for the examples:

  * Construct Your DataClass
    * Load data into KielMAT
    * Load datasets
    * Events in DataClass

  * Run Modules
    * Gait sequence
    * Initial contact detection

This reorganization would better reflect the content and provide clearer guidance for users.

**Paper**
The accompanying paper is very explicit, with numerous references that allow for further exploration in the field of actimetric evaluation. Here, even if i am not an expert of this domain, I use the term "actimetric" and wonder if this term should be integrated into the title to better delineate the library's scope. Indeed, the use of "Motion Data" is quite general and does not necessarily imply the use of IMUs in a mobility assessment context.

My major comment concerns the term "performance" as cited by the authors: "...a new dimension of measuring daily function, that is performance." Performance is implicitly related to the International Classification of Functioning, Disability and Health (ICF). I believe the authors should explicitly reference the ICF in the manuscript.

I also find no information in the "state of need" section regarding the four modules integrated into the library.

*Minor Comments*

  * Provide a reference to the Mobilise-D consortium.
  * Add The MathWorks, Inc. (2022) to the Matlab references.
  * Define FAIR.
