+++ kielmat/test/test_modules.py
@@ -659,28 +659,6 @@ def test_invalid_input_data_type():
         )
 
 
-def test_invalid_index_name():
-    # Initialize the class
-    pam = PhysicalActivityMonitoring()
-
-    # Test with invalid index name
-    data_with_wrong_index_name = acceleration_data.copy()
-    data_with_wrong_index_name.index.name = "wrong_name"
-    with pytest.raises(ValueError):
-        pam.detect(
-            data=data_with_wrong_index_name,
-            acceleration_unit="m/s^2",
-            sampling_freq_Hz=sampling_frequency,
-            thresholds_mg={
-                "sedentary_threshold": 45,
-                "light_threshold": 100,
-                "moderate_threshold": 400,
-            },
-            epoch_duration_sec=5,
-            plot=False,
-        )
-
-
 def test_insufficient_columns():
     # Initialize the class
     pam = PhysicalActivityMonitoring()
