+++ paper/references.bib
@@ -190,7 +190,7 @@
 	doi = {10.3389/fneur.2018.00652}
 }
 
-@article{van2013separating,
+@article{van:2013,
   title={Separating movement and gravity components in an acceleration signal and implications for the assessment of human daily physical activity},
   author={Van Hees, Vincent T and Gorzelniak, Lukas and Dean Le{\'o}n, Emmanuel Carlos and Eder, Martin and Pias, Marcelo and Taherian, Salman and Ekelund, Ulf and Renstr{\"o}m, Frida and Franks, Paul W and Horsch, Alexander and others},
   journal={PloS one},
@@ -211,6 +211,18 @@
   pages = {462--470},
   year = {2020},
   doi = {10.1016/S1474-4422(19)30397-7}
+}
+
+@article{wilkinson:2016,
+  author={Wilkinson, Mark D and Dumontier, Michel and Aalbersberg, IJsbrand Jan and Appleton, Gabrielle and Axton, Myles and Baak, Arie and Blomberg, Niklas and Boiten, Jan-Willem and da Silva Santos, Luiz Bonino and Bourne, Philip E and others},
+  title={The FAIR Guiding Principles for scientific data management and stewardship},
+  journal={Scientific data},
+  volume={3},
+  number={1},
+  pages={1--9},
+  year={2016},
+  publisher={Nature Publishing Group},
+  doi = {10.1038/sdata.2016.18}
 }
 
 % ----------
