+++ paper/references.bib
@@ -264,3 +264,14 @@
   doi= {10.5281/zenodo.7903963}
 }
 
+% ----------
+
+@software{MATLAB:2022,
+  year = {2022},
+  author = {The MathWorks Inc.},
+  title = {MATLAB version: 9.13.0 (R2022b)},
+  publisher = {The MathWorks Inc.},
+  address = {Natick, Massachusetts, United States},
+  url = {https://www.mathworks.com}
+}
+
