+++ paper/references.bib
@@ -245,6 +245,13 @@
   url = {https://github.com/stanfordnmbl/mobile-gaitlab}
 }
 
+@misc{mobilised:2019,
+  author = {},
+  title = {Mobilise-D consortium},
+  year = {2019},
+  url = {https://mobilise-d.eu/},
+}
+
 @misc{mobilised:2023,
   author = {Micó-Amigo, M. Encarna and Bonci, Tecla and Paraschiv-Ionescu, Anisoara and Ullrich, Martin and Kirk, Cameron and Soltani, Abolfazl and Küderle, Arne and Gazit, Eran and Salis, Francesca and Alcock, Lisa and Aminian, Kamiar and Becker, Clemens and Bertuletti, Stefano and Brown, Philip and Buckley, Ellen and Cantu, Alma and Carsin, Anne-Elie and Caruso, Marco and Caulfield, Brian and Cereatti, Andrea and Chiari, Lorenzo and D’Ascanio, Ilaria and Eskofier, Bjoern and Fernstad, Sara and Froehlich, Marcel and Garcia-Aymerich, Judith and Hansen, Clint and Hausdorff, Jeffrey M. and Hiden, Hugo and Hume, Emily and Keogh, Alison and Kluge, Felix and Koch, Sarah and Maetzler, Walter and Megaritis, Dimitrios and Mueller, Arne and Niessen, Martijn and Palmerini, Luca and Schwickert, Lars and Scott, Kirsty and Sharrack, Basil and Sillén, Henrik and Singleton, David and Vereijken, Beatrix and Vogiatzis, Ioannis and Yarnall, Alison J. and Rochester, Lynn and Mazzà, Claudia and Del Din, Silvia and {for the Mobilise-D consortium}},
   title = {Mobilise-D Technical Validation Study Recommended Algorithms},
