+++ paper/references.bib
@@ -190,6 +190,18 @@
 	doi = {10.3389/fneur.2018.00652}
 }
 
+@article{ustun:2003,
+  title={The International Classification of Functioning, Disability and Health: a new tool for understanding disability and health},
+  author={{\"U}stun, T Bedirhan and Chatterji, Somnath and Bickenbach, Jerome and Kostanjsek, Nenad and Schneider, Margie},
+  journal={Disability and rehabilitation},
+  volume={25},
+  number={11-12},
+  pages={565--571},
+  year={2003},
+  publisher={Taylor \& Francis},
+  doi = {10.1080/0963828031000137063}
+}
+
 @article{van:2013,
   title={Separating movement and gravity components in an acceleration signal and implications for the assessment of human daily physical activity},
   author={Van Hees, Vincent T and Gorzelniak, Lukas and Dean Le{\'o}n, Emmanuel Carlos and Eder, Martin and Pias, Marcelo and Taherian, Salman and Ekelund, Ulf and Renstr{\"o}m, Frida and Franks, Paul W and Horsch, Alexander and others},
