### 2024-08-14 2b51cc8
organization of the examples
