+++ docs/examples/index.md
@@ -2,13 +2,13 @@
 
 This is a collection of examples to explain various parts of the KielMAT toolbox.
 
-## Basic KielMAT usage
+## Construct Your DataClass
 
-### [Example 0: Introduction to the KielMAT toolbox](basic_00_intro_kielmat.md)
-### [Example 1: Load data into KielMAT](basic_01_load_Data.md)
+### [Example 0: Load data into KielMAT](basic_00_load_Data.md)
+### [Example 1: Introduction to the KielMAT toolbox](basic_01_intro_kielmat.md)
 ### [Example 2: Events in dataclass](basic_02_events.md)
 
-## Modules of KielMAT
+## Run Modules
 
 ### [Example 1: Gait Sequence Detection](modules_01_gsd.md)
 ### [Example 2: Initial Contact Detection](modules_02_icd.md)
