+++ docs/examples/index.md
@@ -4,8 +4,8 @@ This is a collection of examples to explain various parts of the KielMAT toolbox
 
 ## Construct Your DataClass
 
-### [Example 0: Load data into KielMAT](basic_00_load_Data.md)
-### [Example 1: Introduction to the KielMAT toolbox](basic_01_intro_kielmat.md)
+### [Example 0: Load data into KielMAT](basic_00_load_Data_into_KielMAT.md)
+### [Example 1: Load datasets](basic_01_load_dataset.md)
 ### [Example 2: Events in dataclass](basic_02_events.md)
 
 ## Run Modules
