+++ docs/examples/basic_00_load_Data_into_KielMAT.md
@@ -141,7 +141,7 @@ print(recording.events)
 
 
     
-![png](basic_01_load_Data_files/basic_01_load_Data_14_1.png)
+![png](basic_00_load_Data_into_KielMAT_files/basic_00_load_Data_into_KielMAT_1.png)
     
 
 
