+++ README.md
@@ -59,31 +59,168 @@ classDiagram
 
 ```
  A recording consists of the motion data from one or more tracking systems, where each tracking system may consist motion data from one or more tracked points. Therefore, the motion data (`KielMATRecording.data`) are organized as a dictionary where the dictionary keys refer to the tracking systems, and the corresponding values the actual (raw) data as a `pandas.DataFrame`. The description of data channels (`KielMATRecording.channels`) is availabe as a dictionary with the same keys, and the values contain the channels description.
+
+#### Example 1: Custom Dataset
+You can create a KielMATRecording instance from your own motion data.  Suppose you have motion data from your tracking system in a CSV file structured as follows:
+
 ```python
->>> from kielmat.datasets import mobilised
->>> file_name = "/mnt/neurogeriatrics_data/Mobilise-D/rawdata/sub-3011/Free-living/data.mat"
->>> recording = mobilised.load_recording(file_name, tracking_systems=["SU", "SU_INDIP"], tracked_points=["LowerBack"])
->>> recording.data
-{'SU':         LowerBack_ACCEL_x  ...  LowerBack_BARO_n/a
-0                0.967784  ...         1011.628100
-1                0.969667  ...         1011.628400
-...                   ...  ...                 ...
-993022           0.970579  ...         1012.078703
-993023           0.960542  ...         1002.580321
+      timestamp  YourTrackedPoint_ACCEL_x  YourTrackedPoint_ACCEL_y  YourTrackedPoint_ACCEL_z  YourTrackedPoint_GYRO_x  YourTrackedPoint_GYRO_y  YourTrackedPoint_GYRO_z
+0     0.00       0.1                       0.2                       9.8                       0.01                     0.02                     0.03
+1     0.01       0.1                       0.2                       9.8                       0.01                     0.02                     0.03
+...   ...        ...                       ...                       ...                       ...                      ...                      ...
+```
 
-[993024 rows x 10 columns], 
+You can create a `KielMATRecording` as follows:
 
-'SU_INDIP':      LowerBack_ACCEL_x  ... LowerBack_MAGN_z
-0                0.967986  ...         -5.902833
-1                0.963671  ...          9.501037
-...                   ...  ...               ...
-993022           0.951656  ...        -17.987983
-993023           0.955107  ...        -18.050600
+```python
+import pandas as pd
+from kielmat.utils.kielmat_dataclass import KielMATRecording
 
-[993024 rows x 9 columns]
-}
+# Load your motion data into a pandas DataFrame
+motion_data = pd.read_csv("path_to_your_data.csv")
 
->>> recording.channels
+# Calculate the sampling frequency using the timestamp column
+time_diff = motion_data["timestamp"].diff().dropna()  # Calculate time differences
+sampling_frequency = 1 / time_diff.mean()  # Sampling frequency in Hz
+
+# Drop the timestamp column, as it is not needed in the motion data
+motion_data = motion_data.drop(columns=["timestamp"])
+
+# Define the tracking system and tracked point names
+tracking_system = "YourTrackingSystem"
+tracked_point = "YourTrackedPoint"
+
+# Create the data dictionary
+data = {tracking_system: motion_data}
+
+# Create the channels DataFrame
+channels_info = pd.DataFrame({
+    "name": [f"{tracked_point}_ACCEL_x", f"{tracked_point}_ACCEL_y", 
+             f"{tracked_point}_ACCEL_z", f"{tracked_point}_GYRO_x", 
+             f"{tracked_point}_GYRO_y", f"{tracked_point}_GYRO_z"],
+    "type": ["ACCEL", "ACCEL", "ACCEL", "GYRO", "GYRO", "GYRO"],
+    "component": ["x", "y", "z", "x", "y", "z"],
[…]
