### 2024-08-19 01f03b6
Added entry for `faq.md` in the `mkdocs.yml`
