+++ mkdocs.yml
@@ -64,6 +64,7 @@ nav:
     - Data Preprocessing: utils/preprocessing.md
     - Data Import: utils/importers.md
     - Orientation Estimation: utils/orientation_estimation.md
+  - FAQ: faq.md
   - Contributing: 
     - contributing.md
 
