### 2024-08-19 3239048
Updated `faq.md` with `SSLCertVerificationError`
