### 2024-08-19 362df9b
Fixed some typos in the `faq.md`
