+++ docs/examples/index.md
@@ -4,9 +4,9 @@ This is a collection of examples to explain various parts of the KielMAT toolbox
 
 ## Construct Your DataClass
 
-### [Example 0: Load data into KielMAT](basic_00_load_Data_into_KielMAT.md)
-### [Example 1: Load datasets](basic_01_load_dataset.md)
-### [Example 2: Events in dataclass](basic_02_events.md)
+### [Example 1: Load data into KielMAT](basic_00_load_Data_into_KielMAT.md)
+### [Example 2: Load datasets](basic_01_load_dataset.md)
+### [Example 3: Events in dataclass](basic_02_events.md)
 
 ## Run Modules
 
