### 2024-08-20 84c0003
Added Physical Activity Monitoring example to index
