+++ paper/references.bib
@@ -1,3 +1,13 @@
+@book{ICF:2001,
+    address    = {Geneva},
+    title      = {International classification of functioning, disability and health: {ICF}},
+    isbn       = {978-92-4-154542-6},
+    shorttitle = {International classification of functioning, disability and health},
+    publisher  = {World Health Organization},
+    editor     = {{World Health Organization}},
+    year       = {2001},
+}
+
 @article{Boswell:2023,
   author = {Boswell, Melissa A and Kidzi{\'n}ski, {\L}ukasz and Hicks, Jennifer L and Uhlrich, Scott D and Falisse, Antoine and Delp, Scott L},
   title = {Smartphone videos of the sit-to-stand test predict osteoarthritis and health outcomes in a nationwide study},
