### 2024-08-20 ed3c4c9
Updated references for ICF in paper.md
