+++ paper/paper.md
@@ -41,7 +41,7 @@ output:
 The Kiel Motion Analysis Toolbox (KielMAT) is an open-source Python-based toolbox designed for processing human motion data, following open-science practices. KielMAT offers a range of algorithms for the processing of motion data in neuroscience and biomechanics and currently includes implementations for gait sequence detection, initial contact detection, physical activity monitoring, sit to stand and stand to sit detection algorithms. These algorithms aid in identifying patterns in human motion data on different time scales. The KielMAT is versatile in accepting motion data from various recording modalities, including IMUs that provide acceleration data from specific body locations such as the pelvis or wrist. This flexibility allows researchers to analyze data captured using different hardware setups, ensuring broad applicability across studies. Some of the toolbox algorithms have been developed and validated in clinical cohorts, allowing extracted patters to be used in a clinical context. The modular design of KielMAT allows the toolbox to be easily extended to incorporate relevant algorithms which will be developed in the research community. The toolbox is designed to be user-friendly and is accompanied by a comprehensive documentation and practical examples, while the underlying data structures build on the Motion BIDS specification [@jeung:2023]. The KielMAT toolbox is intended to be used by researchers and clinicians to analyze human motion data from various recording modalities and to promote the utilization of open-source software in the field of human motion analysis.
 
 # Statement of need
-Physical mobility is an essential aspect of health, as impairment in mobility is associated with reduced quality of life, falls, hospitalization, mortality, and other adverse events in many chronic conditions. Traditional mobility measures include patient-reported outcomes, objective clinical assessments, and subjective clinical assessments. These measures are linked to the perception and capacity aspects of health, which often fail to show relevant effects on daily function at an individual level [@maetzler:2021]. Perception involves surveys and patient-reported outcomes that capture how individuals feel about their own functional abilities, while capacity refers to clinical assessments of an individual's ability to perform various tasks. To complement both patient-reported (perception) and clinical (capacity) assessment approaches, digital health technology (DHT) introduces a new paradigm for assessing daily function. By using wearable devices, DHT provides objective insights into an individual's functional performance, directly linking it to the International Classification of Functioning, Disability and Health (ICF) framework for assessing how people perform in everyday life activities. [@warmerdam:2020; @fasano:2020; @maetzler:2021; @hansen:2018; @buckley:2019; @celik:2021; @ustun:2003]. DHT allows an objective impression of how patients function in everyday life and their ability to routinely perform everyday activities [@hansen:2018; @buckley:2019; @celik:2021]. Nonetheless, due to several persisting challenges in this field, current tools and techniques are still in their infancy [@micoamigo:2023]. Many studies often used proprietary software to clinically relevant features of mobility. The development of easy-to-use and open-source software is imperative for transparent features extraction in research and clinical settings. The Kiel Motion Analysis Toolbox (KielMAT) addresses this gap by providing software for human mobility analysis, to be used by motion researchers and clinicians, while promoting open-source practices. The conceptual framework builds on Findable, Accessible, Interoperable and Reusable (FAIR) data principles to encourage the use of open source software as well as facilitate data sharing and reproducibility in the field of human motion analysis [@wilkinson:2016]. The KielMAT comprises several modules which are implemented and validated with different dataset and each serving distinct purposes in human motion analysis:
[…]
