+++ paper/references.bib
@@ -269,7 +269,7 @@
 
 @misc{mobilised:2019,
   author = {},
-  title = {Mobilise-D consortium},
+  title = {{Mobilise-D} consortium},
   year = {2019},
   url = {https://mobilise-d.eu/},
 }
@@ -297,7 +297,7 @@
 
 @software{MATLAB:2022,
   year = {2022},
-  author = {The MathWorks Inc.},
+  author = {{The MathWorks Inc.}},
   title = {MATLAB version: 9.13.0 (R2022b)},
   publisher = {The MathWorks Inc.},
   address = {Natick, Massachusetts, United States},
