### 2024-08-22 #6842 @rmndrs89
## General

> NGMT is implemented in Python (>=3.10) and is available under the Non-Profit Open Software License version 3.0. It can be installed from PyPI or from source on GitHub.
>
> The documentation includes comprehensive installation instructions, a conceptual framework, and tutorial notebooks.
>
> The code utilizes downloadable data from sources such as the Mobilised and OpenNeuro projects. However, it is important to note that these downloads may encounter SSL errors when accessed from a secure network, such as that of a hospital. We hope that you find this a fair workaround.

Hi @aaa34169, thanks for pointing out the possible SSL error. We have researched this a bit, but were not able to reproduce this error ourselves since we are not on our university hospital's network. We have read that this error may be circumvented and have addressed this in the new [FAQ / Troubeshooting](https://neurogeriatricskiel.github.io/KielMAT/faq/) section in the Documentation. Worst case scenario is that users will have to manually navigate to the OpenNeuro or Zenodo repositories to fetch the datasets.

## Functionality

> The library is easily installable with Python 3.10. The local installation using the command
>
> `pip install -e .`
>
> works flawlessly.
>
> The modules serve well-defined purposes.
>
> The library employs data from projects such as Mobilised or OpenNeuro. However, when accessed on a secure network, an "SSLCertVerificationError" exception may be returned. It would be beneficial to mention this potential error and ideally provide information on how to resolve or circumvent it.
> The library is intended for users such as rehabilitation center engineers or academic researchers who may want to use it on institutional workstations. It's important to prevent any potential reluctance from these users.
>
> Regarding the data, I have also noted that the file "exAccelDataMobilise.csv" does not allow the example modules_03_pam to run successfully.

Thanks for this remark, the file is deprecated. The example now runs with [Mobilise-D dataset example](https://github.com/neurogeriatricskiel/KielMAT/blob/main/examples/modules_03_pam.ipynb).

## Documentation

> The entirety of the library is documented and accessible .
>
> To address common errors, I would recommend that the authors create a "Common Issues" or "Frequently Asked Questions" section.

Cool we have added a section [FAQ / Troubeshooting](https://neurogeriatricskiel.github.io/KielMAT/faq/) in the Documentation

> In general, the documentation is intuitive and provides both high-level and low-level information. However, I find that the organization could be improved, particularly with a focus on the concept of DataClass and its usage.
>
> First, I suggest that the horizontal menu place DataClass before modules. Additionally, on the Home page, it is pertinent to describe the "KielMATRecording" object thoroughly, accompanied specifically by an explanation demonstrating how to construct a KielMATRecording instance from the user’s own data, rather than from the Mobilised project dataset.

Yes, we agree, since the functionality of the toolbox is founded on the `KielMATRecording` object, its accurate description is key to user friendliness. We have updated the documentation accordingly by putting the [dataclass](https://neurogeriatricskiel.github.io/KielMAT/dataclass/) section before the [modules](https://neurogeriatricskiel.github.io/KielMAT/modules/). It is noteworthy to mention that the dataclass is also discussed on the [landing page of the documentation](https://neurogeriatricskiel.github.io/KielMAT/#data-classes).

> Moreover, on the Home page, the link https://neurogeriatricskiel.github.io/KielMAT/examples/00_tutorial_basics/ leads to a page with display errors.
>
> In the Examples section, I find the title "Tutorial basics" to be somewhat misleading, as the content of this page explains how to load a dataset and does not explore the modules. I would propose the following organization for the examples:
>
> Construct Your DataClass
>
> Load data into KielMAT
> Load datasets
> Events in DataClass
> Run Modules
>
> Gait sequence
> Initial contact detection
> This reorganization would better reflect the content and provide clearer guidance for users.

We appreciate this valuable feedback, and we agree that the order of the examples should make sense so that users can easily learn how to work with the toolbox. We have updated the [Examples](https://neurogeriatricskiel.github.io/KielMAT/examples/) in the documentation (thereby also the error with regards to `examples/00_tutorial_basics/` has been resolved).

## Paper

> The accompanying paper is very explicit, with numerous references that allow for further exploration in the field of actimetric evaluation. Here, even if i am not an expert of this domain, I use the term "actimetric" and wonder if this term should be integrated into the title to better delineate the library's scope. Indeed, the use of "Motion Data" is quite general and does not necessarily imply the use of IMUs in a mobility assessment context.

We agree that the term "motion data" is generic, however, we envision that its functionality will not be limited to accelerometer- (or inertial sensor-)based movement analysis. We are currently working on adding functionality for marker-based motion capture, and are exploring on how to include functionality for markerless motion capture as well. In that case the term "actimetric" would not cover the functionality anymore.

> My major comment concerns the term "performance" as cited by the authors: "...a new dimension of measuring daily function, that is performance." Performance is implicitly related to the International Classification of Functioning, Disability and Health (ICF). I believe the authors should explicitly reference the ICF in the manuscript.

We absolutely agree with this comment, and therefore explicitly refer to the ICF _"[...] DHT provides objective insights into an individual’s functional performance,
directly linking it to the International Classification of Functioning, Disability and Health (ICF)
framework (Üstun et al., 2003; World Health Organization, 2001) for assessing how people
perform in everyday life activities. [...]"_

> I also find no information in the "state of need" section regarding the four modules integrated into the library.

We have addressed this comment by explicitly mentioning the currently implemented modules in the Statement of Need section, like:

_"The KielMAT comprises several modules which are implemented and validated with dfferent dataset and each serving distinct purposes in human motion analysis:
1. Gait Sequence Detection (GSD): ...
2. Initial Contact Detection (ICD): ...
3. Physical Activity Monitoring (PAM): ..."_

Minor Comments

> Provide a reference to the Mobilise-D consortium.
> Add The MathWorks, Inc. (2022) to the Matlab references.
> Define FAIR.

We have added a reference to the Mobilise-D consortium, and to The Mathworks, Inc. (for MATLAB). We have defined and referenced FAIR as: _"[...] The conceptual framework builds on Findable, Accessible, Interoperable and Reusable (FAIR) data principles to encourage the use of open source software as well as facilitate data sharing and reproducibility in the field of human motion analysis (Wilkinson et al., 2016). [...]"_

Overall, we really appreciate the comments, and hope that our adjustments have increased the quality of the toolbox, documentation and paper. Feel free to let us know what else we can address.
