### 2024-09-03 #6842 @aaa34169
Hi

I am doing it for the end of the week.

From: jwelzel ***@***.***>
Sent: Tuesday, September 3, 2024 9:22 AM
Cc: Fabien Leboeuf ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: NGMT: NeuroGeriatric Motion Toolbox - An Open-Source Python Toolbox for Analyzing Neurological Motion Data from Various Recording Modalities (Issue #6842)

Hi @aaa34169<https://github.com/aaa34169> and @Kevin-Mattheus-Moerman<https://github.com/Kevin-Mattheus-Moerman>,

we would highly appreciate if you could provide us with a rough timeline when the second revision from your site will be completed.

Best,
Julius & Team
