### 2024-09-05 0ebe0d6
update DOI for JOSS
