+++ paper/references.bib
@@ -86,11 +86,19 @@
 	doi = {10.3233/JPD-181498}
 }
 
-@article{jeung:2023,
-  title={Motion-BIDS: extending the Brain Imaging Data Structure specification to organize motion data for reproducible research},
-  author={Jeung, Sein and Cockx, Helena and Appelhoff, Stefan and Berg, Timotheus and Gramann, Klaus and Grothkopp, S{\"o}ren and Warmerdam, Elke and Hansen, Clint and Oostenveld, Robert and Welzel, Julius and others},
-  year={2023},
-  publisher={PsyArXiv}
+@article{jeung:2024,
+  title = {Motion-{{BIDS}}: An Extension to the Brain Imaging Data Structure to Organize Motion Data for Reproducible Research},
+  shorttitle = {Motion-{{BIDS}}},
+  author = {Jeung, Sein and Cockx, Helena and Appelhoff, Stefan and Berg, Timotheus and Gramann, Klaus and Grothkopp, S{\"o}ren and Warmerdam, Elke and Hansen, Clint and Oostenveld, Robert and Welzel, Julius},
+  year = {2024},
+  month = jul,
+  journal = {Scientific Data},
+  volume = {11},
+  number = {1},
+  pages = {716},
+  publisher = {Nature Publishing Group},
+  issn = {2052-4463},
+  doi = {10.1038/s41597-024-03559-8},
 }
 
 @article{kidzinski:2020,
@@ -135,7 +143,8 @@
   number={7},
   pages={e69627},
   year={2013},
-  publisher={Public Library of Science San Francisco, USA}
+  publisher={Public Library of Science San Francisco, USA},
+  doi = {10.1371/journal.pone.0069627}
 }
 
 @article{mazza:2021,
@@ -146,7 +155,8 @@
   number={12},
   pages={e050785},
   year={2021},
-  publisher={British Medical Journal Publishing Group}
+  publisher={British Medical Journal Publishing Group},
+  doi = {10.1136/bmjopen-2021-050785}
 }
 
 @article{micoamigo:2023,
@@ -168,7 +178,8 @@
   number={1},
   pages={1--11},
   year={2019},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi = {10.1186/s12984-019-0494-z}
 }
 
 @inproceedings{paraschiv:2020,
@@ -177,7 +188,8 @@
   booktitle={2020 42nd Annual International Conference of the IEEE Engineering in Medicine \& Biology Society (EMBC)},
   pages={4596--4599},
   year={2020},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/EMBC44109.2020.9176281}
 }
 
 @article{pham:2017,
