### 2024-09-05 #6842 @JuliusWelzel
Hi @Kevin-Mattheus-Moerman,

we updated the .bib entries in https://github.com/neurogeriatricskiel/KielMAT/commit/0ebe0d61076c763a488db23d7db3fe515af865e7.

DOIs should return no error now
