### 2024-09-06 #6842 @Kevin-Mattheus-Moerman
@JuliusWelzel below are some minor comments on the paper:

- [ ] This reference in the paper does not seem to work: `...build on the Motion BIDS specification [@jeung:2023]`. Can you check if it is part of the .bib file and if the correct handle is used in the paper? Should this be `[@jeung:2024]` which is used later in the paper and works?
- [ ] The KielMAT name is explained twice. You can introduce it once, and subsequently keep using the name KielMAT.

Side note, you acknowledge no funding at the moment. This is not a requirement but if you do want to you can do so in the acknowledgements section.
