### 2024-09-12 #6842 @JuliusWelzel
The proof looks great! We're happy to give it the green light unless you need any revisions. Thanks a ton to @davidpagnon and @aaa34169 for their killer feedback. They really helped us level up the paper. Thanks again for @Kevin-Mattheus-Moerman excellent editorial handling.
