### 2024-09-16 #6842 @JuliusWelzel
>  Double check authors and affiliations (including ORCIDs)

All correct

> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

We made the [release 0.0.5](https://github.com/neurogeriatricskiel/KielMAT/releases/tag/release) for the review of JOSS (also see [PR102](https://github.com/neurogeriatricskiel/KielMAT/pull/102))

> * Archive the release on Zenodo/figshare/etc and post the DOI here.

Sourcecode has been archived on [OSF](https://osf.io/8m2sp/) with the following DOI:  10.17605/OSF.IO/8M2SP

> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

They match

> * Make sure that the license listed for the archive is the same as the software license.

It is the same LICENSE file

Let us know if there is anything we can help with :)
