### 2024-09-19 7ce0b8e
Orientation estimation using Versatile Quaternion-based Filter (VQF)
