+++ kielmat/modules/ptd/_pham.py
@@ -5,23 +5,26 @@ import scipy.signal
 from kielmat.utils import preprocessing
 from kielmat.utils import viz_utils
 from typing import Optional, Union, Tuple, List
+from vqf import VQF, BasicVQF, PyVQF
 
 
 class PhamPosturalTransitionDetection:
     """
     This algorithm aims to detect postural transitions (e.g., sit to stand or stand to sit movements)
     using accelerometer and gyroscope data collected from a lower back inertial measurement unit (IMU)
-    sensor.
+    sensor based on [1].
 
     The algorithm is designed to be robust in detecting postural transitions using inertial sensor data
     and provides detailed information about these transitions. It starts by loading the accelerometer and
     gyro data, which includes three columns corresponding to the acceleration and gyro signals across
     the x, y, and z axes, along with the sampling frequency of the data. It first checks the validity of
-    the input data. Then, it calculates the sampling period, selects accelerometer and gyro data. Tilt angle
-    estimation is performed using gyro data in lateral or anteroposterior direction which represent movements
-    or rotations in the mediolateral direction. The tilt angle is decomposed using wavelet transformation to
-    identify stationary periods. Stationary periods are detected using accelerometer variance and gyro variance.
-    Then, peaks in the wavelet-transformed tilt signal are detected as potential postural transition events.
+    the input data. Then, it calculates the sampling period, selects accelerometer and gyro data. Then, it uses 
+    a Versatile Quaternion-based Filter (VQF) to estimate the orientation of the IMU [2]. This helps in correcting 
+    the orientation of accelerometer and gyroscope data. Tilt angle estimation is performed using gyro data in 
+    lateral or anteroposterior direction which represent movements or rotations in the mediolateral direction. 
+    The tilt angle is decomposed using wavelet transformation to identify stationary periods. Stationary periods 
+    are detected using accelerometer variance and gyro variance. Then, peaks in the wavelet-transformed 
+    tilt signal are detected as potential postural transition events.
 
     If there's enough stationary data, further processing is done to estimate the orientation using
     quaternions and to identify the beginning and end of postural transitions using gyro data. Otherwise,
@@ -49,7 +52,6 @@ class PhamPosturalTransitionDetection:
         >>> pham = PhamPosturalTransitionDetection()
         >>> pham.detect(
                 data=input_data,
-                gyro_mediolateral="pelvis_GYRO_y",
                 accel_unit="g",
                 gyro_unit="deg/s",
                 sampling_freq_Hz=200.0,
@@ -70,6 +72,7 @@ class PhamPosturalTransitionDetection:
 
     References:
         [1] Pham et al. (2018). Validation of a Lower Back "Wearable"-Based Sit-to-Stand and Stand-to-Sit Algorithm... https://doi.org/10.3389/fneur.2018.00652
+        [2] D. Laidig and T. Seel. “VQF: Highly Accurate IMU Orientation Estimation with Bias Estimation ... https://doi.org/10.1016/j.inffus.2022.10.014
     """
 
     def __init__(
@@ -96,7 +99,6 @@ class PhamPosturalTransitionDetection:
     def detect(
         self,
         data: pd.DataFrame,
-        gyro_mediolateral: str,
         accel_unit: str,
         gyro_unit: str,
         sampling_freq_Hz: float,
@@ -110,7 +112,6 @@ class PhamPosturalTransitionDetection:
 
         Args:
             data (pd.DataFrame): Input accelerometer and gyro data (N, 6) for x, y, and z axes.
[…]
