### 2024-09-23 d00407d
examples work with publically avaliable datasets
