+++ docs/examples/modules_02_icd.md
@@ -2,7 +2,7 @@
 
 **Author:** Masoud Abedinifar
 
-**Last update:** Wed 27 Mar 2024
+**Last update:** Mon 23 Sep 2024
 
 ## Learning objectives
 By the end of this tutorial, you will be able to: 
@@ -41,38 +41,44 @@ from kielmat.modules.icd import ParaschivIonescuInitialContactDetection
 from kielmat.config import cfg_colors
 ```
 
-    c:\Users\User\Desktop\kiel\NGMT\.venv\lib\site-packages\tqdm\auto.py:21: TqdmWarning: IProgress not found. Please update jupyter and ipywidgets. See https://ipywidgets.readthedocs.io/en/stable/user_install.html
-      from .autonotebook import tqdm as notebook_tqdm
-    
-
 ## Data Preparation
 
 To implement the Paraschiv-Ionescu initial contact algorithm, we load example data from a KeepControl, which is publicly available on [OpenNeuro](https://openneuro.org/datasets/ds005258). 
 
 The participant was walking at a slow speed for a distance of 20m.
 
-#### Refertences
+```python
+# Dataset path
+dataset_path = Path(os.getcwd()) / "_keepcontrol"
 
-.. [`3`] Mazzà, Claudia, et al. "Technical validation of real-world monitoring of gait: a multicentric observational study." BMJ open 11.12 (2021): e050785. http://dx.doi.org/10.1136/bmjopen-2021-050785
-
+# Fetch the dataset
+keepcontrol.fetch_dataset(dataset_path)
+```    
 
 
 ```python
-recording = keepcontrol.load_recording(tracking_systems="imu", tracked_points="pelvis")
-```
+# In this example, we use "imu" as tracking_system and "pelvis" as tracked points.
+tracking_sys = "imu"
+tracked_points = {tracking_sys: ["pelvis"]}
 
-    Warning: The value of the key 'Subject' should be lower case. Converted to lower case.
-    Warning: The value of the key 'Task' should be lower case. Converted to lower case.
-    
+# The 'keepcontrol.load_recording' function is used to load the data from the specified file_path
+recording = keepcontrol.load_recording(
+    dataset_path=dataset_path,
+    id="pp002",
+    task="walkSlow",
+    tracking_systems=[tracking_sys], 
+    tracked_points=tracked_points
+)
 
+# Load lower back acceleration data
+accel_data = recording.data[tracking_sys][
+    ["pelvis_ACCEL_x", "pelvis_ACCEL_y", "pelvis_ACCEL_z"]
+]
 
-```python
-# find the channel with acceleration data
-chns_oi = [ch for ch in recording.channels["imu"]["name"] if "accel" in ch]
-acceleration_data = recording.data["imu"][chns_oi]
-
-# set the sampling frequency
-sampling_frequency = float(recording.channels["imu"]["sampling_frequency"].unique()[0])
+# Get the corresponding sampling frequency directly from the recording
+sampling_frequency = recording.channels[tracking_sys][
+    recording.channels[tracking_sys]["name"] == "pelvis_ACCEL_x"
+]["sampling_frequency"].values[0]
 ```
 
 ## Visualisation of the Data
@@ -80,9 +86,8 @@ The raw acceleration data including components of x, y and z axis is represented
 
 
 ```python
-# Calculate the time values in minutes
-# The 'time_in_minute' array represents time values in minutes, computed based on the length of 'acceleration_data' and 'sampling_frequency'.
-time_in_minute = np.arange(len(acceleration_data)) / (60 * sampling_frequency)
+# Calculate the time values
+time = np.arange(len(accel_data)) / (sampling_frequency)
 
 # Create a figure with a specified size
[…]
