### 2024-09-24 #6842 @Kevin-Mattheus-Moerman
@JuliusWelzel sorry for the delay. I am just checking with some colleagues about the OSF archive. We usually handle ZENODO archives mostly, so I am just checking we are all happy with the OSF link. I'd say it should be fine but it doesn't render the things we usually look for in the same way.
