### 2024-09-25 #6842 @Kevin-Mattheus-Moerman
@JuliusWelzel The OSF link can be acceptable. But can you ensure all authors are listed, trust the license matches your software license and that the version tag is mentioned? You may need to use the description field to mention the version tag.
