### 2024-09-27 #6842 @JuliusWelzel
Hi @Kevin-Mattheus-Moerman, thanks for your feedback on the OSF archive.

The authors are listed in the [projects wiki](https://osf.io/8m2sp/wiki/home/) under authors at the end of the document. The wiki also references the version number. The version number also can be found in the projects metadata under [description](https://osf.io/8m2sp/metadata/osf). The [licence file](https://osf.io/z3xcr) is the same as on GH. I hope that resolves all issues.
