+++ kielmat/modules/ptd/_pham.py
@@ -42,7 +42,7 @@ class PhamPosturalTransitionDetection:
     along with the detected postural transitions.
 
     Methods:
-        detect(data, gyro_mediolateral, accel_unit, gyro_unit, sampling_freq_Hz, dt_data, tracking_system, tracked_point, plot_results):
+        detect(accel_data, gyro_data, sampling_freq_Hz, dt_data, tracking_system, tracked_point, plot_results):
             Detects  sit to stand and stand to sit using accelerometer and gyro signals.
 
         spatio_temporal_parameters():
@@ -51,9 +51,8 @@ class PhamPosturalTransitionDetection:
     Examples:
         >>> pham = PhamPosturalTransitionDetection()
         >>> pham.detect(
-                data=input_data,
-                accel_unit="g",
-                gyro_unit="deg/s",
+                accel_data=accel_data,
+                gyro_data=gyro_data,
                 sampling_freq_Hz=200.0,
                 tracking_system="imu",
                 tracked_point="LowerBack",
@@ -98,9 +97,8 @@ class PhamPosturalTransitionDetection:
 
     def detect(
         self,
-        data: pd.DataFrame,
-        accel_unit: str,
-        gyro_unit: str,
+        accel_data: pd.DataFrame,
+        gyro_data: pd.DataFrame,
         sampling_freq_Hz: float,
         dt_data: Optional[pd.Series] = None,
         tracking_system: Optional[str] = None,
@@ -111,9 +109,8 @@ class PhamPosturalTransitionDetection:
         Detects postural transitions based on the input accelerometer and gyro data.
 
         Args:
-            data (pd.DataFrame): Input accelerometer and gyro data (N, 6) for x, y, and z axes.
-            accel_unit (str): Unit of acceleration data.
-            gyro_unit (str): Unit of gyro data.
+            accel_data (pd.DataFrame): Input accelerometer data (N, 3) for x, y, and z axes.
+            gyro_data (pd.DataFrame): Input gyro data (N, 3) for x, y, and z axes.
             sampling_freq_Hz (float): Sampling frequency of the input data.
             dt_data (pd.Series, optional): Original datetime in the input data. If original datetime is provided, the output onset will be based on that.
             tracking_system (str, optional): Tracking systems.
@@ -130,24 +127,35 @@ class PhamPosturalTransitionDetection:
                 - tracked_points: Name of the tracked points on the body.
         """
         # check if dt_data is a pandas Series with datetime values
-        if dt_data is not None and (
-            not isinstance(dt_data, pd.Series)
-            or not pd.api.types.is_datetime64_any_dtype(dt_data)
-        ):
-            raise ValueError("dt_data must be a pandas Series with datetime values")
-
-        # check if dt_data is provided and if it is a series with the same length as data
-        if dt_data is not None and len(dt_data) != len(data):
-            raise ValueError("dt_data must be a series with the same length as data")
-
+        if dt_data is not None:
+            # Ensure dt_data is a pandas Series
+            if not isinstance(dt_data, pd.Series):
+                raise ValueError("dt_data must be a pandas Series with datetime values")
+            
+            # Ensure dt_data has datetime values
+            if not pd.api.types.is_datetime64_any_dtype(dt_data):
+                raise ValueError("dt_data must be a pandas Series with datetime values")
+            
+            # Ensure dt_data has the same length as accel_data and gyro_data
+            if len(dt_data) != len(accel_data):
+                raise ValueError("dt_data must be a pandas Series with the same length as accel_data and gyro_data")
+        
         # Check if data is a DataFrame
-        if not isinstance(data, pd.DataFrame):
[…]
