+++ kielmat/test/test_modules.py
@@ -918,9 +918,8 @@ def test_pham_postural_transition_algorithm():
 
     # Perform detection with valid input data
     pham.detect(
-        input_data,
-        accel_unit="g",
-        gyro_unit="deg/s",
+        accel_data=input_data.iloc[:, 0:3],
+        gyro_data=input_data.iloc[:, 3:6],
         sampling_freq_Hz=sampling_freq,
         plot_results=False,
     )
@@ -928,7 +927,7 @@ def test_pham_postural_transition_algorithm():
     # Perform spatio-temporal parameter extraction
     pham.spatio_temporal_parameters()
 
-    # dt_data is a pandas Series but not with datetime dtype
+    # Test 1: dt_data is a pandas Series but not with datetime dtype
     dt_data = pd.Series(np.arange(1000))
 
     # Should raise ValueError because dt_data is not datetime
@@ -936,14 +935,13 @@ def test_pham_postural_transition_algorithm():
         ValueError, match="dt_data must be a pandas Series with datetime values"
     ):
         pham.detect(
-            data=input_data,
-            accel_unit="g",
-            gyro_unit="deg/s",
+            accel_data=input_data.iloc[:, 0:3],
+            gyro_data=input_data.iloc[:, 3:6],
             sampling_freq_Hz=200,
             dt_data=dt_data,
         )
 
-    # dt_data is a NumPy array instead of a pandas Series
+    # Test 2: dt_data is a NumPy array instead of a pandas Series
     dt_data = np.array(pd.date_range(start="1/1/2022", periods=1000, freq="S"))
 
     # Should raise ValueError because dt_data is not a pandas Series
@@ -951,29 +949,27 @@ def test_pham_postural_transition_algorithm():
         ValueError, match="dt_data must be a pandas Series with datetime values"
     ):
         pham.detect(
-            data=input_data,
-            accel_unit="g",
-            gyro_unit="deg/s",
+            accel_data=input_data.iloc[:, 0:3],
+            gyro_data=input_data.iloc[:, 3:6],
             sampling_freq_Hz=200,
             dt_data=dt_data,
         )
 
-    # dt_data is a pandas Series of datetime values but of incorrect length
+    # Test 3: dt_data is a pandas Series of datetime values but of incorrect length
     dt_data = pd.Series(pd.date_range("20210101", periods=500))
 
     # Should raise ValueError because dt_data length does not match input_data length
     with pytest.raises(
-        ValueError, match="dt_data must be a series with the same length as data"
+        ValueError, match="dt_data must be a pandas Series with the same length as accel_data and gyro_data"
     ):
         pham.detect(
-            data=input_data,
-            accel_unit="g",
-            gyro_unit="deg/s",
+            accel_data=input_data.iloc[:, 0:3],
+            gyro_data=input_data.iloc[:, 3:6],
             sampling_freq_Hz=200,
             dt_data=dt_data,
         )
 
-    # dt_data is a pandas Series but converted to a NumPy array
+    # Test 4: dt_data is a pandas Series but converted to a NumPy array
     dt_data = pd.Series(pd.date_range(start="1/1/2022", periods=1000, freq="S"))
     dt_data = dt_data.to_numpy()
 
@@ -982,9 +978,8 @@ def test_pham_postural_transition_algorithm():
         ValueError, match="dt_data must be a pandas Series with datetime values"
     ):
         pham.detect(
-            data=input_data,
-            accel_unit="g",
-            gyro_unit="deg/s",
+            accel_data=input_data.iloc[:, 0:3],
+            gyro_data=input_data.iloc[:, 3:6],
[…]

[…]
