+++ kielmat/modules/td/_pham.py
@@ -38,7 +38,7 @@ class PhamTurnDetection:
     the algorithm's performance and provides insights into the dynamics of the detected turns.
 
     Methods:
-        detect(data, gyro_vertical, accel_unit, gyro_unit, sampling_freq_Hz, dt_data, tracking_system, tracked_point, plot_results):
+        detect(accel_data, gyro_data, gyro_vertical, sampling_freq_Hz, dt_data, tracking_system, tracked_point, plot_results):
             Detects turns using accelerometer and gyro signals.
 
             Returns:
@@ -51,10 +51,9 @@ class PhamTurnDetection:
     Examples:
         >>> pham = PhamTurnDetection()
         >>> pham.detect(
-                data=input_data,
+                accel_data=accel_data,
+                gyro_data=gyro_data,
                 gyro_vertical="pelvis_GYRO_x",
-                accel_unit="g",
-                gyro_unit="rad/s",
                 sampling_freq_Hz=200.0,
                 tracking_system="imu",
                 tracked_point="LowerBack",
@@ -98,10 +97,9 @@ class PhamTurnDetection:
 
     def detect(
         self,
-        data: pd.DataFrame,
+        accel_data: pd.DataFrame,
+        gyro_data: pd.DataFrame,
         gyro_vertical: str,
-        accel_unit: str,
-        gyro_unit: str,
         sampling_freq_Hz: float,
         dt_data: Optional[pd.Series] = None,
         tracking_system: Optional[str] = None,
@@ -112,10 +110,9 @@ class PhamTurnDetection:
         Detects truns based on the input accelerometer and gyro data.
 
         Args:
-            data (pd.DataFrame): Input accelerometer and gyro data (N, 6) for x, y, and z axes.
+            accel_data (pd.DataFrame): Input accelerometer data (N, 3) for x, y, and z axes.
+            gyro_data (pd.DataFrame): Input gyro data (N, 3) for x, y, and z axes.
             gyro_vertical (str): The column name that corresponds to the vertical component gyro.
-            accel_unit (str): Unit of acceleration data.
-            gyro_unit (str): Unit of gyro data.
             sampling_freq_Hz (float): Sampling frequency of the input data in Hz.
             dt_data (pd.Series, optional): Original datetime in the input data. If original datetime is provided, the output onset will be based on that.
             tracking_system (str, optional): Tracking systems.
@@ -132,24 +129,35 @@ class PhamTurnDetection:
                 - tracked_points: Name of the tracked points on the body.
         """
         # check if dt_data is a pandas Series with datetime values
-        if dt_data is not None and (
-            not isinstance(dt_data, pd.Series)
-            or not pd.api.types.is_datetime64_any_dtype(dt_data)
-        ):
-            raise ValueError("dt_data must be a pandas Series with datetime values")
-
-        # check if dt_data is provided and if it is a series with the same length as data
-        if dt_data is not None and len(dt_data) != len(data):
-            raise ValueError("dt_data must be a series with the same length as data")
-
+        if dt_data is not None:
+            # Ensure dt_data is a pandas Series
+            if not isinstance(dt_data, pd.Series):
+                raise ValueError("dt_data must be a pandas Series with datetime values")
+            
+            # Ensure dt_data has datetime values
+            if not pd.api.types.is_datetime64_any_dtype(dt_data):
+                raise ValueError("dt_data must be a pandas Series with datetime values")
+            
+            # Ensure dt_data has the same length as input data
+            if len(dt_data) != len(accel_data):
+                raise ValueError("dt_data must be a series with the same length as data")
+    
[…]
