### 2024-10-02 df47956
Apply Black Formatting
