### 2024-10-02 #6842 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
