### 2024-10-02 #6842 @Kevin-Mattheus-Moerman
@JuliusWelzel as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process some final checks:

- [ ] Software license tag listed here matches a tagged release
