### 2024-10-04 b01627c
Formatted the admonition `Note` in the README
