+++ paper/paper.md
@@ -1,81 +1,0 @@
----
-title: "KielMAT: Kiel Motion Analysis Toolbox - An Open-Source Python Toolbox for Analyzing Neurological Motion Data from Various Recording Modalities"
-tags:
-  - Python
-  - Motion capture
-  - Neurology
-  - Accelerometer
-  - Gyroscope
-authors:
-  - name: Masoud Abedinifar^[co-first author]
-    orcid: 0000-0002-4050-9835
-    affiliation: 1
-  - name: Julius Welzel^[co-first author]
-    orcid: 0000-0001-8958-0934
-    affiliation: 1
-  - name: Clint Hansen
-    orcid: 0000-0003-4813-3868
-    affiliation: 1
-  - name: Walter Maetzler
-    orcid: 0000-0002-5945-4694
-    affiliation: 1
-  - name: Robbin Romijnders^[corresponding author]
-    orcid: 0000-0002-2507-0924
-    affiliation: 1
-affiliations:
-  - name: Neurogeriatrics, Department of Neurology, University Hospital Schleswig-Holstein (USKH), Kiel Germany
-    index: 1
-date: 08 Feb 2024
-bibliography: references.bib
-output:
-  quarto::pdf_document:
-    latex_engine: xelatex
-    keep_tex: true
----
-
-
-
-<div style="text-align: justify;">
-
-# Summary
-The Kiel Motion Analysis Toolbox (KielMAT) is an open-source Python-based toolbox designed for processing human motion data, following open-science practices. KielMAT offers a range of algorithms for the processing of motion data in neuroscience and biomechanics and currently includes implementations for gait sequence detection, initial contact detection, physical activity monitoring, sit to stand and stand to sit detection algorithms. These algorithms aid in identifying patterns in human motion data on different time scales. The KielMAT is versatile in accepting motion data from various recording modalities, including IMUs that provide acceleration data from specific body locations such as the pelvis or wrist. This flexibility allows researchers to analyze data captured using different hardware setups, ensuring broad applicability across studies. Some of the toolbox algorithms have been developed and validated in clinical cohorts, allowing extracted patters to be used in a clinical context. The modular design of KielMAT allows the toolbox to be easily extended to incorporate relevant algorithms which will be developed in the research community. The toolbox is designed to be user-friendly and is accompanied by a comprehensive documentation and practical examples, while the underlying data structures build on the Motion BIDS specification [@jeung:2024]. The KielMAT toolbox is intended to be used by researchers and clinicians to analyze human motion data from various recording modalities and to promote the utilization of open-source software in the field of human motion analysis.
-
-# Statement of need
[…]
