+++ paper/references.bib
@@ -1,318 +1,0 @@
-@book{ICF:2001,
-    address    = {Geneva},
-    title      = {International classification of functioning, disability and health: {ICF}},
-    isbn       = {978-92-4-154542-6},
-    shorttitle = {International classification of functioning, disability and health},
-    publisher  = {World Health Organization},
-    editor     = {{World Health Organization}},
-    year       = {2001},
-}
-
-@article{Boswell:2023,
-  author = {Boswell, Melissa A and Kidzi{\'n}ski, {\L}ukasz and Hicks, Jennifer L and Uhlrich, Scott D and Falisse, Antoine and Delp, Scott L},
-  title = {Smartphone videos of the sit-to-stand test predict osteoarthritis and health outcomes in a nationwide study},
-  journal = {npj Digital Medicine},
-  volume = {6},
-  number = {1},
-  pages = {32},
-  year = {2023},
-  publisher={Nature Publishing Group UK London},
-  doi= {10.1038/s41746-023-00775-1}
-}
-
-@article{buckley:2019,
-	author = {Buckley, Christopher and Alcock, Lisa and McArdle, Ríona and Rehman, Rana Zia Ur and Del Din, Silvia and Mazzà, Claudia and Yarnall, Alison J. and Rochester, Lynn},
-  title = {The {Role} of {Movement} {Analysis} in {Diagnosing} and {Monitoring} {Neurodegenerative} {Conditions}: {Insights} from {Gait} and {Postural} {Control}},
-	journal = {Brain Sciences},
-  volume = {9},
-	number = {2},
-  pages = {34},
-  year = {2019},
-  doi = {10.3390/brainsci9020034}
-}
-
-@article{celik:2021,
-  author = {Celik, Y. and Stuart, S. and Woo, W. L. and Godfrey, A.},
-	title = {Gait analysis in neurological populations: {Progression} in the use of wearables},
-	journal = {Medical Engineering \& Physics},
-  volume = {87},
-  pages = {9--29},
-  year = {2021},
-	doi = {10.1016/j.medengphy.2020.11.005}
-}
-
-@article{czech:2019,
-  author = {Matthew D. Czech and Shyamal Patel}, 
-  title = {GaitPy: An Open-Source Python Package for Gait Analysis Using an Accelerometer on the Lower Back}, 
-  journal = {Journal of Open Source Software},
-  volume = {4}, 
-  number = {43}, 
-  pages = {1778}, 
-  year = {2019}, 
-  publisher = {The Open Journal}, 
-  doi = {10.21105/joss.01778} 
-}
-
-@article{fasano:2020,
-	author = {Fasano, Alfonso and Mancini, Martina},
-  title = {Wearable-based mobility monitoring: the long road ahead},
-  journal = {The Lancet Neurology},
-	volume = {19},
-	number = {5},
-	pages = {378--379},
-  year = {2020},
-  doi = {10.1016/S1474-4422(20)30033-8}
-}
-
-@article{gorgolewski:2016,
[…]
