### 2024-10-08 3739575
[ADD] paper joss doi to README
