### 2024-10-10 0200889
[ADD] example how to use gaitmap with KielMAT
