### 2024-10-17 e6d8d09
temporal gait parameters
