+++ kielmat/modules/icd/_paraschiv.py
@@ -18,17 +18,25 @@ class ParaschivIonescuInitialContactDetection:
     The algorithm takes accelerometer data as input, and the vertical acceleration component, and processes each
     specified gait sequence independently. The signal is first detrended and then low-pass filtered. The resulting
     signal is numerically integrated and differentiated using a Gaussian continuous wavelet transformation. The
-    initial contact (IC) events are identified as the positive maximal peaks between successive zero-crossings.
+    initial contact (IC) events are identified as the positive maximal peaks between successive zero-crossings. In 
+    addition, final contact (FC) are identified are obtained from a further CWT differentiation of the signal [3].
 
     Finally, initial contacts information is provided as a DataFrame with columns `onset`, `event_type`, and
-    `tracking_systems`.
+    `tracking_systems`. In addition, the Final contact inofrmation also provided as a DataFrame with columns `onset`, `event_type`, and
+    `tracking_systems`. 
+
+    Additionally, temporal parameters related to gait can be calculated and accessed, including step time, stride time,
+    stance time, single support time, double support time, and cadence [1,4,5].
 
     Methods:
         detect(accel_data, gait_sequences, sampling_freq_Hz):
             Detects initial contacts on the accelerometer signal.
 
+        temporal_parameters():
+            Calculates the temporal parameters of the detected gaits using initial and final contacts information.
+            
     Examples:
-        Find initial contacts based on the detected gait sequence
+        Detect initial contacts, final contacts, and calculate temporal parameters:
 
         >>> icd = ParaschivIonescuInitialContactDetection()
         >>> icd = icd.detect(accel_data=acceleration_data, sampling_freq_Hz=100)
@@ -37,10 +45,26 @@ class ParaschivIonescuInitialContactDetection:
             0   5       initial contact  0          SU
             1   5.6     initial contact  0          SU
 
+        >>> print(icd.final_contacts_)
+               onset       event_type       duration   tracking_systems
+            0  5.300  final contact           0         SU
+            1  6.100  final contact           0         SU
+        
+        >>> icd.temporal_parameters()
+        >>> print(icd.temporal_parameters_)
+            gait_sequence_id      step_time      stride_time    stance_time     swing_time     cadence
+            0                 0   [0.575, 0.75]  [1.325, 1.4]   [0.275, 0.35]   [1.04, 1.04]   97.39
+
     References:
         [1] Paraschiv-Ionescu et al. (2019). Locomotion and cadence detection using a single trunk-fixed accelerometer...
 
         [2] Paraschiv-Ionescu et al. (2020). Real-world speed estimation using single trunk IMU: methodological challenges...
+
+        [3] McCamley et al. (2012) An enhanced estimate of initial contact and final contact instants of time...
+        
+        [4] Din et al. (2015) Validation of an accelerometer to quantify a comprehensive battery of gait characteristics...
+
+        [5] Godfrey et al. (2015) Instrumenting gait with an accelerometer: A system and algorithm examination...
     """
 
     def __init__(
@@ -50,6 +74,9 @@ class ParaschivIonescuInitialContactDetection:
         Initializes the ParaschivIonescuInitialContactDetection instance.
         """
         self.initial_contacts_ = None
+        self.final_contacts_ = None
+        self.temporal_parameters_ = None
+
 
     def detect(
         self,
@@ -105,17 +132,20 @@ class ParaschivIonescuInitialContactDetection:
         # Extract vertical accelerometer data using the specified index
         acc_vertical = accel_data[v_acc_col_name]
 
-        # Initialize an empty list to store the processed output
-        processed_output = []
+        # Initialize lists to store detected initial and final contacts
+        initial_contacts_list = []
+        final_contacts_list = []
 
-        # Initialize an empty list to store all onsets
-        all_onsets = []
[…]
